% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{addTreatmentAndOutcome}
\alias{addTreatmentAndOutcome}
\title{Add treatment and outcome data}
\usage{
addTreatmentAndOutcome(data, cohortData)
}
\arguments{
\item{data}{Covariates for data dimension}

\item{cohortData}{cohortMethodData object}
}
\value{
Returns input covariates, with extra columns treatment and outcome that are 1/0
}
\description{
Appends columns to covariates that includes treatment and outcome for that patient
}

