% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adjustedKm}
\alias{adjustedKm}
\title{Compute a weight-adjusted Kaplan-Meier curve}
\usage{
adjustedKm(weight, time, y)
}
\arguments{
\item{weight}{Vector of observation weights}

\item{time}{Vector of event times}

\item{y}{Vector outcomes (0 indicates censoring, 1 indicates event-of-interest)}
}
\description{
Compute a weight-adjusted Kaplan-Meier curve
}
