% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{appendAnalysisId}
\alias{appendAnalysisId}
\title{Appends analysis Id onto new codes}
\usage{
appendAnalysisId(codes, analysisId)
}
\arguments{
\item{codes}{\code{ffdf} with columns: \code{SOURCE_CODE} and \code{TARGET_CONCEPT_ID}}

\item{analysisId}{identifier for data dimension}
}
\value{
Returns \code{codes} with analysisId added
}
\description{
This function puts the analysisId in front of the new codes. This is to prevent confusion in situations
where ICD9 codes are used for both inpatient and outpatient purposes
}

