% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{calculateRanks}
\alias{calculateRanks}
\title{Calculates bias and exp ranks for each covariate}
\usage{
calculateRanks(data, cohortData, fudge)
}
\arguments{
\item{data}{Covariate information for data dimension, as ffdf object with columns \code{rowId}, \code{covariateId},
\code{treatment}, \code{outcome}}

\item{cohortData}{\code{cohortData} object constructed by \code{getDbCohortData}}

\item{fudge}{Constant to avoid divide by zero}
}
\value{
Returns ffdf object with the following columns: \describe{
\item{covariateId}{Covariate identifier}
\item{biasRank}{The value of abs(log(bias)) for each covariate, ranked decreasingly}
\item{expRank}{The value of abs(log(RRce)) for each covariate, ranked decreasingly}}
Returns list to work with sapply.
}
\description{
This function calculates the bias and exp ranks, as described in the hdps implementation, and the relative risk of each \code{covariateId} for a
data dimension.
}

