% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{checkCmInstallation}
\alias{checkCmInstallation}
\title{Check is CohortMethod and its dependencies are correctly installed}
\usage{
checkCmInstallation(connectionDetails)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the function
\code{createConnectionDetails} in the \code{DatabaseConnector} package.}
}
\description{
Check is CohortMethod and its dependencies are correctly installed
}
\details{
This function checks whether CohortMethod and its dependencies are correctly installed. This will
check the database connectivity, large scale regresion engine (Cyclops), and large data object
handling (ff).
}
