% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CohortMethod.R
\docType{data}
\name{cohortDataSimulationProfile}
\alias{cohortDataSimulationProfile}
\title{A simulation profile}
\usage{
data(cohortDataSimulationProfile)
}
\description{
A simulation profile
}
\keyword{datasets}

