% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDbInterface_R6.R
\name{CohortDbInterface}
\alias{CohortDbInterface}
\title{CohortDbInterface}
\description{
Class to interface with database.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CohortDbInterface-new}{\code{CohortDbInterface$new()}}
\item \href{#method-CohortDbInterface-disconnect}{\code{CohortDbInterface$disconnect()}}
\item \href{#method-CohortDbInterface-createCohorts}{\code{CohortDbInterface$createCohorts()}}
\item \href{#method-CohortDbInterface-getCohorts}{\code{CohortDbInterface$getCohorts()}}
\item \href{#method-CohortDbInterface-countCohorts}{\code{CohortDbInterface$countCohorts()}}
\item \href{#method-CohortDbInterface-sampleCohorts}{\code{CohortDbInterface$sampleCohorts()}}
\item \href{#method-CohortDbInterface-countRaw}{\code{CohortDbInterface$countRaw()}}
\item \href{#method-CohortDbInterface-getOutcomes}{\code{CohortDbInterface$getOutcomes()}}
\item \href{#method-CohortDbInterface-rmTempTables}{\code{CohortDbInterface$rmTempTables()}}
\item \href{#method-CohortDbInterface-extractCovarDat}{\code{CohortDbInterface$extractCovarDat()}}
\item \href{#method-CohortDbInterface-clone}{\code{CohortDbInterface$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortDbInterface-new"></a>}}
\if{latex}{\out{\hypertarget{method-CohortDbInterface-new}{}}}
\subsection{Method \code{new()}}{
Initializer method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortDbInterface$new(
  connectionDetails,
  tempEmulationSchema,
  cdmDatabaseSchema,
  exposureDatabaseSchema,
  exposureTable,
  cdmVersion,
  outcomeTable,
  outcomeDatabaseSchema
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connectionDetails}}{\if{html}{\out{<ConnectionDetails>}} created by
\link[DatabaseConnector]{createConnectionDetails} >= 6.0.0}

\item{\code{tempEmulationSchema}}{\if{html}{\out{
<character>
}}}

\item{\code{cdmDatabaseSchema}}{\if{html}{\out{
<character>
}}}

\item{\code{exposureDatabaseSchema}}{\if{html}{\out{
<character>
}}}

\item{\code{exposureTable}}{\if{html}{\out{
<character>
}}}

\item{\code{cdmVersion}}{\if{html}{\out{
<character>
}}}

\item{\code{outcomeTable}}{\if{html}{\out{
<character>
}}}

\item{\code{outcomeDatabaseSchema}}{\if{html}{\out{
<character>
}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortDbInterface-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-CohortDbInterface-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
Method to disconnect from database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortDbInterface$disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortDbInterface-createCohorts"></a>}}
\if{latex}{\out{\hypertarget{method-CohortDbInterface-createCohorts}{}}}
\subsection{Method \code{createCohorts()}}{
Method to create cohorts in the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortDbInterface$createCohorts(
  targetId,
  comparatorId,
  studyStartDate,
  studyEndDate,
  firstExposureOnly,
  removeDuplicateSubjects,
  washoutPeriod,
  restrictToCommonPeriod
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{targetId}}{\if{html}{\out{
<numeric>
}}}

\item{\code{comparatorId}}{\if{html}{\out{
<numeric>
}}}

\item{\code{studyStartDate}}{\if{html}{\out{
<character>
}}}

\item{\code{studyEndDate}}{\if{html}{\out{
<character>
}}}

\item{\code{firstExposureOnly}}{\if{html}{\out{
<logical>
}}}

\item{\code{removeDuplicateSubjects}}{\if{html}{\out{<character>}} One of: "keep all", "keep first", "remove all"}

\item{\code{washoutPeriod}}{\if{html}{\out{
<numeric>
}}}

\item{\code{restrictToCommonPeriod}}{\if{html}{\out{
<logical>
}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortDbInterface-getCohorts"></a>}}
\if{latex}{\out{\hypertarget{method-CohortDbInterface-getCohorts}{}}}
\subsection{Method \code{getCohorts()}}{
Retrieve cohorts from the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortDbInterface$getCohorts(sampled, targetId)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sampled}}{\if{html}{\out{
<logical>
}}}

\item{\code{targetId}}{\if{html}{\out{
<numeric>
}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<data.frame>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortDbInterface-countCohorts"></a>}}
\if{latex}{\out{\hypertarget{method-CohortDbInterface-countCohorts}{}}}
\subsection{Method \code{countCohorts()}}{
Counts cohorts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortDbInterface$countCohorts(targetId)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{targetId}}{\if{html}{\out{
<numeric>
}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<data.frame>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortDbInterface-sampleCohorts"></a>}}
\if{latex}{\out{\hypertarget{method-CohortDbInterface-sampleCohorts}{}}}
\subsection{Method \code{sampleCohorts()}}{
Method to sample cohorts in the database to the specified maxCohortSize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortDbInterface$sampleCohorts(maxCohortSize)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{maxCohortSize}}{\if{html}{\out{
<numeric>
}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortDbInterface-countRaw"></a>}}
\if{latex}{\out{\hypertarget{method-CohortDbInterface-countRaw}{}}}
\subsection{Method \code{countRaw()}}{
Count overall exposed population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortDbInterface$countRaw(
  targetId,
  comparatorId,
  studyStartDate,
  studyEndDate
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{targetId}}{\if{html}{\out{
<numeric>
}}}

\item{\code{comparatorId}}{\if{html}{\out{
<numeric>
}}}

\item{\code{studyStartDate}}{\if{html}{\out{
<character>
}}}

\item{\code{studyEndDate}}{\if{html}{\out{
<character>
}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
<data.frame>
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortDbInterface-getOutcomes"></a>}}
\if{latex}{\out{\hypertarget{method-CohortDbInterface-getOutcomes}{}}}
\subsection{Method \code{getOutcomes()}}{
Method to retrieve the outcomes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortDbInterface$getOutcomes(metaData, outcomeIds, sampled)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metaData}}{\if{html}{\out{
<list>
}}}

\item{\code{outcomeIds}}{\if{html}{\out{
<numeric>
}}}

\item{\code{sampled}}{\if{html}{\out{
<logical>
}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
metaData
\if{html}{\out{<list>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortDbInterface-rmTempTables"></a>}}
\if{latex}{\out{\hypertarget{method-CohortDbInterface-rmTempTables}{}}}
\subsection{Method \code{rmTempTables()}}{
Remove temp tables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortDbInterface$rmTempTables(sampled)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sampled}}{\if{html}{\out{
<logical>
}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortDbInterface-extractCovarDat"></a>}}
\if{latex}{\out{\hypertarget{method-CohortDbInterface-extractCovarDat}{}}}
\subsection{Method \code{extractCovarDat()}}{
Wrapper to extract covariate data. See
\link[FeatureExtraction]{getDbCovariateData}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortDbInterface$extractCovarDat(cohortTable, covariateSettings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cohortTable}}{\if{html}{\out{
<character>
}}}

\item{\code{covariateSettings}}{\if{html}{\out{<CovariateSettings>}} class. See
\link[FeatureExtraction]{createCovariateSettings}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortDbInterface-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CohortDbInterface-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortDbInterface$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
