% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMethodData.R
\docType{class}
\name{CohortMethodData-class}
\alias{CohortMethodData-class}
\alias{CohortMethodData}
\alias{show,CohortMethodData-method}
\alias{summary,CohortMethodData-method}
\title{Cohort Method Data}
\usage{
\S4method{show}{CohortMethodData}(object)

\S4method{summary}{CohortMethodData}(object)
}
\arguments{
\item{object}{An object of type \code{CohortMethodData}.}
}
\description{
\code{CohortMethodData} is an S4 class that inherits from \link[FeatureExtraction:CovariateData-class]{CoviarateData}, which in turn inherits from \link[Andromeda:Andromeda-class]{Andromeda}. It contains information on the cohorts, their
outcomes, and baseline covariates. Information about multiple outcomes can be captured at once for
efficiency reasons.

A \code{CohortMethodData} is typically created using \code{\link[=getDbCohortMethodData]{getDbCohortMethodData()}}, can only be saved using
\code{\link[=saveCohortMethodData]{saveCohortMethodData()}}, and loaded using \code{\link[=loadCohortMethodData]{loadCohortMethodData()}}.
}
