% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMethodData_R6.R
\name{CohortMethodData}
\alias{CohortMethodData}
\title{CohortMethodData}
\description{
Class to build COhortMethodData.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CohortMethodData-new}{\code{CohortMethodData$new()}}
\item \href{#method-CohortMethodData-createCohorts}{\code{CohortMethodData$createCohorts()}}
\item \href{#method-CohortMethodData-downSample}{\code{CohortMethodData$downSample()}}
\item \href{#method-CohortMethodData-getCohorts}{\code{CohortMethodData$getCohorts()}}
\item \href{#method-CohortMethodData-buildCovariateData}{\code{CohortMethodData$buildCovariateData()}}
\item \href{#method-CohortMethodData-clone}{\code{CohortMethodData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortMethodData-new"></a>}}
\if{latex}{\out{\hypertarget{method-CohortMethodData-new}{}}}
\subsection{Method \code{new()}}{
Initializer method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMethodData$new(
  metaData,
  outcomeIds,
  firstExposureOnly,
  removeDuplicateSubjects,
  restrictToCommonPeriod,
  washoutPeriod,
  maxCohortSize,
  covariateSettings,
  cohortDbInterface
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metaData}}{\if{html}{\out{<list>}} containing: targetId, comparatorId, studyStartDate, studyEndDate}

\item{\code{outcomeIds}}{\if{html}{\out{
<numeric>
}}}

\item{\code{firstExposureOnly}}{\if{html}{\out{
<logical>
}}}

\item{\code{removeDuplicateSubjects}}{\if{html}{\out{<character>}} One of: "keep all", "keep first", "remove all"}

\item{\code{restrictToCommonPeriod}}{\if{html}{\out{
<logical>
}}}

\item{\code{washoutPeriod}}{\if{html}{\out{
<numeric>
}}}

\item{\code{maxCohortSize}}{\if{html}{\out{
<numeric>
}}}

\item{\code{covariateSettings}}{\if{html}{\out{<CovariateSettings>}} class}

\item{\code{cohortDbInterface}}{\if{html}{\out{<CohortDbInterface>}} R6 Class}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortMethodData-createCohorts"></a>}}
\if{latex}{\out{\hypertarget{method-CohortMethodData-createCohorts}{}}}
\subsection{Method \code{createCohorts()}}{
Wrapper for \code{CohortDbInterface$createCohorts}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMethodData$createCohorts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortMethodData-downSample"></a>}}
\if{latex}{\out{\hypertarget{method-CohortMethodData-downSample}{}}}
\subsection{Method \code{downSample()}}{
Downsample cohorts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMethodData$downSample()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortMethodData-getCohorts"></a>}}
\if{latex}{\out{\hypertarget{method-CohortMethodData-getCohorts}{}}}
\subsection{Method \code{getCohorts()}}{
Wrapper for \code{cohortDbInterface$getCohorts}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMethodData$getCohorts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortMethodData-buildCovariateData"></a>}}
\if{latex}{\out{\hypertarget{method-CohortMethodData-buildCovariateData}{}}}
\subsection{Method \code{buildCovariateData()}}{
Build covariate data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMethodData$buildCovariateData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortMethodData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CohortMethodData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMethodData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
