% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMethodData_R6.R
\name{CohortMethodDataR6}
\alias{CohortMethodDataR6}
\title{CohortMethodData}
\description{
Class to build COhortMethodData.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CohortMethodDataR6-new}{\code{CohortMethodDataR6$new()}}
\item \href{#method-CohortMethodDataR6-createCohorts}{\code{CohortMethodDataR6$createCohorts()}}
\item \href{#method-CohortMethodDataR6-downSample}{\code{CohortMethodDataR6$downSample()}}
\item \href{#method-CohortMethodDataR6-getCohorts}{\code{CohortMethodDataR6$getCohorts()}}
\item \href{#method-CohortMethodDataR6-buildCovariateData}{\code{CohortMethodDataR6$buildCovariateData()}}
\item \href{#method-CohortMethodDataR6-clone}{\code{CohortMethodDataR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortMethodDataR6-new"></a>}}
\if{latex}{\out{\hypertarget{method-CohortMethodDataR6-new}{}}}
\subsection{Method \code{new()}}{
Initializer method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMethodDataR6$new(
  metaData,
  outcomeIds,
  firstExposureOnly,
  removeDuplicateSubjects,
  restrictToCommonPeriod,
  washoutPeriod,
  maxCohortSize,
  covariateSettings,
  cohortDbInterface
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metaData}}{\if{html}{\out{<list>}} containing: targetId, comparatorId, studyStartDate, studyEndDate}

\item{\code{outcomeIds}}{\if{html}{\out{
<numeric>
}}}

\item{\code{firstExposureOnly}}{\if{html}{\out{
<logical>
}}}

\item{\code{removeDuplicateSubjects}}{\if{html}{\out{<character>}} One of: "keep all", "keep first", "remove all"}

\item{\code{restrictToCommonPeriod}}{\if{html}{\out{
<logical>
}}}

\item{\code{washoutPeriod}}{\if{html}{\out{
<numeric>
}}}

\item{\code{maxCohortSize}}{\if{html}{\out{
<numeric>
}}}

\item{\code{covariateSettings}}{\if{html}{\out{<CovariateSettings>}} class}

\item{\code{cohortDbInterface}}{\if{html}{\out{<CohortDbInterface>}} R6 Class}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortMethodDataR6-createCohorts"></a>}}
\if{latex}{\out{\hypertarget{method-CohortMethodDataR6-createCohorts}{}}}
\subsection{Method \code{createCohorts()}}{
Wrapper for \code{CohortDbInterface$createCohorts}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMethodDataR6$createCohorts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortMethodDataR6-downSample"></a>}}
\if{latex}{\out{\hypertarget{method-CohortMethodDataR6-downSample}{}}}
\subsection{Method \code{downSample()}}{
Downsample cohorts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMethodDataR6$downSample()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortMethodDataR6-getCohorts"></a>}}
\if{latex}{\out{\hypertarget{method-CohortMethodDataR6-getCohorts}{}}}
\subsection{Method \code{getCohorts()}}{
Wrapper for \code{cohortDbInterface$getCohorts}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMethodDataR6$getCohorts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortMethodDataR6-buildCovariateData"></a>}}
\if{latex}{\out{\hypertarget{method-CohortMethodDataR6-buildCovariateData}{}}}
\subsection{Method \code{buildCovariateData()}}{
Build covariate data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMethodDataR6$buildCovariateData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortMethodDataR6-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CohortMethodDataR6-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMethodDataR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
