% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMethod.R
\docType{data}
\name{cohortMethodDataSimulationProfile}
\alias{cohortMethodDataSimulationProfile}
\title{A simulation profile}
\usage{
data(cohortMethodDataSimulationProfile)
}
\description{
A simulation profile
}
\keyword{datasets}
