% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompetingRisks.R
\name{combineCompetingStudyPopulations}
\alias{combineCompetingStudyPopulations}
\title{Create competing risk study population}
\usage{
combineCompetingStudyPopulations(mainPopulation, competingRiskPopulation)
}
\arguments{
\item{mainPopulation}{A data frame describing the population. This should at least have a
\code{rowId} column corresponding to the \code{rowId} column in the
\link{CohortMethodData} covariates object and a \code{treatment} column.}

\item{competingRiskPopulation}{A data frame ....}
}
\description{
Creates competing risk study population from two overlaping study populations
}
\examples{
# TODO

}
