% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{combineData}
\alias{combineData}
\title{Combines covariate, treatment, and outcome}
\usage{
combineData(cohortData, codes, dimCovariateId)
}
\arguments{
\item{cohortData}{\code{cohortData} object constructed by \code{getDbCohortData}}

\item{codes}{List of \code{ffdf} source code mappings for each data dimension with columns:
\code{SOURCE_CODE} and \code{TARGET_CONCEPT_ID}}

\item{dimCovariateId}{List of \code{ff_vector} covariate identifiers for each data dimension}
}
\value{
Returns \code{list} of \code{ffdf} by data dimension with the following columns: \describe{
\item{rowId}{Patient identifier}
\item{covariateId}{New source code for data dimension}
\item{treatment}{1 if patient is in treatment cohort; 0 otherwise}
\item{outcome}{1 if patient is in outcome cohort; 0 otherwise} }
}
\description{
This function combines covariate, treatment, and outcome information, then lists them by data dimension.
It breaks down \code{cohortData$covariates} by data dimension, converts the \code{covariateId} to the \code{SOURCE_CODE}
that matches the \code{conceptId} associated with that \code{covariateId}, then adds \code{treatment} and \code{outcome}
as binary indicators for that patient (\code{rowId}).
}

