% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{combineDataHelper}
\alias{combineDataHelper}
\title{Helper for combineData}
\usage{
combineDataHelper(sourceId, rowIds, treatments, outcomes)
}
\arguments{
\item{sourceId}{An ff vector of with length = length(\code{cohortData$covariates}) of converted source codes for a data
dimension.}

\item{rowIds}{An ff vector of patient identifiers}

\item{treatments}{An ff vector of binary treatment identifier for each entry of \code{cohortData$covariates}}

\item{outcomes}{An ff vector of binary outcome identifier for each entry of \code{cohortData$covariates}}
}
\value{
Returns ffdf object with the four combined ff vectors, only on covariate indices appropriate for that data dimension.
\code{covariateId} is converted to a factor, if it is not already one.
}
\description{
Helper for combineData
}

