% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{combineFunction}
\alias{combineFunction}
\title{Sequentially applies FUN to elements of data}
\usage{
combineFunction(data, FUN)
}
\arguments{
\item{data}{List of elements}

\item{FUN}{Function to apply, takes two arguments}
}
\value{
Returns aggregated output of applying FUN to data
}
\description{
This function applies and combines a function over successive elements of a list. For example, if \code{data} is a list of
\code{data.frame} and \code{FUN = rbind} it will return a \code{data.frame} of all the elements combined.
}

