% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{combineWithOtherCovariateRef}
\alias{combineWithOtherCovariateRef}
\title{Combines new covariateRef entries with those from demographics and predefined.}
\usage{
combineWithOtherCovariateRef(newCovariateRef, demographicsCovariateRef,
  predefinedCovariateRef, covariateIdIndex)
}
\arguments{
\item{newCovariateRef}{New covariate reference entries generated via \code{createNewCovRef}}

\item{demographicsCovariateRef}{Covariate reference entries for demographics covariates}

\item{predefinedCovariateRef}{Covariate reference entries for predefined covariates}

\item{covariateIdIndex}{Index from \code{covariateId} of data dimensions and demographics to a numeric index that fits format of
\code{cohortData}. Generated by \code{createCovariateIndex}, has columns \code{covariateId} and \code{index}.}
}
\value{
Returns combined \code{covariateRef}, with demographics covariates changed to new index.
}
\description{
This function combines the new \code{covariateRef} entries for the data dimensions with the demographics \code{covariateRef} entries
from cohortData. The demographics \code{covariateId} column is changed into the new index given in \code{covariateIdIndex}.
}

