% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{combineWithOtherCovariates}
\alias{combineWithOtherCovariates}
\title{Combines covariate information from data dimensions with that from demographics and predefined.}
\usage{
combineWithOtherCovariates(newCovariates, demographicsCovariates,
  predefinedCovariates)
}
\arguments{
\item{newCovariates}{An ffdf object with columns \code{rowId}, \code{covariateId}, \code{treatment}, \code{outcome}. This includes all
of the data dimensions and the expanded covariates with "_sporadic" and "_frequent" tags.}

\item{demographicsCovariates}{An ffdf object with columns \code{rowId}, \code{covariateId}, \code{covariateValue}.}

\item{predefinedCovariates}{An ffdf object with columns \code{rowId}, \code{covariateId}, \code{covariateValue}.}
}
\value{
Returns one ffdf object with columns \code{rowId}, \code{covariateId}, \code{covariateValue}.
}
\description{
This function combines the covariate information from all the data dimensions with that from demographics covariates and predefined covariates.
}
\details{
The \code{covariateValue} for the data dimensions is set to be 1. The \code{treatment} and \code{outcome} columns are dropped to match
the format of \code{cohortData}.
}

