% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Balance.R
\name{computeBalanceMetaAnalysis}
\alias{computeBalanceMetaAnalysis}
\title{Perform balance meta-analysis}
\usage{
computeBalanceMetaAnalysis(
  balances,
  threshold = 0.1,
  alpha = 0.05,
  maxCores = 1
)
}
\arguments{
\item{balances}{A list of balance data frames as generated by \code{\link[=computeCovariateBalance]{computeCovariateBalance()}}.}

\item{threshold}{Threshold value for the absolute value of the standardized difference of means (ASDM).
If the ASDM exceeds this threshold it will be marked as unbalanced. (Hripcsak et al. 2025)}

\item{alpha}{The family-wise alpha for testing whether the absolute value of the standardized
difference of means is greater than the threshold. If not provided, any value greater
than the threshold will be marked as unbalanced.}

\item{maxCores}{How many parallel cores should be used?}
}
\value{
A new balance data frame.
}
\description{
Perform balance meta-analysis
}
\details{
Perform a random-effects meta-analysis on the covariate balance standardized difference of means.
This is recommended when using balance hypothesis testing, as any single database may be
underpowered to detect imbalances that are systemic across the databases. (Hripcsak et al. 2025)
}
\references{
Hripcsak G, Zhang L, Chen Y, Li K, Suchard MA, Ryan PB, Schuemie MJ (2025)
Assessing Covariate Balance with Small Sample Sizes. Stat Med. 2025 Aug;44(18-19):e70212.
}
