% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PsFunctions.R
\name{computeCovariateBalance}
\alias{computeCovariateBalance}
\title{Compute covariate balance before and after matching and trimming}
\usage{
computeCovariateBalance(restrictedCohorts, cohortMethodData,
  outcomeConceptId = NULL)
}
\arguments{
\item{restrictedCohorts}{A data frame containing the people that are remaining after matching
and/or trimming.}

\item{cohortMethodData}{An object of type \code{cohortMethodData} as generated using
\code{getDbCohortMethodData}.}

\item{outcomeConceptId}{The concept ID of the outcome. Persons marked for removal for the outcome
                           will be removed when computing the balance before matching/trimming.}
}
\value{
Returns a date frame describing the covariate balance before and after matching/trimming.
}
\description{
For every covariate, prevalence in treatment and comparator groups before and after
matching/trimming are computed.
}
\details{
The restrictedCohorts data frame should have at least the following columns: \tabular{lll}{
\verb{rowId} \tab(integer) \tab A unique identifier for each row (e.g. the person ID) \cr
\verb{treatment} \tab(integer) \tab Column indicating whether the person is in the treated (1) or
comparator (0)\cr \tab \tab group \cr }
}

