% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrsFunctions.R
\name{computeDrsAuc}
\alias{computeDrsAuc}
\title{Compute the area under the ROC curve}
\usage{
computeDrsAuc(data, confidenceIntervals = FALSE)
}
\arguments{
\item{data}{A data frame with at least the two columns described below}

\item{confidenceIntervals}{Compute 95 percent confidence intervals (computationally expensive for
large data sets)}
}
\value{
A data frame holding the AUC and its 95 percent confidence interval
}
\description{
\code{computeDrsAuc} computes the area under the ROC curve of the disease risk score within the union of
the target and comparator cohorts.
}
\details{
The data frame should have a least the following two columns: \tabular{lll}{ \verb{outcomeCount}
\tab(integer) \tab Column counting the number of outcomes observed during the time at risk\cr \tab
\tab (0) group \cr \verb{diseaseRiskScore} \tab(numeric) \tab disease risk score \cr }
}
\examples{
outcomeCount <- rep(0:1, each = 100)
propensityScore <- c(rnorm(100, mean = 0.4, sd = 0.25), rnorm(100, mean = 0.6, sd = 0.25))
data <- data.frame(outcomeCount = outcomeCount, diseaseRiskScore = diseaseRiskScore)
data <- data[data$diseaseRiskScore > 0 & data$diseaseRiskScore < 1, ]
computeDrsAuc(data)

}
