% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{computeMdrr}
\alias{computeMdrr}
\title{Compute the minimum detectable relative risk}
\usage{
computeMdrr(population, alpha = 0.05, power = 0.8, twoSided = TRUE,
  modelType = "cox")
}
\arguments{
\item{population}{A data frame describing the study population as created using the
\code{\link{createStudyPopulation}} function. This should at least have these
columns: subjectId, treatment, outcomeCount, timeAtRisk.}

\item{alpha}{Type I error.}

\item{power}{1 - beta, where beta is the type II error.}

\item{twoSided}{Consider a two-sided test?}

\item{modelType}{The type of outcome model that will be used. Possible values are
"logistic", "poisson", or "cox". Currently only "cox" is supported.}
}
\value{
A data frame with the MDRR and some counts.
}
\description{
Compute the minimum detectable relative risk
}
\details{
Compute the minimum detectable relative risk (MDRR) and expected standard error (SE) for a given
study population, using the actual observed sample size and number of outcomes. Currently, only
computations for Cox models are implemented. For Cox model, the computations by Schoenfeld (1983) is used.
}
\references{
Schoenfeld DA (1983) Sample-size formula for the proportional-hazards regression model, Biometrics,
39(3), 499-503
}
