% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{convertCovariateId}
\alias{convertCovariateId}
\title{Converts covariateId to new index}
\usage{
convertCovariateId(covariates, index)
}
\arguments{
\item{covariates}{Covariates to convert to new index; covariateId is a factor. Has columns \code{rowId}, \code{covariateId}, \code{covariateValue}}

\item{index}{Index to use. Has column \code{covariateId} which match those from \code{covariates}, and column \code{index}}
}
\value{
Returns input \code{covariates} with \code{covariateId} converted to new index, which is numeric
}
\description{
Converts covariateId to new index
}

