% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convertToCyclopsDataObject}
\alias{convertToCyclopsDataObject}
\title{Convert data from two data frames or ffdf objects into a CyclopsData object}
\usage{
convertToCyclopsDataObject(outcomes, covariates, modelType = "lr",
  addIntercept = TRUE, offsetAlreadyOnLogScale = FALSE,
  makeCovariatesDense = NULL, checkSorting = TRUE, checkRowIds = TRUE,
  quiet = FALSE)
}
\arguments{
\item{outcomes}{A data frame or ffdf object containing the outcomes with predefined columns (see below).}

\item{covariates}{A data frame or ffdf object containing the covariates with predefined columns (see below).}

\item{modelType}{Cyclops model type. Current supported types are "pr", "cpr", lr", "clr", or "cox"}

\item{addIntercept}{Add an intercept to the model?}

\item{offsetAlreadyOnLogScale}{Is the time variable already on a log scale?}

\item{checkSorting}{Check if the data is sorted appropriately, and if not, sort.}

\item{checkRowIds}{Check if all rowIds in the covariates appear in the outcomes.}

\item{quiet}{If true, (warning) messages are surpressed.}

\item{useOffsetCovariate}{Use the time variable in the model as an offset?}
}
\value{
An object of type cyclopsData
}
\description{
\code{convertToCyclopsDataObject} loads data from two data frames or ffdf objects, and inserts it into a Cyclops data object.
}
\details{
These columns are expected in the outcome object:
\tabular{lll}{
  \verb{stratumId}    \tab(integer) \tab (optional) Stratum ID for conditional regression models \cr
  \verb{rowId}  	\tab(integer) \tab Row ID is used to link multiple covariates (x) to a single outcome (y) \cr
  \verb{y}    \tab(real) \tab The outcome variable \cr
  \verb{time}    \tab(real) \tab For models that use time (e.g. Poisson or Cox regression) this contains time (e.g. number of days) \cr
}

These columns are expected in the covariates object:
\tabular{lll}{
  \verb{stratumId}    \tab(integer) \tab (optional) Stratum ID for conditional regression models \cr
  \verb{rowId}  	\tab(integer) \tab Row ID is used to link multiple covariates (x) to a single outcome (y) \cr
  \verb{covariateId}    \tab(integer) \tab A numeric identifier of a covariate  \cr
  \verb{covariateValue}    \tab(real) \tab The value of the specified covariate \cr
}

Note: If checkSorting is turned off, the outcome table should be sorted by stratumId (if present)
and then rowId except for Cox regression when the table should be sorted by
stratumId (if present), -time, y, and rowId. The covariate table should be sorted by stratumId
(if present), rowId and covariateId except for Cox regression when the table should be sorted by
stratumId (if present), -time, y, and rowId.
}
\examples{
#Convert infert dataset to Cyclops format:
covariates <- data.frame(stratumId = rep(infert$stratum,2),
                         rowId = rep(1:nrow(infert),2),
                         covariateId = rep(1:2,each=nrow(infert)),
                         covariateValue = c(infert$spontaneous,infert$induced))
outcomes <- data.frame(stratumId = infert$stratum,
                       rowId = 1:nrow(infert),
                       y = infert$case)
#Make sparse:
covariates <- covariates[covariates$covariateValue != 0,]

#Create Cyclops data object:
cyclopsData <- convertToCyclopsDataObject(outcomes,covariates,modelType = "clr",addIntercept = FALSE)

#Fit model:
fit <- fitCyclopsModel(cyclopsData,prior = prior("none"))
}

