% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettingsObjects.R
\name{convertUntypedListToCmAnalysesSpecifications}
\alias{convertUntypedListToCmAnalysesSpecifications}
\title{Convert untyped list to SccsAnalysesSpecifications}
\usage{
convertUntypedListToCmAnalysesSpecifications(untypedList)
}
\arguments{
\item{untypedList}{A list of untyped objects. For example, these could be objects from a call
to \code{jsonlite::fromJSON()}. Importantly, \code{simplifyDataFrame} must be set to
\code{FALSE} when doing so.}
}
\value{
An object of type \code{SccsAnalysesSpecifications}.
}
\description{
Convert untyped list to SccsAnalysesSpecifications
}
