% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{covariateIdToFactor}
\alias{covariateIdToFactor}
\title{Converts covariateId of rawCodes to a factor}
\usage{
covariateIdToFactor(codes)
}
\arguments{
\item{codes}{Ffdf object with columns \code{SOURCE_CODE} and \code{TARGET_CONCEPT_ID}}
}
\value{
Returns the input codes with \code{SOURCE_CODE} as a factor
}
\description{
This function converts the SOURCE_CODE column acquired from the OMOP database to a factor, if it weren't already a factor
}

