% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createCcdData.ffdf}
\alias{createCcdData.ffdf}
\title{Convert data from two ffdf objects into a CcdData object}
\usage{
createCcdData.ffdf(outcomes, covariates, modelType = "lr",
  addIntercept = TRUE, useOffsetCovariate = NULL,
  offsetAlreadyOnLogScale = FALSE, sortCovariates = TRUE,
  makeCovariatesDense = NULL, batchBytes = getOption("ffbatchbytes"))
}
\arguments{
\item{outcomes}{A ffdf object containing the outcomes with predefined columns (see below).}

\item{covariateSql}{A ffdf object containing the covariates with predefined columns (see below).}

\item{modelType}{CCD model type. Current supported types are "ls", "pr", "lr", "clr", "sccs", or "cox"}

\item{addIntercept}{Add an intercept to the model?}

\item{useOffsetCovariate}{Use the time variable in the model as an offset?}

\item{offsetAlreadyOnLogScale}{Is the time variable already on a log scale?}

\item{sortCovariates}{Do the covariates still need to be sorted?}

\item{batchBytes}{Number of bytes to be read from the ffdf objects at a time. Larger batch sizes lead to fewer conversion calls
and should be more efficient, but may lead to out-of-memory errors.}
}
\value{
An object of type CcdModel
}
\description{
\code{createCcdData.ffdf} loads data from two ffdf objects, and inserts it into a CCD data object.
}
\details{
These columns are expected in the outcome object:
\tabular{lll}{
  \verb{row_id}  	\tab(integer) \tab Row ID is used to link multiple covariates (x) to a single outcome (y) \cr
  \verb{stratum_id}  	\tab(integer) \tab Stratum ID for conditional regression models \cr
  \verb{y}    \tab(real) \tab The outcome variable \cr
  \verb{time}    \tab(real) \tab For models that use time (e.g. Poisson or Cox regression) this contains time (e.g. number of days) \cr
}
The outcome table should be sorted by row_id

These columns are expected in the covariates object:
\tabular{lll}{
  \verb{row_id}  	\tab(integer) \tab Row ID is used to link multiple covariates (x) to a single outcome (y) \cr
  \verb{covariate_id}    \tab(integer) \tab A numeric identifier of a covariate  \cr
  \verb{covariate_value}    \tab(real) \tab The value of the specified covariate \cr
}
The covariate table should be sorted by row_id and covariate_id
}
\examples{
\dontrun{
  connectionDetails <- createConnectionDetails(dbms="sql server", server="RNDUSRDHIT07.jnj.com", schema="test")
  connection <- connect(connectionDetails)

  outcomes <- dbGetQuery.ffdf(conn,"SELECT * FROM outcomes ORDER BY row_id")
  covariates <- dbGetQuery.ffdf(conn,"SELECT * FROM covariates ORDER BY row_id, covariate_id")

  ccdData <- dbGetCcdInput(connection,outcomes,covariates,modelType = "lr")

  dbDisconnect(connection)

  ccdFit <- fitCcdModel(ccdData, prior = prior("normal",0.01))
}
}

