% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettingsObjects.R
\name{createCmAnalysis}
\alias{createCmAnalysis}
\title{Create a CohortMethod analysis specification}
\usage{
createCmAnalysis(
  analysisId = 1,
  description = "",
  getDbCohortMethodDataArgs,
  createStudyPopulationArgs,
  createPsArgs = NULL,
  trimByPsArgs = NULL,
  truncateIptwArgs = NULL,
  matchOnPsArgs = NULL,
  stratifyByPsArgs = NULL,
  computeSharedCovariateBalanceArgs = NULL,
  computeCovariateBalanceArgs = NULL,
  fitOutcomeModelArgs = NULL
)
}
\arguments{
\item{analysisId}{An integer that will be used later to refer to this specific
set of analysis choices.}

\item{description}{A short description of the analysis.}

\item{getDbCohortMethodDataArgs}{An object representing the arguments to be used when calling
the \code{\link[=getDbCohortMethodData]{getDbCohortMethodData()}} function.}

\item{createPsArgs}{An object representing the arguments to be used when calling
the \code{\link[=createPs]{createPs()}} function.}

\item{trimByPsArgs}{An object representing the arguments to be used when calling
the \code{\link[=trimByPs]{trimByPs()}} function.}

\item{truncateIptwArgs}{An object representing the arguments to be used when calling
the \code{\link[=truncateIptw]{truncateIptw()}} function.}

\item{matchOnPsArgs}{An object representing the arguments to be used when calling
the \code{\link[=matchOnPs]{matchOnPs()}} function.}

\item{stratifyByPsArgs}{An object representing the arguments to be used when calling
the \code{\link[=stratifyByPs]{stratifyByPs()}} function.}

\item{computeSharedCovariateBalanceArgs}{An object representing the arguments to be used when calling
the \code{\link[=computeCovariateBalance]{computeCovariateBalance()}} function per target-comparator-analysis.}

\item{computeCovariateBalanceArgs}{An object representing the arguments to be used when calling
the \code{\link[=computeCovariateBalance]{computeCovariateBalance()}} function per target-comparator-outcome-analysis.}

\item{fitOutcomeModelArgs}{An object representing the arguments to be used when calling
the \code{\link[=fitOutcomeModel]{fitOutcomeModel()}} function.}

\item{createStudyPopArgs}{An object representing the arguments to be used when calling
the \code{\link[=createStudyPopulation]{createStudyPopulation()}} function.}
}
\value{
An object of type \code{CmAnalysis}, to be used with the \link{runCmAnalyses} function.
}
\description{
Create a CohortMethod analysis specification
}
\details{
Create a set of analysis choices, to be used with the \code{\link[=runCmAnalyses]{runCmAnalyses()}} function.

Providing a NULL value for any of the argument applies the corresponding step will not be executed.
For example, if \code{createPsArgs = NULL}, no propensity scores will be computed.
}
