% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettingsObjects.R
\name{createCmDiagnosticThresholds}
\alias{createCmDiagnosticThresholds}
\title{Create CohortMethod diagnostics thresholds}
\usage{
createCmDiagnosticThresholds(
  mdrrThreshold = 10,
  easeThreshold = 0.25,
  sdmThreshold = 0.1,
  equipoiseThreshold = 0.2,
  generalizabilitySdmThreshold = 1
)
}
\arguments{
\item{mdrrThreshold}{What is the maximum allowed minimum detectable relative risk
(MDRR)?}

\item{easeThreshold}{What is the maximum allowed expected absolute systematic error
(EASE).}

\item{sdmThreshold}{What is the maximum allowed standardized difference of mean (SDM)? If
any covariate has an SDM exceeding this threshold, the diagnostic will
fail.}

\item{equipoiseThreshold}{What is the minimum required equipoise?}

\item{generalizabilitySdmThreshold}{What is the maximum allowed standardized difference of mean
(SDM)when comparing the population before and after PS
adjustments? If the SDM is greater than this value, the diagnostic
will fail.}
}
\value{
An object of type \code{CmDiagnosticThresholds}.
}
\description{
Threshold used when calling \code{\link[=exportToCsv]{exportToCsv()}} to determine if we pass or fail diagnostics.
}
