% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalyses.R
\name{createCmMultiThreadingSettings}
\alias{createCmMultiThreadingSettings}
\title{Create CohortMethod multi-threading settings}
\usage{
createCmMultiThreadingSettings(
  getDbCohortMethodDataThreads = 1,
  createPsThreads = 1,
  psCvThreads = 1,
  createStudyPopThreads = 1,
  trimMatchStratifyThreads = 1,
  computeSharedBalanceThreads = 1,
  computeBalanceThreads = 1,
  prefilterCovariatesThreads = 1,
  fitOutcomeModelThreads = 1,
  outcomeCvThreads = 1
)
}
\arguments{
\item{getDbCohortMethodDataThreads}{The number of parallel threads to use for building the
cohortMethod data objects.}

\item{createPsThreads}{The number of parallel threads to use for fitting the
propensity models.}

\item{psCvThreads}{The number of parallel threads to use for the cross-
validation when estimating the hyperparameter for the
propensity model. Note that the total number of CV threads at
one time could be \code{createPsThreads * psCvThreads}.}

\item{createStudyPopThreads}{The number of parallel threads to use for creating the study
population.}

\item{trimMatchStratifyThreads}{The number of parallel threads to use for trimming, matching
and stratifying.}

\item{computeSharedBalanceThreads}{The number of parallel threads to use for computing shared covariate
balance.}

\item{computeBalanceThreads}{The number of parallel threads to use for computing covariate
balance.}

\item{prefilterCovariatesThreads}{The number of parallel threads to use for prefiltering covariates.}

\item{fitOutcomeModelThreads}{The number of parallel threads to use for fitting the outcome
models.}

\item{outcomeCvThreads}{The number of parallel threads to use for the cross-
validation when estimating the hyperparameter for the outcome
model. Note that the total number of CV threads at one time
could be \code{fitOutcomeModelThreads * outcomeCvThreads}.}
}
\value{
An object of type \code{CmMultiThreadingSettings}.
}
\description{
Create CohortMethod multi-threading settings
}
\seealso{
\code{\link[=createDefaultCmMultiThreadingSettings]{createDefaultCmMultiThreadingSettings()}}
}
