% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1.R
\name{createCmTable1}
\alias{createCmTable1}
\title{Create a table 1}
\usage{
createCmTable1(balance,
  specifications = getDefaultCmTable1Specifications(),
  beforeTargetPopSize, beforeComparatorPopSize, afterTargetPopSize,
  afterComparatorPopSize, beforeLabel = "Before matching",
  afterLabel = "After matching", targetLabel = "Target",
  comparatorLabel = "Comparator", percentDigits = 1,
  stdDiffDigits = 2)
}
\arguments{
\item{balance}{A data frame created by the \code{computeCovariateBalance} funcion.}

\item{specifications}{Specifications of which covariates to display, and how.}

\item{beforeTargetPopSize}{The number of people in the target cohort before matching/stratification/trimming,
to mention in the table header. If not provide, no number will be included in the header.}

\item{beforeComparatorPopSize}{The number of people in the comparator cohort before matching/stratification/trimming,
to mention in the table header. If not provide, no number will be included in the header.}

\item{afterTargetPopSize}{The number of people in the target cohort after matching/stratification/trimming,
to mention in the table header. If not provide, no number will be included in the header.}

\item{afterComparatorPopSize}{The number of people in the comparator cohort after matching/stratification/trimming,
to mention in the table header. If not provide, no number will be included in the header.}

\item{beforeLabel}{Label for identifying columns before matching / stratification / trimming.}

\item{afterLabel}{Label for identifying columns after matching / stratification / trimming.}

\item{targetLabel}{Label for identifying columns of the target cohort.}

\item{comparatorLabel}{Label for identifying columns of the comparator cohort.}

\item{percentDigits}{Number of digits to be used for percentages.}

\item{stdDiffDigits}{Number of digits to be used for the standardized differences.}
}
\value{
A data frame with the formatted table 1.
}
\description{
Creates a formatted table of cohort characteristics, to be included in publications or reports.
}
