% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Simulation.R
\name{createCohortDataSimulationProfile}
\alias{createCohortDataSimulationProfile}
\title{Create simulation profile}
\usage{
createCohortDataSimulationProfile(cohortData)
}
\arguments{
\item{cohortData}{An object of type \code{cohortData} as generated using \code{getDbCohortData}.}
}
\value{
An object of type \code{cohortDataSimulationProfile}.
}
\description{
\code{createCohortDataSimulationProfile} creates a profile based on the provided cohortData object,
which can be used to generate simulated data that has similar characteristics.
}
\details{
The output of this function is an object that can be used by the \code{simulateCohortData} function
to generate a cohortData object.
}

