% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettingsObjects.R
\name{createComputeRiskDifferenceArgs}
\alias{createComputeRiskDifferenceArgs}
\title{Create a parameter object for the function \code{\link[=fitOutcomeModel]{fitOutcomeModel()}} for computing the risk difference}
\usage{
createComputeRiskDifferenceArgs(timePoint = 365)
}
\arguments{
\item{timePoint}{The point in time (days relative to index date) at which the cumulative risk difference should be
computed.}
}
\value{
An object of type \code{ComputeRiskDifferenceArgs}.
}
\description{
Create a parameter object for the function \code{\link[=fitOutcomeModel]{fitOutcomeModel()}} for computing the risk difference
}
\details{
Create an object defining the parameter values.
}
