% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{createCovariateIndex}
\alias{createCovariateIndex}
\title{Creates numeric covariateId index}
\usage{
createCovariateIndex(covariates)
}
\arguments{
\item{covariates}{Covariate information of combined data dimension and demographic data. Includes column \code{covariateId}}
}
\value{
Returns ffdf object with columns \code{covariateId} and \code{index}
}
\description{
Creates numeric covariateId index from covariates.
}

