% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createCreateDrsArgs}
\alias{createCreateDrsArgs}
\title{Create a parameter object for the function createDrs}
\usage{
createCreateDrsArgs(modelType = "logistic")
}
\arguments{
\item{modelType}{The type of disease risk model. Can be either 'logistic' or 'survival'.}
}
\description{
Create a parameter object for the function createDrs
}
\details{
Create an object defining the parameter values.
}
