% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettingsObjects.R
\name{createCreateStudyPopulationArgs}
\alias{createCreateStudyPopulationArgs}
\title{Create a parameter object for the function \code{\link[=createStudyPopulation]{createStudyPopulation()}}}
\usage{
createCreateStudyPopulationArgs(
  removeSubjectsWithPriorOutcome = TRUE,
  priorOutcomeLookback = 99999,
  minDaysAtRisk = 1,
  maxDaysAtRisk = 99999,
  riskWindowStart = 0,
  startAnchor = "cohort start",
  riskWindowEnd = 0,
  endAnchor = "cohort end",
  censorAtNewRiskWindow = FALSE
)
}
\arguments{
\item{removeSubjectsWithPriorOutcome}{Remove subjects that have the outcome prior to the risk window start?}

\item{priorOutcomeLookback}{How many days should we look back when identifying prior outcomes?}

\item{minDaysAtRisk}{The minimum required number of days at risk. Risk windows with fewer days than this number are removed from the analysis.}

\item{maxDaysAtRisk}{The maximum allowed number of days at risk. Risk windows that are longer will be truncated to this number of days.}

\item{riskWindowStart}{The start of the risk window (in days) relative to the startAnchor.}

\item{startAnchor}{The anchor point for the start of the risk window. Can be "cohort start" or "cohort end".}

\item{riskWindowEnd}{The end of the risk window (in days) relative to the endAnchor.}

\item{endAnchor}{The anchor point for the end of the risk window. Can be "cohort start" or "cohort end".}

\item{censorAtNewRiskWindow}{If a subject is in multiple cohorts, should time-at-risk be censored when the new time-at-risk starts to prevent overlap?}
}
\value{
An object of type \code{CreateStudyPopulationArgs}.
}
\description{
Create a parameter object for the function \code{\link[=createStudyPopulation]{createStudyPopulation()}}
}
\details{
Create an object defining the parameter values.
}
