% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsFunctions.R
\name{createDefaultBayesSettings}
\alias{createDefaultBayesSettings}
\title{Settings object for Bayesian Bridge model}
\usage{
createDefaultBayesSettings(
  n_iter = 10000,
  n_burnin = 1000,
  bridge_exponent = 0.5,
  regularizing_slab_size = 1,
  thin = 1,
  n_status_update = 100,
  init = list(global_scale = 0.1),
  coef_sampler_type = "cg",
  params_to_fix = c(),
  local_scale_sampler_type = "all",
  params_to_save = c("coef", "global_scale", "logp"),
  fixed_effects = NULL,
  mixture = NULL
)
}
\description{
Returns a default list of settings for a Bayesian Bridge logistic model to fit the propensity score.
}
