% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalyses.R
\name{createDefaultMultiThreadingSettings}
\alias{createDefaultMultiThreadingSettings}
\title{Create default CohortMethod multi-threading settings}
\usage{
createDefaultMultiThreadingSettings(maxCores)
}
\arguments{
\item{maxCores}{Maximum number of CPU cores to use.}
}
\value{
An object of type \code{CmMultiThreadingSettings}.
}
\description{
Create CohortMethod multi-threading settings based on the maximum number of cores to be
used.
}
\examples{
settings <- createDefaultMultiThreadingSettings(10)

}
\seealso{
\code{\link[=createMultiThreadingSettings]{createMultiThreadingSettings()}}
}
