% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrsFunctions.R
\name{createDrs}
\alias{createDrs}
\title{Create disease risk scores}
\usage{
createDrs(
  cohortMethodData,
  population = NULL,
  coefficients,
  modelType = "logistic"
)
}
\arguments{
\item{cohortMethodData}{An object of type \code{cohortMethodData} as generated using
\code{getDbCohortMethodData}.}

\item{population}{A data frame describing the population. This should at least have a
'rowId' column corresponding to the rowId column in the
\code{cohortMethodData} covariates object and a 'treatment' column.
If population is not specified, the full population in the
\code{cohortMethodData} will be used.}

\item{coefficients}{When using an external disease risk model, the coefficients. See details for
requirements.}

\item{modelType}{The type of disease risk model. Can be either 'logistic' or 'survival'.}
}
\description{
\code{createDrs} creates disease risk scores.
}
\details{
The \code{coefficients} argument must be a named vector, where the first element is the intercept, and for
the other elements the names are the covariate IDs.
}
