% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Analyses.R
\name{createDrugComparatorOutcomes}
\alias{createDrugComparatorOutcomes}
\title{Create drug-comparator-outcomes combinations.}
\usage{
createDrugComparatorOutcomes(targetDrugConceptId, comparatorDrugConceptId,
  outcomeConceptIds, indicationConceptIds = c(), exclusionConceptIds = c(),
  excludedCovariateConceptIds = c(), includedCovariateConceptIds = c())
}
\arguments{
\item{targetDrugConceptId}{A concept ID indentifying the target drug in the exposure
table. If multiple strategies for picking the target will be
tested in the analysis, a named list of numbers can be provided
instead. In the analysis, the name of the number to be used can
be specified using the #' \code{targetType} parameter in the
\code{\link{createCmAnalysis}} function.}

\item{comparatorDrugConceptId}{A concept ID indentifying the comparator drug in the exposure
table. If multiple strategies for picking the comparator will
be tested in the analysis, a named list of numbers can be
provided instead. In the analysis, the name of the number to be
used can be specified using the #' \code{comparatorType}
parameter in the \code{\link{createCmAnalysis}} function.}

\item{outcomeConceptIds}{A vector of concept IDs indentifying the outcome(s) in the
outcome table.}

\item{indicationConceptIds}{A vector of concept IDs identifying conditions that are
required to appear prior to or on the index date. If multiple
strategies for picking the indication will be tested in the
analysis, a named list of vectors can be provided instead. In
the analysis, the name of the vector to be used can be
specified using the \code{indicationType} parameter in the
\code{\link{createCmAnalysis}} function.}

\item{exclusionConceptIds}{A list of concept IDs that cannot appear on or before the index
date. This argument is to be used only for exclusion criteria
that are specific to the drug-comparator combination.}

\item{excludedCovariateConceptIds}{A list of concept IDs that cannot be used to construct
covariates. This argument is to be used only for exclusion
concepts that are specific to the drug-comparator combination.}

\item{includedCovariateConceptIds}{A list of concept IDs that must be used to construct
                                     covariates. This argument is to be used only for inclusion
                                     concepts that are specific to the drug-comparator combination.}
}
\description{
Create drug-comparator-outcomes combinations.
}
\details{
Create a set of hypotheses of interest, to be used with the \code{\link{runCmAnalyses}} function.
}

