% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createGetDbCohortMethodDataArgs}
\alias{createGetDbCohortMethodDataArgs}
\title{Create a parameter object for the function getDbCohortMethodData}
\usage{
createGetDbCohortMethodDataArgs(washoutWindow = 183,
  indicationLookbackWindow = 183, studyStartDate = "", studyEndDate = "",
  exclusionConceptIds = c(), outcomeConditionTypeConceptIds = c(),
  cdmVersion = "4", excludeDrugsFromCovariates = TRUE, covariateSettings)
}
\arguments{
\item{washoutWindow}{The mininum required continuous observation time prior to
indexdate for a person to be included in the cohort.}

\item{indicationLookbackWindow}{NA}

\item{studyStartDate}{A calendar date specifying the minimum date that a cohort
indexdate can appear. Date format is 'yyyymmdd'.}

\item{studyEndDate}{A calendar date specifying the maximum date that a cohort
indexdate can appear. Date format is 'yyyymmdd'.}

\item{exclusionConceptIds}{A list of CONCEPT_IDs used to restrict the cohorts, based on
anydescendant conditions/drugs/procedures occurring at least
onceanytime prior to the cohort index date.}

\item{outcomeConditionTypeConceptIds}{A list of TYPE_CONCEPT_ID values that will restrictcondition
occurrences.  Only applicable if outcomeTable
=CONDITION_OCCURRENCE.}

\item{cdmVersion}{Define the OMOP CDM version used: currently support "4" and
"5".}

\item{excludeDrugsFromCovariates}{Should the target and comparator drugs (and their
descendantconcepts) be excluded from the covariates? Note
that this willwork if the drugs are actualy drug concept IDs
(and not cohortIDs).}

\item{covariateSettings}{An object of type covariateSettings as created using
                                        thecreateCovariateSettings function in the
                                        PatientLevelPrediction package..}
}
\description{
Create a parameter object for the function getDbCohortMethodData
}
\details{
Create an object defining the parameter values.
}

