% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createMatchOnPsAndCovariatesArgs}
\alias{createMatchOnPsAndCovariatesArgs}
\title{Create parameter object for the function matchOnPsAndCovariates}
\usage{
createMatchOnPsAndCovariatesArgs(caliper = 0.25,
  caliperScale = "standardized", maxRatio = 1, covariateIds)
}
\arguments{
\item{caliper}{The caliper for matching. A caliper is the distance which is acceptable forany
match. Observations which are outside of the caliper are dropped. Acaliper of
0 means no caliper is used.}

\item{caliperScale}{The scale on which the caliper is defined. Two scales
aresupported:caliperScale = 'propensity score' or caliperScale
='standardized'. On the standardized scale, the caliper is interpreted
instandard deviations of the propensity score distribution.}

\item{maxRatio}{The maximum number of persons int the comparator arm to be matched to
eachperson in the treatment arm. A maxRatio of 0 means no maximum: all
comparatorswill be assigned to a treated person.}

\item{covariateIds}{One or more covariate IDs in the cohortMethodData object on which
                      subjectsshould be also matched.}
}
\description{
Create parameter object for the function matchOnPsAndCovariates
}
\details{
Create an object defining the parameter values. To be used in the
\code{\link{createCohortMethodAnalysis}} function.
}

