% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{createNewCovRef}
\alias{createNewCovRef}
\title{Generates new covariateRef entries}
\usage{
createNewCovRef(data, dimensionInfo, rawCodes, covariateIdIndex, cohortData)
}
\arguments{
\item{data}{Covariate information for data dimension, as ffdf object with columns \code{rowId}, \code{covariateId},
\code{treatment}, \code{outcome}}

\item{dimensionInfo}{The dimensionTable entry for data dimension, which includes the \code{truncatedNames} entry}

\item{rawCodes}{An ffdf object of the mapping from concept ID (column \code{TARGET_CONCEPT_ID}) to data dimension code
(column \code{SOURCE_CODE})}

\item{covariateIdIndex}{Index to convert \code{covariateId} of \code{data}, which is type \code{factor}, to a numeric index.}

\item{cohortData}{\code{cohortData} object constructed by \code{getDbCohortData}}
}
\value{
Returns ffdf object with columns: \describe{
\item{covariateId}{Newly indexed covariate identifier}
\item{covariateName}{New name for covariate, with appropriate recurrence tag}
\item{analysisId}{Analysis identifier for data dimension}
\item{conceptId}{Concept identifier for covariate. 0 if code was truncated} }
}
\description{
This function creates \code{covariateRef} entries for each data dimension. The new \code{covariateName} is taken from cohortData, if there
was no truncation of codes, or generated via the \code{truncatedNames} entry in \code{dimensionInfo}. Additionally, the appropriate
marker of "(once)", "(sporadic)", or "(frequent)" is appended to the beginning of the \code{covariateName}.
}
\details{
If there is no \code{truncatedNames} for a truncated code, the \code{covariateName} is "covariate name not available."
Furthermore, if a code was truncated for a data dimension, then \code{conceptId} is meaningless and will display 0.
}

