% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{createResultsDataModel}
\alias{createResultsDataModel}
\title{Create the results data model tables on a database server.}
\usage{
createResultsDataModel(
  connectionDetails = NULL,
  databaseSchema,
  tablePrefix = ""
)
}
\arguments{
\item{connectionDetails}{DatabaseConnector connectionDetails instance @seealso\link[DatabaseConnector:createConnectionDetails]{DatabaseConnector::createConnectionDetails}}

\item{databaseSchema}{The schema on the server where the tables will be created.}

\item{tablePrefix}{(Optional)  string to insert before table names for database table names}
}
\description{
Create the results data model tables on a database server.
}
\details{
Only PostgreSQL and SQLite servers are supported.
}
