% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createStratifyByDrsArgs}
\alias{createStratifyByDrsArgs}
\title{Create a parameter object for the function stratifyByDrs}
\usage{
createStratifyByDrsArgs(
  numberOfStrata = 5,
  stratificationColumns = c(),
  baseSelection = "all"
)
}
\arguments{
\item{numberOfStrata}{How many strata? The boundaries of the strata are
automaticallydefined to contain equal numbers of target persons.}

\item{stratificationColumns}{Names of one or more columns in the data data.frame on whichsubjects
should also be stratified in addition to stratification ondisease
risk score.}

\item{baseSelection}{What is the base selection of subjects where the strata bounds areto
be determined? Strata are defined as equally-sized strata insidethis
selection. Possible values are "all", "target", and "comparator".}
}
\description{
Create a parameter object for the function stratifyByDrs
}
\details{
Create an object defining the parameter values.
}
