% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyPopulation.R
\name{createStudyPopulation}
\alias{createStudyPopulation}
\title{Create a study population}
\usage{
createStudyPopulation(cohortMethodData, outcomeId, firstExposureOnly = TRUE,
  washoutPeriod = 180, removeDuplicateSubjects = TRUE,
  removeSubjectsWithPriorOutcome = TRUE, priorOutcomeLookback = 99999,
  riskWindowStart = 0, addExposureDaysToStart = FALSE, riskWindowEnd = 0,
  addExposureDaysToEnd = TRUE, modelType = "logistic",
  dropExtraColumns = TRUE)
}
\arguments{
\item{cohortMethodData}{An object of type \code{cohortMethodData} as generated using
\code{getDbCohortMethodData}.}

\item{outcomeId}{The  ID of the outcome.}

\item{firstExposureOnly}{Should only the first exposure per subject be included? Note that
this is typically done in the \code{createStudyPopulation} function,}

\item{washoutPeriod}{The mininum required continuous observation time prior to index
date for a person to be included in the cohort.}

\item{removeDuplicateSubjects}{Remove subjects that are in both the treated and comparator cohort?}

\item{removeSubjectsWithPriorOutcome}{Remove subjects that have the outcome prior to the risk window start?}

\item{priorOutcomeLookback}{How many days should we look back when identifying prior outcomes?}

\item{riskWindowStart}{The start of the risk window (in days) relative to the index date (+
days of exposure if the \code{addExposureDaysToStart} parameter is
specified).}

\item{addExposureDaysToStart}{Add the length of exposure the start of the risk window?}

\item{riskWindowEnd}{The end of the risk window (in days) relative to the index data (+
days of exposure if the \code{addExposureDaysToEnd} parameter is
specified).}

\item{addExposureDaysToEnd}{Add the length of exposure the risk window?}

\item{modelType}{The type of outcome model that will be used. Possible values are
"logistic", "poisson", or "cox".}
}
\value{
A data frame specifying the study population. This data frame will have the following columns:
\describe{
\item{rowId}{A unique identifier for an exposure}
\item{subjectId}{The person ID of the subject}
\item{cohortStartdate}{The index date}
\item{outcome}{The outcome variable, which value depends on the modelType}
\item{time}{Optional: for survival models this is the time to event or end of risk windows, for poisson models this is the time to the end of the risk window.}
}
}
\description{
Create a study population
}
\details{
Create a study population by enforcing certain inclusion and exclusion criteria, defining
a risk window, and determining which outcomes fall inside the risk window.
}

