% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createTrimByIptwArgs}
\alias{createTrimByIptwArgs}
\title{Create a parameter object for the function trimByIptw}
\usage{
createTrimByIptwArgs(maxWeight = 10, estimator = "ate")
}
\arguments{
\item{maxWeight}{The maximum allowed IPTW.}

\item{estimator}{The type of estimator. Options are estimator = "ate" for the average treatment effect, and estimator = "att"for the average treatment effect in the treated.}
}
\description{
Create a parameter object for the function trimByIptw
}
\details{
Create an object defining the parameter values.
}
