% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createTrimByPsToEquipoiseArgs}
\alias{createTrimByPsToEquipoiseArgs}
\title{Create parameter object for the function trimByPsToEquipoise}
\usage{
createTrimByPsToEquipoiseArgs(bounds = c(0.25, 0.75))
}
\arguments{
\item{bounds}{The upper and lower bound on the preference score for keeping persons}
}
\description{
Create parameter object for the function trimByPsToEquipoise
}
\details{
Create an object defining the parameter values. To be used in the
\code{\link{createCohortMethodAnalysis}} function.
}

