% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettingsObjects.R
\name{createTruncateIptwArgs}
\alias{createTruncateIptwArgs}
\title{Create a parameter object for the function \code{\link[=truncateIptw]{truncateIptw()}}}
\usage{
createTruncateIptwArgs(maxWeight = 10)
}
\arguments{
\item{maxWeight}{The maximum allowed IPTW.}
}
\value{
An object of type \code{TruncateIptwArgs}.
}
\description{
Create a parameter object for the function \code{\link[=truncateIptw]{truncateIptw()}}
}
\details{
Create an object defining the parameter values.
}
