% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{deletePredefinedCodes}
\alias{deletePredefinedCodes}
\title{Removes predefined codes from data dimensions}
\usage{
deletePredefinedCodes(rawCodes, predefinedConceptIds)
}
\arguments{
\item{rawCodes}{Ffdf object with columns: \code{SOURCE_CODE} and \code{TARGET_CONCEPT_ID}}

\item{predefinedConceptIds}{Vector of concept identifiers for predefined covariates}
}
\value{
Returns \code{rawCodes} with all instances of (\code{TARGET_CONCEPT_ID} = predefined covariate) removed
}
\description{
Removes predefined codes from data dimensions
}

