% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{deleteRepeatCodes}
\alias{deleteRepeatCodes}
\title{Keeps only unique code mappings}
\usage{
deleteRepeatCodes(codes)
}
\arguments{
\item{codes}{\code{ffdf} with columns: \code{SOURCE_CODE} and \code{TARGET_CONCEPT_ID}}
}
\value{
Returns \code{codes} with relevant rows deleted.
}
\description{
This function deletes duplicate entries in the truncated codes, and also ambiguous mappings from one
\code{TARGET_CONCEPT_ID} to multiple \code{SOURCE_CODE}
}

