% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{.truncateSimulationProfile}
\alias{.truncateSimulationProfile}
\title{truncateSimulationProfile}
\usage{
.truncateSimulationProfile(profile, minCellCount = 5)
}
\arguments{
\item{profile}{Object of class \code{CohortDataSimulationProfile}}

\item{minCellCount}{Number of cases below which prevalence will be set to zero}
}
\value{
Modified copy of supplied simulation profile
}
\description{
Set to zero all low-prevalence covariates in the supplied simulation table
in order to prevent identification of persons
}
\details{
Truncate low-prevalence covariates in a simulation profile
}
