% Generated by roxygen2 (4.0.1): do not edit by hand
\name{estimateEffect}
\alias{estimateEffect}
\title{Create an outcome model, and compute the relative risk}
\usage{
estimateEffect(cohortData, strata = NULL, riskWindowStart = 0,
  riskWindowEnd = 9999, addExposureDaysToEnd = FALSE,
  useCovariates = TRUE, modelType = "clr")
}
\arguments{
\item{cohortData}{An object of type \code{cohortData} as generated using \code{dbGetCohortData}}

\item{strata}{A data frame specifying the strata. This data frame should have at least the following columns:
\code{RowId} and \code{StratumId}}

\item{riskWindowEnd}{The maximum length (in days) of the risk window.}

\item{useCovariates}{Whether to use the covariate matrix in the cohortData in the outcome model}

\item{modelType}{The type of model to be fitted. See details for options}
}
\value{
A data frame holding the effect estimate
}
\description{
\code{estimateEffect} creates an outcome model, and computes the relative risk
}
\details{
The data frame should have a least the following two columns:
\tabular{ll}{
  \verb{lr}      \tab Logistic regression  \cr
  \verb{clr}     \tab Conditional logistic regression \cr
  \verb{cox}     \tab Cox regression (conditional or unconditional, depending on whether \code{stata} is specified) \cr
  \verb{pr}      \tab Poisson regression  \cr
  \verb{cpr}     \tab Conditional Poisson regression \cr
}
}
\examples{
#todo
}

