% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{expandCovariates}
\alias{expandCovariates}
\title{Adds new covariates for sporadic and frequent occurrences}
\usage{
expandCovariates(data)
}
\arguments{
\item{data}{Covariate information for data dimension, as ffdf object with columns \code{rowId}, \code{covariateId},
\code{treatment}, \code{outcome}}
}
\value{
Returns data as ffdf object with same columns, with added lines for sporadic and frequent codes. Sporadic codes are designated by
appending "_sporadic" to the \code{covariateId}, while frequent codes are designated by appending "_frequent" to the \code{covariateId}.
}
\description{
This function adds new covariates named sporadic and frequent according to hdps implementation, for a data dimension.
}
\details{
Sporadic indicates a code occurs for a patient more than the median occurrences for that code; frequent indicates more than 75th percentile.
Once the new codes are generated, repeated codes in the original data are dropped so that "occurs at least once" is a binary variable.
}

