% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyPopulation.R
\name{getAttritionTable}
\alias{getAttritionTable}
\title{Get the attrition table for a population}
\usage{
getAttritionTable(object)
}
\arguments{
\item{object}{Either an object of type \link{CohortMethodData}, a population object generated by
functions like \code{\link[=createStudyPopulation]{createStudyPopulation()}}, or an object of type
\code{outcomeModel}.}
}
\value{
A \code{tibble} specifying the number of people and exposures in the population after specific steps
of filtering.
}
\description{
Get the attrition table for a population
}
