% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsFunctions.R
\name{getBayesPs}
\alias{getBayesPs}
\title{Get propensity score using Bayesian Bridge model}
\usage{
getBayesPs(covariates, outcomes, settings)
}
\description{
Returns the samples from the Gibbs sampler as well as a propensity score estimate from prediction of model
}
