% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{getConceptId}
\alias{getConceptId}
\title{Retrieves conceptId that matches analysisId}
\usage{
getConceptId(analysisId, cohortData)
}
\arguments{
\item{analysisId}{Analysis ID for data dimension}

\item{cohortData}{\code{cohortData} object constructed by \code{getDbCohortData}}
}
\value{
Returns a \code{ff_vector} of type \code{double} of concept IDs
}
\description{
Retrieves list of concept IDs that match the analysisId for a data dimension. These concept IDs will be used to query the CDM for
the source codes for that data dimension.
}

