% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{getCovariateId}
\alias{getCovariateId}
\title{Retrieves covariateId that matches analysisId}
\usage{
getCovariateId(analysisId, cohortData)
}
\arguments{
\item{analysisId}{Analysis ID for data dimension}

\item{cohortData}{\code{cohortData} object constructed by \code{getDbCohortData}}
}
\value{
Returns a \code{ff_vector} of type \code{double} of covariate IDs
}
\description{
Retrieves list of covariate IDs that match the analysisId for a data dimension.
}

