% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDbCohortData}
\alias{getDbCohortData}
\title{Get the cohort data from the server}
\usage{
getDbCohortData(connectionDetails, cdmSchema = "CDM4_SIM",
  resultsSchema = "scratch", targetDrugConceptId = 755695,
  comparatorDrugConceptId = 739138, indicationConceptIds = 439926,
  washoutWindow = 183, indicationLookbackWindow = 183,
  studyStartDate = "", studyEndDate = "", exclusionConceptIds = c(4027133,
  4032243, 4146536, 2002282, 2213572, 2005890, 43534760, 21601019),
  outcomeConceptIds = 194133, outcomeConditionTypeConceptIds = c(38000215,
  38000216, 38000217, 38000218, 38000183, 38000232),
  exposureSchema = cdmSchema, exposureTable = "drug_era",
  outcomeSchema = cdmSchema, outcomeTable = "condition_occurrence",
  useCovariateDemographics = TRUE, useCovariateConditionOccurrence = TRUE,
  useCovariateConditionOccurrence365d = TRUE,
  useCovariateConditionOccurrence30d = TRUE,
  useCovariateConditionOccurrenceInpt180d = TRUE,
  useCovariateConditionEra = FALSE, useCovariateConditionEraEver = FALSE,
  useCovariateConditionEraOverlap = FALSE,
  useCovariateConditionGroup = FALSE, useCovariateDrugExposure = FALSE,
  useCovariateDrugExposure365d = FALSE, useCovariateDrugExposure30d = FALSE,
  useCovariateDrugEra = FALSE, useCovariateDrugEra365d = FALSE,
  useCovariateDrugEra30d = FALSE, useCovariateDrugEraOverlap = FALSE,
  useCovariateDrugEraEver = FALSE, useCovariateDrugGroup = FALSE,
  useCovariateProcedureOccurrence = FALSE,
  useCovariateProcedureOccurrence365d = FALSE,
  useCovariateProcedureOccurrence30d = FALSE,
  useCovariateProcedureGroup = FALSE, useCovariateObservation = FALSE,
  useCovariateObservation365d = FALSE, useCovariateObservation30d = FALSE,
  useCovariateObservationBelow = FALSE,
  useCovariateObservationAbove = FALSE,
  useCovariateObservationCount365d = FALSE,
  useCovariateConceptCounts = FALSE, useCovariateRiskScores = FALSE,
  useCovariateInteractionYear = FALSE, useCovariateInteractionMonth = FALSE,
  excludedCovariateConceptIds = c(4027133, 4032243, 4146536, 2002282, 2213572,
  2005890, 43534760, 21601019), deleteCovariatesSmallCount = 100)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{cdmSchema}{}

\item{resultsSchema}{}

\item{targetDrugConceptId}{}

\item{comparatorDrugConceptId}{}

\item{indicationConceptIds}{}

\item{washoutWindow}{}

\item{indicationLookbackWindow}{}

\item{studyStartDate}{}

\item{studyEndDate}{}

\item{exclusionConceptIds}{}

\item{outcomeConceptIds}{}

\item{outcomeConditionTypeConceptIds}{}

\item{exposureSchema}{}

\item{exposureTable}{}

\item{outcomeSchema}{}

\item{outcomeTable}{}

\item{useCovariateDemographics}{}

\item{useCovariateConditionOccurrence}{}

\item{useCovariateConditionOccurrence365d}{}

\item{useCovariateConditionOccurrence30d}{}

\item{useCovariateConditionOccurrenceInpt180d}{}

\item{useCovariateConditionEra}{}

\item{useCovariateConditionEraEver}{}

\item{useCovariateConditionEraOverlap}{}

\item{useCovariateConditionGroup}{}

\item{useCovariateDrugExposure}{}

\item{useCovariateDrugExposure365d}{}

\item{useCovariateDrugExposure30d}{}

\item{useCovariateDrugEra}{}

\item{useCovariateDrugEra365d}{}

\item{useCovariateDrugEra30d}{}

\item{useCovariateDrugEraOverlap}{}

\item{useCovariateDrugEraEver}{}

\item{useCovariateDrugGroup}{}

\item{useCovariateProcedureOccurrence}{}

\item{useCovariateProcedureOccurrence365d}{}

\item{useCovariateProcedureOccurrence30d}{}

\item{useCovariateProcedureGroup}{}

\item{useCovariateObservation}{}

\item{useCovariateObservation365d}{}

\item{useCovariateObservation30d}{}

\item{useCovariateObservationBelow}{}

\item{useCovariateObservationAbove}{}

\item{useCovariateObservationCount365d}{}

\item{useCovariateConceptCounts}{}

\item{useCovariateRiskScores}{}

\item{useCovariateInteractionYear}{}

\item{useCovariateInteractionMonth}{}

\item{excludedCovariateConceptIds}{}

\item{deleteCovariatesSmallCount}{}

\item{sourceName}{}
}
\value{
Returns an object of type \code{cohortData}, containing information on the cohorts, their outcomes,
and baseline covariates.
}
\description{
Todo: add description
}
\details{
Todo: add details
}

