% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDbCohortDataObject}
\alias{getDbCohortDataObject}
\title{Get the cohort data from the server}
\usage{
getDbCohortDataObject(connectionDetails, cdmSchema = "CDM4_SIM",
  resultsSchema = "scratch", targetDrugConceptId = 755695,
  comparatorDrugConceptId = 739138, indicationConceptIds = 439926,
  washoutWindow = 183, indicationLookbackWindow = 183,
  studyStartDate = "", studyEndDate = "", exclusionConceptIds = c(4027133,
  4032243, 4146536, 2002282, 2213572, 2005890, 43534760, 21601019),
  outcomeConceptIds = 194133, outcomeConditionTypeConceptIds = c(38000215,
  38000216, 38000217, 38000218, 38000183, 38000232), maxOutcomeCount = 1,
  exposureTable = "DRUG_ERA", outcomeTable = "CONDITION_OCCURRENCE",
  useCovariateDemographics = TRUE, useCovariateConditionOccurrence = TRUE,
  useCovariateConditionEra = FALSE, useCovariateConditionGroup = FALSE,
  useCovariateDrugExposure = FALSE, useCovariateDrugEra = FALSE,
  useCovariateDrugGroup = FALSE, useCovariateProcedureOccurrence = FALSE,
  useCovariateProcedureGroup = FALSE, useCovariateObservation = FALSE,
  useCovariateConceptCounts = FALSE, useCovariateRiskScores = FALSE,
  useCovariateInteractionYear = FALSE, useCovariateInteractionMonth = FALSE,
  excludedCovariateConceptIds = c(4027133, 4032243, 4146536, 2002282, 2213572,
  2005890, 43534760, 21601019), deleteCovariatesSmallCount = 100)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{cdmSchema}{}

\item{resultsSchema}{}

\item{targetDrugConceptId}{}

\item{comparatorDrugConceptId}{}

\item{indicationConceptIds}{}

\item{washoutWindow}{}

\item{indicationLookbackWindow}{}

\item{studyStartDate}{}

\item{studyEndDate}{}

\item{exclusionConceptIds}{}

\item{outcomeConceptIds}{}

\item{outcomeConditionTypeConceptIds}{}

\item{maxOutcomeCount}{}

\item{exposureTable}{}

\item{outcomeTable}{}

\item{useCovariateDemographics}{}

\item{useCovariateConditionOccurrence}{}

\item{useCovariateConditionEra}{}

\item{useCovariateConditionGroup}{}

\item{useCovariateDrugExposure}{}

\item{useCovariateDrugEra}{}

\item{useCovariateDrugGroup}{}

\item{useCovariateProcedureOccurrence}{}

\item{useCovariateProcedureGroup}{}

\item{useCovariateObservation}{}

\item{useCovariateConceptCounts}{}

\item{useCovariateRiskScores}{}

\item{useCovariateInteractionYear}{}

\item{useCovariateInteractionMonth}{}

\item{excludedCovariateConceptIds}{}

\item{deleteCovariatesSmallCount}{}
}
\value{
Returns an object of type \code{cohortData}, containing information on the cohorts, their outcomes,
and baseline covariates.
}
\description{
Todo: add description
}
\details{
Todo: add details
}

