% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataLoadingSaving.R
\name{getDbCohortMethodData}
\alias{getDbCohortMethodData}
\title{Get the cohort data from the server}
\usage{
getDbCohortMethodData(connectionDetails, cdmDatabaseSchema,
  oracleTempSchema = cdmDatabaseSchema, targetDrugConceptId,
  comparatorDrugConceptId, indicationConceptIds = c(), washoutWindow = 183,
  indicationLookbackWindow = 183, studyStartDate = "", studyEndDate = "",
  exclusionConceptIds = c(), outcomeConceptIds,
  outcomeConditionTypeConceptIds = c(),
  exposureDatabaseSchema = cdmDatabaseSchema, exposureTable = "drug_era",
  outcomeDatabaseSchema = cdmDatabaseSchema,
  outcomeTable = "condition_occurrence", cdmVersion = "4",
  excludeDrugsFromCovariates = TRUE, covariateSettings)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
                                    function \code{createConnectionDetails} in the
                                    \code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance.  Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{oracleTempSchema}{For Oracle only: the name of the database schema where you want
all temporary tables to be managed. Requires create/insert
permissions to this database.}

\item{targetDrugConceptId}{A unique identifier to define the target cohort.  If
exposureTable = DRUG_ERA, targetDrugConceptId is a CONCEPT_ID
and all descendant concepts within that CONCEPT_ID will be used
to define the cohort.  If exposureTable <> DRUG_ERA,
targetDrugConceptId is used to select the cohort_concept_id in
the cohort-like table.}

\item{comparatorDrugConceptId}{A unique identifier to define the comparator cohort.  If
exposureTable = DRUG_ERA, comparatorDrugConceptId is a
CONCEPT_ID and all descendant concepts within that CONCEPT_ID
will be used to define the cohort.  If exposureTable <>
DRUG_ERA, comparatorDrugConceptId is used to select the
cohort_concept_id in the cohort-like table.}

\item{indicationConceptIds}{A list of CONCEPT_IDs used to restrict the target and comparator
cohorts, based on any descendant condition of this list
occurring at least once within the indicationLookbackWindow
prior to the cohort index date. If no concept IDs are specified,
the cohorts are not restricted to any indication.}

\item{washoutWindow}{The mininum required continuous observation time prior to index
date for a person to be included in the cohort.}

\item{indicationLookbackWindow}{The window to look back prior to cohort index date to identify
records of a indication condition.  Only applicable if
indicationConceptIds != ''.}

\item{studyStartDate}{A calendar date specifying the minimum date that a cohort index
date can appear. Date format is 'yyyymmdd'.}

\item{studyEndDate}{A calendar date specifying the maximum date that a cohort index
date can appear. Date format is 'yyyymmdd'.}

\item{exclusionConceptIds}{A list of CONCEPT_IDs used to restrict the cohorts, based on any
descendant conditions/drugs/procedures occurring at least once
anytime prior to the cohort index date.}

\item{outcomeConceptIds}{A list of CONCEPT_IDs used to define outcomes.  If
outcomeTable = CONDITION_OCCURRENCE, the list is a set of
ancestor CONCEPT_IDs, and all occurrences of all descendant
concepts will be selected.  If outcomeTable <>
CONDITION_OCCURRENCE, the list contains records found in
COHORT_DEFINITION_ID field.}

\item{outcomeConditionTypeConceptIds}{A list of TYPE_CONCEPT_ID values that will restrict
condition occurrences.  Only applicable if outcomeTable =
CONDITION_OCCURRENCE.}

\item{exposureDatabaseSchema}{The name of the database schema that is the location where the
exposure data used to define the exposure cohorts is available.
If exposureTable = DRUG_ERA, exposureDatabaseSchema is not used
by assumed to be cdmSchema.  Requires read permissions to this
database.}

\item{exposureTable}{The tablename that contains the exposure cohorts.  If
exposureTable <> DRUG_ERA, then expectation is exposureTable has
format of COHORT table: cohort_concept_id, SUBJECT_ID,
COHORT_START_DATE, COHORT_END_DATE.}

\item{outcomeDatabaseSchema}{The name of the database schema that is the location where
the data used to define the outcome cohorts is available. If
exposureTable = CONDITION_ERA, exposureDatabaseSchema is not
used by assumed to be cdmSchema.  Requires read permissions
to this database.}

\item{outcomeTable}{The tablename that contains the outcome cohorts.  If
outcomeTable <> CONDITION_OCCURRENCE, then expectation is
outcomeTable has format of COHORT table:
COHORT_DEFINITION_ID, SUBJECT_ID, COHORT_START_DATE,
COHORT_END_DATE.}

\item{cdmVersion}{Define the OMOP CDM version used: currently support "4" and "5".}

\item{excludeDrugsFromCovariates}{Should the target and comparator drugs (and their descendant
concepts) be excluded from the covariates? Note that this will
work if the drugs are actualy drug concept IDs (and not cohort
IDs).}

\item{covariateSettings}{An object of type \code{covariateSettings} as created using the
\code{createCovariateSettings} function in the
\code{PatientLevelPrediction} package..}
}
\value{
Returns an object of type \code{cohortMethodData}, containing information on the cohorts, their
outcomes, and baseline covariates. Information about multiple outcomes can be captured at once for
efficiency reasons. This object is a list with the following components: \describe{
\item{outcomes}{An ffdf object listing the outcomes per person, including the time to event, and
the outcome conncept ID. Outcomes are not yet filtered based on risk window, since this is done at
a later stage.} \item{cohorts}{An ffdf object listing the persons in each cohort, listing their
exposure status as well as the time to the end of the observation period and time to the end of the
cohort (usually the end of the exposure era).} \item{covariates}{An ffdf object listing the
baseline covariates per person in the two cohorts. This is done using a sparse representation:
covariates with a value of 0 are omitted to save space.} \item{exclude}{An ffdf object listing for
each outcome concept ID the persons that need to be excluded from the analysis because of prior
outcomes.} \item{covariateRef}{An ffdf object describing the covariates that have been extracted.}
\item{metaData}{A list of objects with information on how the cohortMethodData object was
constructed.} } The generic \code{summary()} function has been implemented for this object.
}
\description{
This function executes a large set of SQL statements against the database in OMOP CDM format to
extract the data needed to perform the analysis.
}
\details{
Based on the parameters, the treatment and comparator cohorts are constructed. Baseline covariates
at or before the index date are extracted, as well as outcomes occurring on or after the index
date. The treatment and comparator cohorts can be identified using the drug_era table, or through
user-defined cohorts in a cohort table either inside the CDM instance or in a separate schema.
Similarly, outcomes are identified using the condition_occurrence or condition_era table, or
through user-defined cohorts in a cohort table either inside the CDM instance or in a separate
schema. Covariates are automatically extracted from the appropriate tables within the CDM.This
function calls the \code{getDbCovariates} and \code{getDbOutcomes} functions.

Important: The target and comparator drug must not be included in the covariates, including any
descendant concepts. If the \code{targetDrugConceptId} and \code{comparatorDrugConceptId} arguments represent
real concept IDs, you can set the \code{excludeDrugsFromCovariates} argument to TRUE and automatically the drugs
and their descendants will be excluded from the covariates. However, if the \code{targetDrugConceptId} and
\code{comparatorDrugConceptId} arguments do not represent concept IDs, you will need to manually add the drugs
and descendants to the \code{excludedCovariateConceptIds} of the \code{covariateSettings} argument.
}

