% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1.R
\name{getDefaultCmTable1Specifications}
\alias{getDefaultCmTable1Specifications}
\title{Get the default table 1 specifications}
\usage{
getDefaultCmTable1Specifications()
}
\value{
A specifications objects.
}
\description{
Loads the default specifications for a table 1, to be used with the \code{\link{createTable1}}
function.

Important: currently only works for binary covariates.
}
