% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{getDemographicsCovariateRef}
\alias{getDemographicsCovariateRef}
\title{Retrieves demographics elements of covariate Ref}
\usage{
getDemographicsCovariateRef(cohortData, analysisIds)
}
\arguments{
\item{cohortData}{\code{cohortData} object constructed by \code{getDbCohortData}}

\item{analysisIds}{Vector of analysis identifiers associated with demographic covariates.}
}
\value{
Returns ffdf of elements of \code{cohortData$covariateRef} that relate to demographic information.
}
\description{
Retrieves demographics elements of covariate Ref
}

