% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{getDemographicsCovariates}
\alias{getDemographicsCovariates}
\title{Retrieves demographics elements of covariates}
\usage{
getDemographicsCovariates(cohortData, covariateRef)
}
\arguments{
\item{cohortData}{\code{cohortData} object constructed by \code{getDbCohortData}}

\item{covariateRef}{Covariate reference information associated with demographic covariates.}
}
\value{
Returns ffdf of elements of \code{cohortData$covariates} that relate to demographic information.
}
\description{
Retrieves demographics elements of covariates
}

