% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalyses.R
\name{getDiagnosticsSummary}
\alias{getDiagnosticsSummary}
\title{Get a summary report of the analyses diagnostics}
\usage{
getDiagnosticsSummary(outputFolder)
}
\arguments{
\item{outputFolder}{Name of the folder where all the outputs have been written to.}
}
\value{
A tibble containing summary diagnostics for each outcome-covariate-analysis combination.
}
\description{
Get a summary report of the analyses diagnostics
}
