% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{getDimensionAnalysisId}
\alias{getDimensionAnalysisId}
\title{Retrieves analysisId}
\usage{
getDimensionAnalysisId(dimensionInfo)
}
\arguments{
\item{dimensionInfo}{Information for a data dimension, as acquired from \code{getDimensionTable}}
}
\value{
Returns the numeric value for \code{analysisId}
}
\description{
Retrieves \code{analysisId} from data dimension information
}

