% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{getDimensionSql}
\alias{getDimensionSql}
\title{Retrieves sql query}
\usage{
getDimensionSql(dimensionInfo)
}
\arguments{
\item{dimensionInfo}{Information for a data dimension, as acquired from \code{getDimensionTable}}
}
\value{
Returns the sql query string
}
\description{
Retrieves \code{sql} from data dimension information
}

