% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{getDimensionTable}
\alias{getDimensionTable}
\title{Loads up data dimensions.}
\usage{
getDimensionTable(dimensions)
}
\arguments{
\item{dimensions}{A character vector of data dimension names}
}
\value{
Returns an list of \code{data.frame}, one for each data dimension. Each \code{data.frame} has the following columns: \describe{
\item{name}{Name of data dimension}
\item{analysisId}{Analysis identifier}
\item{sql}{Sql query (as string) that will retrieve mapping from OMOP concept ID to source code for specific data dimension}
\item{truncate}{String that can be parsed and evaluated into a function that truncates source code to desired granularity}
\item{truncatedNames}{String that can be parsed and evluated to return a \code{data.frame} with columns \code{code} and \code{name}.
This is for cases where the source code is truncated and new names for truncated codes are available} }
}
\description{
This function loads up information related to each data dimension provided. Data dimensions are hard coded here
}
\details{
If a code is not truncated, it's covariate name remains unchanged. If it is truncated and no name table is available, truncatedNames = ""
and the covariate name will become "covariate name not available."
}

