% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrsFunctions.R
\name{getDrsModel}
\alias{getDrsModel}
\title{Get the disease risk model}
\usage{
getDrsModel(diseaseRiskScore, cohortMethodData)
}
\arguments{
\item{diseaseRiskScore}{The disease risk scores as generated using the \code{\link{createDrs}} function.}

\item{cohortMethodData}{An object of type \code{cohortMethodData} as generated using
\code{\link{getDbCohortMethodData}}.}
}
\description{
\code{getDrsModel} shows the disease risk model
}
\details{
Shows the coefficients and names of the covariates with non-zero coefficients.
}
