% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalyses.R
\name{getFileReference}
\alias{getFileReference}
\title{Get file reference}
\usage{
getFileReference(outputFolder)
}
\arguments{
\item{outputFolder}{Name of the folder where all the outputs have been written to.}
}
\value{
A tibble containing file names of artifacts generated for each target-comparator-outcome-analysis combination.
}
\description{
Get file reference
}
