% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFullOutcomeModel}
\alias{getFullOutcomeModel}
\title{Get the full outcome model}
\usage{
getFullOutcomeModel(outcomeModel, cohortData)
}
\arguments{
\item{outcomeModel}{An object of type \code{outcomeModel} as generated using he \code{createOutcomeMode} function.}

\item{cohortData}{An object of type \code{cohortData} as generated using \code{dbGetCohortData}.}
}
\description{
\code{getFullOutcomeModel} shows the full outcome model, so showing the betas of all variables
included in the outcome model, not just the treatment variable.
}
\details{
Shows the coefficients and names of the covariates with non-zero coefficients.
}
\examples{
#todo
}

