% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{getICD9Dx}
\alias{getICD9Dx}
\title{Loads ICD9 truncated names table}
\usage{
getICD9Dx(file)
}
\arguments{
\item{file}{Name of file}
}
\value{
Returns \code{data.frame} with columns: \describe{
\item{code}{Truncated ICD9 codes to 3 digits}
\item{name}{Name of ICD9 code} }
}
\description{
This function loads the saved table that contains the names of ICD9 codes truncated to 3 digits from the file ICD9Dx.Rdata
}

