% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{getNewCovariateRef}
\alias{getNewCovariateRef}
\title{Get covariateRef entries for covariates}
\usage{
getNewCovariateRef(covariates, cohortData)
}
\arguments{
\item{covariates}{Ffdf of covariates}

\item{cohortData}{CohortMethodData object}
}
\value{
Returns ffdf of covariateRef entries corresponding to covariates
}
\description{
Get covariateRef entries for covariates
}

