% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{getPredefinedCovariateRef}
\alias{getPredefinedCovariateRef}
\title{Retrieves predefined elements of covariate Ref}
\usage{
getPredefinedCovariateRef(cohortData, conceptIds)
}
\arguments{
\item{cohortData}{\code{cohortData} object constructed by \code{getDbCohortData}}

\item{conceptIds}{Vector of concept identifiers associated with predfined covariates.}
}
\value{
Returns ffdf of elements of \code{cohortData$covariateRef} that relate to predefined covariates.
}
\description{
Retrieves predefined elements of covariate Ref
}

