% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{getPredefinedCovariates}
\alias{getPredefinedCovariates}
\title{Retrieves prefefined elements of covariates}
\usage{
getPredefinedCovariates(cohortData, predefinedCovariateRef)
}
\arguments{
\item{cohortData}{\code{cohortData} object constructed by \code{getDbCohortData}}

\item{predefinedCovariateRef}{Ffdf of covariate reference entries associated with predefined covariates.}
}
\value{
Returns ffdf of elements of \code{cohortData$covariates} that relate to predefined covariates.
}
\description{
Retrieves prefefined elements of covariates
}

