% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsFunctions.R
\name{getPsModel}
\alias{getPsModel}
\title{Get the propensity model}
\usage{
getPsModel(propensityScore, cohortMethodData)
}
\arguments{
\item{propensityScore}{The propensity scores as generated using the \code{\link[=createPs]{createPs()}} function.}

\item{cohortMethodData}{An object of type \link{CohortMethodData} as generated using
\code{\link[=getDbCohortMethodData]{getDbCohortMethodData()}}.}
}
\value{
A tibble.
}
\description{
Returns the coefficients and names of the covariates with non-zero coefficients.
}
