% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export.R
\name{getResultsDataModel}
\alias{getResultsDataModel}
\title{Get results data model}
\usage{
getResultsDataModel()
}
\value{
A tibble specifying the tables and columns generated by the \code{\link[=exportToCsv]{exportToCsv()}} function.
}
\description{
Get results data model
}
