% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{getResultsDataModelSpecifications}
\alias{getResultsDataModelSpecifications}
\title{Get specifications for CohortMethod results data model}
\usage{
getResultsDataModelSpecifications()
}
\value{
A tibble data frame object with specifications
}
\description{
Get specifications for CohortMethod results data model
}
