% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viewer.R
\name{insertExportedResultsInSqlite}
\alias{insertExportedResultsInSqlite}
\title{Insert exported results into a SQLite database}
\usage{
insertExportedResultsInSqlite(sqliteFileName, exportFolder, cohorts)
}
\arguments{
\item{sqliteFileName}{The name of the SQLite file to store the results in. If the
file does not exist it will be created.}

\item{exportFolder}{The folder containing the CSV files to upload, as generated
using the \code{\link[=exportToCsv]{exportToCsv()}} function.}

\item{cohorts}{A data frame describing the cohorts used in the study. Should
include the target, comparator, and outcome of interest cohorts.
The data frame should at least have a \code{cohortId} and \code{cohortName}
columns.}
}
\value{
Does not return anything. Called for the side effect of inserting data into the
SQLite database.
}
\description{
Insert exported results into a SQLite database
}
