% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvalidationPolicy.R
\name{InvalidationPolicy}
\alias{InvalidationPolicy}
\title{Invalidation Policy}
\description{
Determines which artifacts should be deleted when analysis settings change,
based on dependency relationships between components.
}
\details{
When settings change, not all artifacts need to be regenerated. This class
implements the logic to determine minimal invalidation scope based on which
settings components have changed.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InvalidationPolicy-computeInvalidationScope}{\code{InvalidationPolicy$computeInvalidationScope()}}
\item \href{#method-InvalidationPolicy-getInvalidationMessage}{\code{InvalidationPolicy$getInvalidationMessage()}}
\item \href{#method-InvalidationPolicy-clone}{\code{InvalidationPolicy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidationPolicy-computeInvalidationScope"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidationPolicy-computeInvalidationScope}{}}}
\subsection{Method \code{computeInvalidationScope()}}{
Compute which file patterns should be deleted based on changed components
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidationPolicy$computeInvalidationScope(changedComponents)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{changedComponents}}{A list (as returned by SettingsHasher$compareSettingsComponents)
with logical values for each component}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Deletion cascade logic:
\itemize{
\item If loadArgsChanged: all downstream artifacts (everything)
\item Else if studyPopArgsChanged: populations, PS models, strata, outcomes
\item Else if psArgsChanged: PS models, strata, outcomes
\item Else if strataArgsChanged: strata and outcomes
\item Else if outcomeModelArgsChanged: outcome models only
\item Else if only balanceArgsChanged: balance files only
\item Else if analyticsChanged (only new outcomes): nothing (outcomes are outcome-specific)
}
}

\subsection{Returns}{
Character vector of file patterns/globs to delete. Empty vector if no deletion needed.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidationPolicy-getInvalidationMessage"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidationPolicy-getInvalidationMessage}{}}}
\subsection{Method \code{getInvalidationMessage()}}{
Get a human-readable message describing what will be deleted
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidationPolicy$getInvalidationMessage(changedComponents)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{changedComponents}}{A list of changed components}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character string suitable for displaying to the user
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidationPolicy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidationPolicy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidationPolicy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
