% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMethodData.R
\name{isCohortMethodData}
\alias{isCohortMethodData}
\title{Check whether an object is a CohortMethodData object}
\usage{
isCohortMethodData(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
A logical value.
}
\description{
Check whether an object is a CohortMethodData object
}
