% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isSorted}
\alias{isSorted}
\title{Check if data is sorted by one or more columns}
\usage{
isSorted(data, columnNames, ascending = rep(TRUE, length(columnNames)))
}
\arguments{
\item{data}{Either a data.frame of ffdf object.}

\item{columnNames}{Vector of one or more column names.}

\item{ascending}{Logical vector indicating the data should be sorted ascending or descending
according the specified columns.}
}
\value{
True or false
}
\description{
\code{isSorted} checks wether data is sorted by one or more specified columns.
}
\details{
This function currently only supports checking for sorting on numeric values.
}
\examples{
x <- data.frame(a = runif(1000),b = runif(1000))
x <- round(x,digits=2)
isSorted(x,c("a","b"))

x <- x[order(x$a,x$b),]
isSorted(x,c("a","b"))

x <- x[order(x$a,-x$b),]
isSorted(x,c("a","b"),c(TRUE,FALSE))
}

