% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viewer.R
\name{launchResultsViewer}
\alias{launchResultsViewer}
\title{Launch Shiny app using}
\usage{
launchResultsViewer(connectionDetails, databaseSchema)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the
\code{\link[DatabaseConnector:createConnectionDetails]{DatabaseConnector::createConnectionDetails()}} function.}

\item{databaseSchema}{The name of the database schema where the results were
written using \link{uploadExportedResults(}.}
}
\value{
Does not return anything. Is called for the side-effect of launching the Shiny
app.
}
\description{
Launch Shiny app using
}
