% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viewer.R
\name{launchResultsViewerUsingSqlite}
\alias{launchResultsViewerUsingSqlite}
\title{Launch Shiny app using a SQLite database}
\usage{
launchResultsViewerUsingSqlite(sqliteFileName)
}
\arguments{
\item{sqliteFileName}{The name of the SQLite file where the results were stored
using the \code{\link[=insertExportedResultsInSqlite]{insertExportedResultsInSqlite()}} function.}
}
\value{
Does not return anything. Is called for the side-effect of launching the Shiny
app.
}
\description{
Launch Shiny app using a SQLite database
}
