% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettingsObjects.R
\name{loadCmAnalysisList}
\alias{loadCmAnalysisList}
\title{Load a list of CmAnalysis from file}
\usage{
loadCmAnalysisList(file)
}
\arguments{
\item{file}{The name of the file}
}
\value{
A list of objects of type \code{CmAnalysis}.
}
\description{
Load a list of objects of type \code{CmAnalysis} from file. The file is in JSON format.
}
