% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadCohortDataObject}
\alias{loadCohortDataObject}
\title{Load the cohort data from a folder}
\usage{
loadCohortDataObject(file)
}
\arguments{
\item{file}{The name of the folder containing the data.}
}
\value{
An object of class cohortData.
}
\description{
\code{loadCohortDataObject} loads an object of type cohortData from a folder in the file system.
}
\details{
The data will be written to a set of files in the folder specified by the user.
}
\examples{
#todo
}

