% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Analyses.R
\name{loadCohortMethodAnalysisList}
\alias{loadCohortMethodAnalysisList}
\title{Load a list of cohortMethodAnalysis from file}
\usage{
loadCohortMethodAnalysisList(file)
}
\arguments{
\item{file}{The name of the file}
}
\value{
A list of objects of type \code{cohortMethodAnalysis}.
}
\description{
Load a list of objects of type \code{cohortMethodAnalysis} from file. The file is in JSON format.
}

