% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMethodData.R
\name{loadCohortMethodData}
\alias{loadCohortMethodData}
\title{Load the cohort method data from a file}
\usage{
loadCohortMethodData(file)
}
\arguments{
\item{file}{The name of the file containing the data.}
}
\value{
An object of class \link{CohortMethodData}.
}
\description{
Loads an object of type \link{CohortMethodData} from a file in the file system.
}
