% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Analyses.R
\name{loadDrugComparatorOutcomesList}
\alias{loadDrugComparatorOutcomesList}
\title{Load a list of drugComparatorOutcomes from file}
\usage{
loadDrugComparatorOutcomesList(file)
}
\arguments{
\item{file}{The name of the file}
}
\value{
A list of objects of type \code{drugComparatorOutcome}.
}
\description{
Load a list of objects of type \code{drugComparatorOutcomes} from file. The file is in JSON format.
}

