% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{loadFfdfList}
\alias{loadFfdfList}
\title{Loads list of ffdf objects}
\usage{
loadFfdfList(file)
}
\arguments{
\item{file}{Name of folder}
}
\description{
Loads list of ffdf objects saved by \code{saveFfdfList}
}

