% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettingsObjects.R
\name{loadTargetComparatorOutcomesList}
\alias{loadTargetComparatorOutcomesList}
\title{Load a list of \code{TargetComparatorOutcomes} from file}
\usage{
loadTargetComparatorOutcomesList(file)
}
\arguments{
\item{file}{The name of the file}
}
\value{
A list of objects of type \code{TargetComparatorOutcomes}.
}
\description{
Load a list of objects of type \code{TargetComparatorOutcomes} from file. The file is in JSON format.
}
