% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotCovariateBalanceScatterPlot}
\alias{plotCovariateBalanceScatterPlot}
\title{Create a scatterplot of the covariate balance}
\usage{
plotCovariateBalanceScatterPlot(balance, fileName = NULL)
}
\arguments{
\item{balance}{A data frame created by the \code{computeCovariateBalance} funcion.}

\item{fileName}{Name of the file where the plot should be saved, for example 'plot.png'. See the
function \code{ggsave} in the ggplot2 package for supported file formats.}
}
\description{
Create a scatterplot of the covariate balance, showing all variables with balance before and after
matching on the x and y axis respectively. Requires running \code{computeCovariateBalance} first.
}

