% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrsFunctions.R
\name{plotDrs}
\alias{plotDrs}
\title{Plot the disease risk score distribution}
\usage{
plotDrs(
  data,
  targetLabel = "Target",
  comparatorLabel = "Comparator",
  title = NULL,
  fileName = NULL
)
}
\arguments{
\item{data}{A data frame with at least the two columns described below}

\item{targetLabel}{A label to us for the target cohort.}

\item{comparatorLabel}{A label to us for the comparator cohort.}

\item{title}{Optional: the main title for the plot.}

\item{fileName}{Name of the file where the plot should be saved, for example 'plot.png'.
See the function \code{ggsave} in the ggplot2 package for supported file
formats.}
}
\value{
A ggplot object. Use the \code{\link[ggplot2]{ggsave}} function to save to file in a different
format.
}
\description{
\code{plotDrs} shows the disease risk score distribution
}
\details{
The data frame should have a least the following two columns: \tabular{lll}{ \verb{treatment}
\tab(integer) \tab Column indicating whether the person is in the target (1) or comparator\cr \tab
\tab (0) group \cr \verb{diseaseRiskScore} \tab(numeric) \tab disease risk score \cr }
}
\examples{
treatment <- rep(0:1, each = 100)
diseaseRiskScore <- c(rnorm(100, mean = 0.4, sd = 0.25), rnorm(100, mean = 0.6, sd = 0.25))
data <- data.frame(treatment = treatment, diseaseRiskScore = diseaseRiskScore)
data <- data[data$diseaseRiskScore > 0 & data$diseaseRiskScore < 1, ]
plotDrs(data)

}
