% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{plotFollowUpDistribution}
\alias{plotFollowUpDistribution}
\title{Plot the distribution of follow-up time}
\usage{
plotFollowUpDistribution(population, targetLabel = "Target",
  comparatorLabel = "Comparator", yScale = "percent",
  logYScale = FALSE, dataCutoff = 0.95, title = NULL,
  fileName = NULL)
}
\arguments{
\item{population}{A data frame describing the study population as created using the
\code{\link{createStudyPopulation}} function. This should at least have
these columns: treatment, timeAtRisk.}

\item{targetLabel}{A label to us for the target cohort.}

\item{comparatorLabel}{A label to us for the comparator cohort.}

\item{yScale}{Should be either 'percent' or 'count'.}

\item{logYScale}{Should the Y axis be on the log scale?}

\item{dataCutoff}{Fraction of the data (number censored) after which the graph will not be
shown.}

\item{title}{The main title of the plot.}

\item{fileName}{Name of the file where the plot should be saved, for example 'plot.png'.
See the function \code{ggsave} in the ggplot2 package for supported file
formats.}
}
\value{
A ggplot object. Use the \code{\link[ggplot2]{ggsave}} function to save to file in a different
format.
}
\description{
Plot the distribution of follow-up time
}
\details{
Plot the distribution of follow-up time, stratified by treatment group.Follow-up time is defined as
time-at-risk, so not censored at the outcome.
}
