% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{processICD9Dx}
\alias{processICD9Dx}
\title{Function used to load and process ICD9 codes}
\usage{
processICD9Dx(file)
}
\arguments{
\item{file}{File name}
}
\value{
Returns \code{data.frame} with columns: \describe{
\item{code}{Truncated ICD9 codes to 3 digits}
\item{name}{Name of ICD9 code} }
}
\description{
This function was used in processing ICD9Dx.txt
}

