% Generated by roxygen2 (4.0.1): do not edit by hand
\name{psAuc}
\alias{psAuc}
\title{Compute the area under the ROC curve}
\usage{
psAuc(data)
}
\arguments{
\item{data}{A data frame with at least the two columns described below}
}
\value{
A data frame holding the AUC and its 95% confidence interval
}
\description{
\code{psAuc} shows the area under the ROC curve os the propensity score
}
\details{
The data frame should have a least the following two columns:
\tabular{lll}{
  \verb{treatment}          \tab(integer) \tab Column indicating whether the person is in the treated (1) or comparator (0) group  \cr
  \verb{propensityScore}   \tab(real)    \tab Propensity score \cr
}
}
\examples{
treatment = rep(0:1, each = 100)
propensityScore = c(rnorm(100,mean=0.4, sd=0.25),rnorm(100,mean=0.6, sd=0.25))
data <- data.frame(treatment = treatment, propensityScore = propensityScore)
data <- data[data$propensityScore > 0 & data$propensityScore < 1,]
psAuc(data)
}

