% Generated by roxygen2 (4.0.1): do not edit by hand
\name{psPlot}
\alias{psPlot}
\title{Plot the propensity score distribution}
\usage{
psPlot(data, scale = "preference")
}
\arguments{
\item{data}{A data frame with at least the two columns described below}

\item{scale}{The scale of the graph. Two scales are supported: \code{
scale = "propensity"} or  \code{#' scale = "preference"}. The preference score scale is defined by Walker
et al (2013).}
}
\description{
\code{psPlot} shows the propensity (or preference) score distribution
}
\details{
The data frame should have a least the following two columns:
\tabular{lll}{
  \verb{treatment}          \tab(integer) \tab Column indicating whether the person is in the treated (1) or comparator (0) group  \cr
  \verb{propensityScore}   \tab(real)    \tab Propensity score \cr
}
}
\examples{
treatment = rep(0:1, each = 100)
propensityScore = c(rnorm(100,mean=0.4, sd=0.25),rnorm(100,mean=0.6, sd=0.25))
data <- data.frame(treatment = treatment, propensityScore = propensityScore)
data <- data[data$propensityScore > 0 & data$propensityScore < 1,]
psPlot(data)
}
\references{
Walker AM, Patrick AR, Lauer MS, Hornbrook MC, Marin MG, Platt R, Roger VL, Stang P, and Schneeweiss S.
(2013) A tool for assessing the feasibility of comparative effectiveness research, Comparative Effective
Research, 3, 11-20
}

