% Generated by roxygen2 (4.0.1): do not edit by hand
\name{psPlotCovariateBalance}
\alias{psPlotCovariateBalance}
\title{Create a plot of the covariate balance}
\usage{
psPlotCovariateBalance(balance, topSumTreatedBeforeMatching = 0,
  type = "xy")
}
\arguments{
\item{balance}{A data frame created by the \code{psComputeCovariateBalance} funcion.}

\item{topSumTreatedBeforeMatching}{The number of covariates to be shown. Only the covariates with
the highest sum in the treated group before matching are shown.}

\item{type}{The type of plot. Options are \code{type = "rowpervariable"} or \code{type = "xy"}}
}
\value{
A plot showing the covariate imbalance before and after matching
}
\description{
Create a plot of the covariate balance. Requires running \code{psComputeCovariateBalance} first.
}

