% Generated by roxygen2 (4.0.2): do not edit by hand
\name{psPlotCovariateBalanceTopVariables}
\alias{psPlotCovariateBalanceTopVariables}
\title{Plot variables with largest imbalance}
\usage{
psPlotCovariateBalanceTopVariables(balance, n = 20, fileName = NULL)
}
\arguments{
\item{balance}{A data frame created by the \code{psComputeCovariateBalance} funcion.}

\item{fileName}{Name of the file where the plot should be saved, for example 'plot.png'. See the
function \code{ggsave} in the ggplot2 package for supported file formats.}
}
\description{
Create a plot showing those variables having the largest imbalance before matching, and
those variables having the largest imbalance after matching. Requires running
\code{psComputeCovariateBalance} first.
}

