% Generated by roxygen2 (4.0.1): do not edit by hand
\name{psShowModel}
\alias{psShowModel}
\title{Show the propensity model}
\usage{
psShowModel(propensityScore, connectionDetails, cdmSchema)
}
\arguments{
\item{propensityScore}{An object of type \code{cohortData} as generated using \code{dbGetCohortData}}

\item{connectionDetails}{An R object of type ConnectionDetail (details for the function that contains server info, database type, optionally username/password, port)}

\item{cdmSchema}{Database schema that contains the vocabulary}
}
\description{
\code{psShowModel} shows the propensity score model
}
\details{
Shows the coefficients and names of the covariates with non-zero coefficients
}
\examples{
#todo
}

