% Generated by roxygen2 (4.0.1): do not edit by hand
\name{psStratify}
\alias{psStratify}
\title{Stratify persons by propensity score}
\usage{
psStratify(data, numberOfStrata = 5, stratificationColumns = c())
}
\arguments{
\item{data}{A data frame with the three columns described below}

\item{numberOfStrata}{How many strata? The boundaries of the strata are automatically defined to
contain equal numbers of treated persons.}

\item{stratificationColumns}{Names of one or more columns in the \code{data} data.frame on which subjects should also be
stratified in addition to stratification on propensity score.}
}
\value{
Returns a date frame with the same columns as the input data plus one extra column: stratumId.
}
\description{
\code{psStratify} uses the provided propensity scores to stratify persons.
}
\details{
The data frame should have the following three columns:
\tabular{lll}{
  \verb{row_id}              \tab(integer) \tab A unique identifier for each row (e.g. the person ID) \cr
  \verb{treatment}           \tab(integer) \tab Column indicating whether the person is in the treated (1) or comparator (0) group  \cr
  \verb{propensity_score}    \tab(real)    \tab Propensity score \cr
}
}
\examples{
row_id = 1:200
treatment = rep(0:1, each = 100)
propensity_score = c(runif(100,min=0,max=1),runif(100,min=0,max=1))
data <- data.frame(row_id = row_id, treatment = treatment, propensity_score = propensity_score)
result <- psStratify(data,5)
}

