% Generated by roxygen2 (4.0.1): do not edit by hand
\name{psTrim}
\alias{psTrim}
\title{Trim persons by propensity score}
\usage{
psTrim(data, trimFraction = 0.05)
}
\arguments{
\item{data}{A data frame with the three columns described below#'}

\item{method}{Whether \code{matching} or \code{stratification} should be performed.}

\item{trimFraction}{This fraction will be removed from each treatment group. In the treatment group, persons
with the highest propensity scores will be removed, in the comparator group person with the lowest scores will be removed.}
}
\value{
Returns a date frame with the same three columns as the input.
}
\description{
\code{psTrim} uses the provided propensity scores to trim subjects with extreme scores.
}
\details{
The data frame should have the following three columns:
\tabular{lll}{
  \verb{row_id}             \tab(integer) \tab A unique identifier for each row (e.g. the person ID) \cr
  \verb{treatment}  	       \tab(integer) \tab Column indicating whether the person is in the treated (1) or comparator (0) group  \cr
  \verb{propensity_score}   \tab(real)    \tab Propensity score \cr
}
}
\examples{
rowId = 1:2000
treatment = rep(0:1, each = 1000)
propensityScore = c(runif(1000,min=0,max=1),runif(1000,min=0,max=1))
data <- data.frame(rowId = rowId, treatment = treatment, propensityScore = propensityScore)
result <- psTrim(data,0.05)
}

