% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ComparatorUnderSampling.R
\name{recomputePsForFullData}
\alias{recomputePsForFullData}
\title{Recompute the PS for the full data set}
\usage{
recomputePsForFullData(psSample, cohortMethodDataSample, cohortMethodData)
}
\arguments{
\item{psSample}{The propensity scores as created on the sample data.}

\item{cohortMethodDataSample}{The sample data on which the PS model was fitted.}

\item{cohortMethodData}{The full data.}
}
\value{
A new propensity score object.
}
\description{
Recompute the PS for the full data set
}
\details{
After using the \code{\link{sampleCohorts}} or \code{\link{sampleComparator}} to reduce the
population size, this function can be used to apply a propensity model fitted on the sample to the
full data.
}

