% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{removeLowBias}
\alias{removeLowBias}
\title{Keeps only covariates with highest bias.}
\usage{
removeLowBias(data, bias, cutoffIndex)
}
\arguments{
\item{data}{List covariate information for data dimensions, as ffdf objects with columns \code{rowId}, \code{covariateId},
\code{treatment}, \code{outcome}}

\item{bias}{List of bias for each data dimension's covariates, as ffdf objects with columns \code{covariateId}, \code{RR}, \code{bias}}

\item{cutoffIndex}{Number of total covariates to keep}
}
\value{
Returns \code{data} with only the highest bias covariates.
}
\description{
This function eliminates data associated with covariates that have low bias. Only the top \code{cutoffIndex} codes are kept.
}

