% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{removeLowBiasHelper}
\alias{removeLowBiasHelper}
\title{Helper for removeLowBias}
\usage{
removeLowBiasHelper(data, bias, index)
}
\arguments{
\item{data}{Covariate information for each data dimension,}

\item{bias}{Bias information for all covariates}

\item{index}{Index to cut off}
}
\value{
Returns \code{data} with proper codes removed
}
\description{
Helper for removeLowBias
}

