% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{removeLowRank}
\alias{removeLowRank}
\title{Keeps top covariates by bias or exposure rank}
\usage{
removeLowRank(data, rankings, rankCutoff, useExpRank)
}
\arguments{
\item{data}{List of covariates, corresponding to data dimensions}

\item{rankings}{List of rankings generated by \code{calculateRanks}}

\item{rankCutoff}{Number of covariates to keep}

\item{useExpRank}{Boolean to use exposure instead of bias}
}
\value{
Removes input covariates with only top covariates kept
}
\description{
Keeps top number of covariates by bias or exposure rank for inclusion in propensity score
}

