% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{removeLowRankHelper}
\alias{removeLowRankHelper}
\title{Helper for removeLowRank}
\usage{
removeLowRankHelper(data, rankNames)
}
\arguments{
\item{data}{Covariate information for each data dimension,}

\item{rankNames}{Names of covariates to keep}
}
\value{
Returns \code{data} with proper codes removed
}
\description{
Helper for removeLowRank
}

