% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{removePredefinedCovariates}
\alias{removePredefinedCovariates}
\title{Remove predefined covariates}
\usage{
removePredefinedCovariates(cohortData, conceptIds, icd9, icd9AnalysisIds)
}
\arguments{
\item{cohortData}{CohortMethodData object}

\item{conceptIds}{Predefined concept IDs to include and exclude}

\item{icd9}{Predefined icd9 condition codes to include and exclude, with E replaced by 10 and V replaced by 11}

\item{icd9AnalysisIds}{List of analysisIds that include icd9 condition codes}
}
\value{
Returns covariates table of cohortData with predefined codes removed
}
\description{
Removes from covariates table of CohortMethodData the predefined codes to include and exclude. The included covariates have already been
extracted
}

