% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{removeRareCodes}
\alias{removeRareCodes}
\title{Removes codes with insufficient patients}
\usage{
removeRareCodes(data, lowPopCutoff, dimensionCutoff = NULL)
}
\arguments{
\item{data}{Covariate information for data dimension, as ffdf object with columns \code{rowId}, \code{covariateId},
\code{treatment}, \code{outcome}}

\item{lowPopCutoff}{Threshold for number of unique patients that need to have a \code{covariateId} to keep data}

\item{dimensionCutoff}{Maximum number of covariates to keep from this data dimension, ordered by prevalence.}
}
\value{
Returns \code{data} with rare codes removed. Returns a list to work with sapply.
}
\description{
This function removes data for covariates that have insufficient number of unique pataients.
}
\details{
If prevalence of a covariate is > 0.5 among people in data dimension, prevalence is subtracted from 1
}

