% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{removeRareCovariates}
\alias{removeRareCovariates}
\title{Keeps only a certain number of covariates per data dimensions, ranked by prevalence}
\usage{
removeRareCovariates(data, dimensionCutoff, totalPopulation)
}
\arguments{
\item{data}{List of 3 ffdf of covariates for data dimension and expanded codes}

\item{dimensionCutoff}{Number of covariates to keep}

\item{totalPopulation}{Total population in cohort study}
}
\value{
Returns list of 3 ffdf with only the kept covariates
}
\description{
Keeps the top number of covariates, by prevalence, from covariates for data dimension. Codes will only be
kept for (>median) and (>q75) if they are kept in the base category.
}

