% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ComparatorUnderSampling.R
\name{sampleCohorts}
\alias{sampleCohorts}
\title{Sample the target or comparator group down}
\usage{
sampleCohorts(cohortMethodData, treatedSampleSize = 10000,
  comparatorSampleSize = 20000)
}
\arguments{
\item{cohortMethodData}{The original cohortMethodData.}

\item{treatedSampleSize}{The sampe size for the treated cohort.}

\item{comparatorSampleSize}{The sample size for the comparator cohort.}
}
\value{
An object of type \code{cohortMethodData} with the sampled populations.
}
\description{
Sample the target or comparator group down
}
\details{
When the target and/or comparator cohorts are extremely large, it may be more efficient to only use
a sample to fit the propensity model. This function creates a new cohortMethodData object with
sampled populations.
}

