% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparatorUnderSampling.R
\name{sampleComparator}
\alias{sampleComparator}
\title{Sample the comparator group down}
\usage{
sampleComparator(cohortMethodData, comparatorToTreatedRatio = 2)
}
\arguments{
\item{cohortMethodData}{The original cohortMethodData.}

\item{comparatorToTreatedRatio}{The ratio between comparator and treated group.}
}
\value{
An object of type \code{cohortMethodData} with the sampled populations.
}
\description{
Sample the comparator group down
}
\details{
When the comparator group is extremely large, it may be more efficient to only use a sample to fit
the propensity model. This function creates a new cohortMethodData object where to comparator group
is sampled down to a size relative to the treated group.
}

