% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettingsObjects.R
\name{saveCmAnalysisList}
\alias{saveCmAnalysisList}
\title{Save a list of CmAnalysis to file}
\usage{
saveCmAnalysisList(CmAnalysisList, file)
}
\arguments{
\item{CmAnalysisList}{A list of objects of type \code{CmAnalysis} as created using the \code{\link[=createCmAnalysis]{createCmAnalysis()}} function.}

\item{file}{The name of the file where the results will be written}
}
\description{
Write a list of objects of type \code{CmAnalysis} to file. The file is in JSON format.
}
