% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Analyses.R
\name{saveCohortMethodAnalysisList}
\alias{saveCohortMethodAnalysisList}
\title{Save a list of cohortMethodAnalysis to file}
\usage{
saveCohortMethodAnalysisList(cohortMethodAnalysisList, file)
}
\arguments{
\item{cohortMethodAnalysisList}{The cohortMethodAnalysis list to be written to file}

\item{file}{The name of the file where the results will be written}
}
\description{
Write a list of objects of type \code{cohortMethodAnalysis} to file. The file is in JSON format.
}

