% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMethodData.R
\name{saveCohortMethodData}
\alias{saveCohortMethodData}
\title{Save the cohort method data to file}
\usage{
saveCohortMethodData(cohortMethodData, file)
}
\arguments{
\item{cohortMethodData}{An object of type \link{CohortMethodData} as generated using
\code{\link[=getDbCohortMethodData]{getDbCohortMethodData()}}.}

\item{file}{The name of the file where the data will be written. If the file already
exists it will be overwritten.}
}
\value{
Returns no output.
}
\description{
Saves an object of type \link{CohortMethodData} to a file.
}
