% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Analyses.R
\name{saveDrugComparatorOutcomeList}
\alias{saveDrugComparatorOutcomeList}
\title{Save a list of drugComparatorOutcome to file}
\usage{
saveDrugComparatorOutcomeList(drugComparatorOutcomeList, file)
}
\arguments{
\item{drugComparatorOutcomeList}{The drugComparatorOutcome list to be written to file}

\item{file}{The name of the file where the results will be written}
}
\description{
Write a list of objects of type \code{drugComparatorOutcome} to file. The file is in JSON format.
}

