% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{saveDrugComparatorOutcomesList}
\alias{saveDrugComparatorOutcomesList}
\title{Save a list of drugComparatorOutcome to file}
\usage{
saveDrugComparatorOutcomesList(drugComparatorOutcomesList, file)
}
\arguments{
\item{drugComparatorOutcomesList}{The drugComparatorOutcomes list to be written to file}

\item{file}{The name of the file where the results will be written}
}
\description{
Write a list of objects of type \code{drugComparatorOutcomes} to file. The file is in JSON format.
}
