% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{saveFfdfList}
\alias{saveFfdfList}
\title{Saves a list of ffdf objects}
\usage{
saveFfdfList(ffdfList, file)
}
\arguments{
\item{ffdfList}{List of ffdf objects to save}

\item{file}{Name of folder to create / save in}
}
\description{
This function saves a list of ffdf objects in a new folder. The names of the ffdf elements are also preserved.
}

