% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{separateCovariates}
\alias{separateCovariates}
\title{Extracts covariates for data dimension}
\usage{
separateCovariates(analysisId, covariates)
}
\arguments{
\item{analysisId}{Analysis ID for data dimension}

\item{covariates}{Covariate information from cohortMethodData}
}
\value{
List of ffdf containing covariates that correspond to analysisId, analysisId+1 (>median), analysisId+2 (>q75)
}
\description{
Extracts covariates for data dimension into 3 categories - input analysisId, analysisId+1, analysisId+2
for at least once, more than median, and more than 75 quantile codes
}

