% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettingsHasher.R
\name{SettingsHasher}
\alias{SettingsHasher}
\title{Settings Hasher}
\description{
Compute deterministic hashes for individual settings components to detect
which parts of the analysis specification have changed.
}
\details{
This class hashes each settings component independently using JSON serialization
and MD5 digests. This allows fine-grained detection of which artifact types need
to be invalidated when specifications change.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SettingsHasher-hashLoadArgs}{\code{SettingsHasher$hashLoadArgs()}}
\item \href{#method-SettingsHasher-hashStudyPopArgs}{\code{SettingsHasher$hashStudyPopArgs()}}
\item \href{#method-SettingsHasher-hashPsArgs}{\code{SettingsHasher$hashPsArgs()}}
\item \href{#method-SettingsHasher-hashTrimByPsArgs}{\code{SettingsHasher$hashTrimByPsArgs()}}
\item \href{#method-SettingsHasher-hashMatchOnPsArgs}{\code{SettingsHasher$hashMatchOnPsArgs()}}
\item \href{#method-SettingsHasher-hashStratifyByPsArgs}{\code{SettingsHasher$hashStratifyByPsArgs()}}
\item \href{#method-SettingsHasher-hashOutcomeModelArgs}{\code{SettingsHasher$hashOutcomeModelArgs()}}
\item \href{#method-SettingsHasher-hashBalanceArgs}{\code{SettingsHasher$hashBalanceArgs()}}
\item \href{#method-SettingsHasher-compareSettingsComponents}{\code{SettingsHasher$compareSettingsComponents()}}
\item \href{#method-SettingsHasher-clone}{\code{SettingsHasher$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SettingsHasher-hashLoadArgs"></a>}}
\if{latex}{\out{\hypertarget{method-SettingsHasher-hashLoadArgs}{}}}
\subsection{Method \code{hashLoadArgs()}}{
Hash the data loading arguments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SettingsHasher$hashLoadArgs(getDbCohortMethodDataArgs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{getDbCohortMethodDataArgs}}{An object of type \code{GetDbCohortMethodDataArgs}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character string containing the MD5 hash (32 hex characters)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SettingsHasher-hashStudyPopArgs"></a>}}
\if{latex}{\out{\hypertarget{method-SettingsHasher-hashStudyPopArgs}{}}}
\subsection{Method \code{hashStudyPopArgs()}}{
Hash the study population arguments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SettingsHasher$hashStudyPopArgs(createStudyPopulationArgs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{createStudyPopulationArgs}}{An object of type \code{CreateStudyPopulationArgs}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character string containing the MD5 hash (32 hex characters)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SettingsHasher-hashPsArgs"></a>}}
\if{latex}{\out{\hypertarget{method-SettingsHasher-hashPsArgs}{}}}
\subsection{Method \code{hashPsArgs()}}{
Hash the propensity score arguments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SettingsHasher$hashPsArgs(createPsArgs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{createPsArgs}}{An object of type \code{CreatePsArgs}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character string containing the MD5 hash (32 hex characters)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SettingsHasher-hashTrimByPsArgs"></a>}}
\if{latex}{\out{\hypertarget{method-SettingsHasher-hashTrimByPsArgs}{}}}
\subsection{Method \code{hashTrimByPsArgs()}}{
Hash arguments for propensity score trimming
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SettingsHasher$hashTrimByPsArgs(trimByPsArgs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trimByPsArgs}}{An object of type \code{TrimByPsArgs}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character string containing the MD5 hash (32 hex characters)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SettingsHasher-hashMatchOnPsArgs"></a>}}
\if{latex}{\out{\hypertarget{method-SettingsHasher-hashMatchOnPsArgs}{}}}
\subsection{Method \code{hashMatchOnPsArgs()}}{
Hash arguments for propensity score matching
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SettingsHasher$hashMatchOnPsArgs(matchOnPsArgs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{matchOnPsArgs}}{An object of type \code{MatchOnPsArgs}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character string containing the MD5 hash (32 hex characters)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SettingsHasher-hashStratifyByPsArgs"></a>}}
\if{latex}{\out{\hypertarget{method-SettingsHasher-hashStratifyByPsArgs}{}}}
\subsection{Method \code{hashStratifyByPsArgs()}}{
Hash arguments for propensity score stratification
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SettingsHasher$hashStratifyByPsArgs(stratifyByPsArgs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stratifyByPsArgs}}{An object of type \code{StratifyByPsArgs}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character string containing the MD5 hash (32 hex characters)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SettingsHasher-hashOutcomeModelArgs"></a>}}
\if{latex}{\out{\hypertarget{method-SettingsHasher-hashOutcomeModelArgs}{}}}
\subsection{Method \code{hashOutcomeModelArgs()}}{
Hash arguments for outcome model fitting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SettingsHasher$hashOutcomeModelArgs(fitOutcomeModelArgs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fitOutcomeModelArgs}}{An object of type \code{FitOutcomeModelArgs}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character string containing the MD5 hash (32 hex characters)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SettingsHasher-hashBalanceArgs"></a>}}
\if{latex}{\out{\hypertarget{method-SettingsHasher-hashBalanceArgs}{}}}
\subsection{Method \code{hashBalanceArgs()}}{
Hash arguments for covariate balance computation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SettingsHasher$hashBalanceArgs(computeCovariateBalanceArgs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{computeCovariateBalanceArgs}}{An object of type \code{ComputeCovariateBalanceArgs}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character string containing the MD5 hash (32 hex characters)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SettingsHasher-compareSettingsComponents"></a>}}
\if{latex}{\out{\hypertarget{method-SettingsHasher-compareSettingsComponents}{}}}
\subsection{Method \code{compareSettingsComponents()}}{
Compare old and new specifications to identify which components changed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SettingsHasher$compareSettingsComponents(oldSpecs, newSpecs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{oldSpecs}}{An object of type \code{CmAnalysesSpecifications} (old version)}

\item{\code{newSpecs}}{An object of type \code{CmAnalysesSpecifications} (new version)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of logical values indicating which components changed:
\itemize{
\item loadArgsChanged: whether data loading arguments differ
\item studyPopArgsChanged: whether study population arguments differ
\item psArgsChanged: whether propensity score arguments differ
\item strataArgsChanged: whether stratification arguments (trim/match/stratify) differ
\item outcomeModelArgsChanged: whether outcome model arguments differ
\item balanceArgsChanged: whether balance computation arguments differ
\item analyticsChanged: whether outcome of interest or analysis IDs changed
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SettingsHasher-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SettingsHasher-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SettingsHasher$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
