% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulateCohortData}
\alias{simulateCohortData}
\title{Generate simulated data}
\usage{
simulateCohortData(cohortDataSimulationProfile, n = 10000)
}
\arguments{
\item{cohortDataSimulationProfile}{An object of type \code{cohortDataSimulationProfile} as
generated using the \code{createCohortDataSimulationProfile} function.}

\item{n}{The size of the population to be generated.}
}
\value{
An object of type \code{cohortData}.
}
\description{
\code{simulateCohortData} creates a cohortData object with simulated data.
}
\details{
This function generates simulated data that is in many ways similar to the original data on which
the simulation profile is based. The contains same outcome, comparator, and outcome concept IDs,
and the covariates and their 1st order statistics should be comparable.
}

