% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{sqlMapply}
\alias{sqlMapply}
\title{Applies mapply to sql queries}
\usage{
sqlMapply(connection, sqlList, paramNames, parameters)
}
\arguments{
\item{connection}{Connection to sql database}

\item{sqlList}{List of sql queries to execute}

\item{paramNames}{List of parameter names}

\item{parameters}{List of list of parameters to substitute. length(parameters) = length(paramNames); length(parameters[[1]]) = length(sqlList)}
}
\value{
Returns list of completed sql queries.
}
\description{
This function works like \code{mapply} for a list of sql queries. Each sql query will be rendered, translated, and executed with proper
parameters.
}
\details{
This function currently assumes that all sql queries use the same parameters, and that values for each parameter are provided.
Here is a visualization of how the function works:

sqlMapply(connection, sqlList = list("query1 @a @b @c", "query2 @a @b @c"), paramNames = list("a", "b", "c"),
parameters = list(list(1,2), list(3,4), list(5,6)))

-> list(SqlRender::renderSql("query1 @a @b @c", a = 1, b = 3, c = 5), SqlRender::renderSql("query2 @a @b @c", a = 2, b = 4, c = 6)
}

