% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalyses.R
\name{summarizeAnalyses}
\alias{summarizeAnalyses}
\title{Create a summary report of the analyses}
\usage{
summarizeAnalyses(referenceTable, outputFolder, calibrate = TRUE)
}
\arguments{
\item{referenceTable}{A \link[dplyr:reexports]{dplyr::tibble} as created by the \link{runCmAnalyses} function.}

\item{outputFolder}{Name of the folder where all the outputs have been written to.}

\item{calibrate}{Perform empirical calibration? See details.}
}
\value{
A tibble containing summary statistics for each target-comparator-outcome-analysis combination.
}
\description{
Create a summary report of the analyses
}
\details{
When \code{calibrate = TRUE}, the \code{trueEffectSize} column in the \code{\link[=createOutcome]{createOutcome()}} function will be
used to identify negative (\code{trueEffectSize = 1}) and positive controls, These will then be used
to fit a systematic error model, which in turn will be used to calibrate p-values and confidence
intervals. If fewer than 5 controls have effect size estimates, no calibration will be performed.
}
