% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsFunctions.R
\name{trimByIptw}
\alias{trimByIptw}
\title{Remove subjects with a high IPTW}
\usage{
trimByIptw(population, maxWeight = 10, estimator = "ate")
}
\arguments{
\item{population}{A data frame with at least the three columns described below.}

\item{maxWeight}{The maximum allowed IPTW.}

\item{estimator}{The type of estimator. Options are \code{estimator = "ate"} for the average treatment
effect, and \code{estimator = "att"}for the average treatment effect in the treated.}
}
\value{
Returns a tibble with the same  columns as the input, as well as a \code{weights} column containing the
IPTW.
}
\description{
Compute the inverse probability of treatment weights (IPTW) using the propensity scores, and remove
subjects having a weight higher than the user-specified threshold.
}
\details{
The data frame should have the following three columns:
\itemize{
\item rowId (numeric): A unique identifier for each row (e.g. the person ID).
\item treatment (integer): Column indicating whether the person is in the target (1) or comparator (0) group.
\item propensityScore (numeric): Propensity score.
}
}
\examples{
rowId <- 1:2000
treatment <- rep(0:1, each = 1000)
propensityScore <- c(runif(1000, min = 0, max = 1), runif(1000, min = 0, max = 1))
data <- data.frame(rowId = rowId, treatment = treatment, propensityScore = propensityScore)
result <- trimByIptw(data)

}
