% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsFunctions.R
\name{trimByPsToEquipoise}
\alias{trimByPsToEquipoise}
\title{Keep only persons in clinical equipoise}
\usage{
trimByPsToEquipoise(population, bounds = c(0.3, 0.7))
}
\arguments{
\item{population}{A data frame with at least the three columns described below.}

\item{bounds}{The upper and lower bound on the preference score for keeping persons.}
}
\value{
Returns a tibble with the same three columns as the input.
}
\description{
Use the preference score to trim subjects that are not in clinical equipoise
}
\details{
The data frame should have the following three columns:
\itemize{
\item rowId (numeric): A unique identifier for each row (e.g. the person ID).
\item treatment (integer): Column indicating whether the person is in the target (1) or comparator (0) group.
\item propensityScore (numeric): Propensity score.
}
}
\examples{
rowId <- 1:2000
treatment <- rep(0:1, each = 1000)
propensityScore <- c(runif(1000, min = 0, max = 1), runif(1000, min = 0, max = 1))
data <- data.frame(rowId = rowId, treatment = treatment, propensityScore = propensityScore)
result <- trimByPsToEquipoise(data)

}
\references{
Walker AM, Patrick AR, Lauer MS, Hornbrook MC, Marin MG, Platt R, Roger VL, Stang P, and
Schneeweiss S. (2013) A tool for assessing the feasibility of comparative effectiveness research,
Comparative Effective Research, 3, 11-20
}
