% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trimMatchStratify}
\alias{trimMatchStratify}
\title{Trim, match or stratify persons by propensity score}
\usage{
trimMatchStratify(propensityScores, method = "matching", caliper = 0,
  matchRatio = 1, strata = 5, trimFraction = 0.05)
}
\arguments{
\item{propensityScores}{A data frame with three columns: \code{row_id} (person ID), \code{treat} and \code{propensityScore}.}

\item{method}{Whether \code{matching} or \code{stratification} should be performed.}

\item{caliper}{The caliper for matching. A caliper is the distance which is acceptable for
any match. Observations which are outside of the caliper are dropped. The caliper is interpreted to be in
standardized units. For example, caliper=.25 means that all matches not equal to or within .25 standard
deviations are dropped. A caliper of 0 means no caliper is used.}

\item{matchRatio}{Number of persons in comparator arm to be matched to each person in the treatment arm.}

\item{strata}{If stratification is performed, how many strata?
If not specified, all analyses will be executed. See \code{data(analysesDetails)} for a list of all Achilles analyses and their Ids.}

\item{trimFraction}{This fraction will be removed from each treatment group. In the treatment group, persons
with the highest propensity scores will be removed, in the comparator group person with the lowest scores will be removed.}
}
\value{
Returns a date frame with two columns: \code{row_id} and \code{stratum_id}
}
\description{
\code{trimMatchStratify} uses the provided propensity scores to trim subjects with extreme
scores, and either match or stratify persons.
}
\details{
Todo
}
\examples{
\dontrun{

}
}

