% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{truncateRawCodes}
\alias{truncateRawCodes}
\title{Applies truncating function to source codes}
\usage{
truncateRawCodes(rawCodes, dimensionTable)
}
\arguments{
\item{rawCodes}{Ffdf object with columns: \code{SOURCE_CODE} and \code{TARGET_CONCEPT_ID}}

\item{dimensionTable}{Data dimension information, including truncating function}
}
\value{
Returns \code{rawCodes} with \code{SOURCE_CODE} truncated if required
}
\description{
Truncates source codes according to data dimension
}
\details{
The input is obtained from the CDM, and the \code{SOURCE_CODE} can be a factor (integer) or a number (double).
Returns a list to work properly with mapply.
}

