% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CohortMethod.R
\docType{data}
\name{vignetteAnalysisSummary}
\alias{vignetteAnalysisSummary}
\title{Analysis summary data for the vignette}
\usage{
data(vignetteAnalysisSummary)
}
\description{
Analysis summary data for the vignette
}
\keyword{datasets}

