% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMethod.R
\docType{data}
\name{vignetteOutcomeModel1}
\alias{vignetteOutcomeModel1}
\title{Outcome data for the vignette}
\usage{
data(vignetteOutcomeModel1)
}
\description{
Outcome data for the vignette
}
\keyword{datasets}

