% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{calculateBias}
\alias{calculateBias}
\title{Calculates bias and relative risk for each covariate}
\usage{
calculateBias(data, cohortData, RR0Constant)
}
\arguments{
\item{data}{Covariate information for data dimension, as ffdf object with columns \code{rowId}, \code{covariateId},
\code{treatment}, \code{outcome}}

\item{cohortData}{\code{cohortData} object constructed by \code{getDbCohortData}}

\item{RR0Constant}{Large constant to use when RR = 0 to prevent divide by zero error}
}
\value{
Returns ffdf object with the following columns: \describe{
\item{covariateId}{Covariate identifier}
\item{RR}{Relative risk of covariate on outcome}
\item{bias}{The value of abs(log(bias)) for each covariate, ranked decreasingly}}
Returns list to work with sapply.
}
\description{
This function calculates the bias, as described in the hdps implementation, and the relative risk of each \code{covariateId} for a
data dimension.
}

