% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{calculateRanks}
\alias{calculateRanks}
\title{Calculates bias and exposure ranks}
\usage{
calculateRanks(data, cohortData, fudge)
}
\arguments{
\item{data}{Covariate information for data dimension with treatment and outcome}

\item{cohortData}{CohortMethodData object}

\item{fudge}{Fudge factor to use to avoid divide by 0. If fudge = 0, relative risks and bias with 0 or Inf values are thrown out}
}
\value{
Returns ffdf with columns \describe{
\item{covariateId}{covariate ID}
\item{biasRank}{bias score for covariate = abs(log(bias))}
\item{expRank}{exposure score for covariate = abs(log(prevalence_treated/prevalence_control))}
\item{PC0}{prevalence of covariate among control, for breaking ties}
\item{PC1}{prevalence of covariate among treated, for breaking ties}}
}
\description{
Calculates bias and exposure ranks
}

