% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMethod.R
\docType{package}
\name{CohortMethod-package}
\alias{CohortMethod}
\alias{CohortMethod-package}
\title{CohortMethod: New-user cohort method with large scale propensity and outcome models}
\description{
CohortMethod is an R package for performing new-user cohort studies
	in an observational database in the OMOP Common Data Model. It extracts the
	necessary data from a database and uses a large set of covariates for both the 
	propensity and outcome model, including for example all drugs, diagnoses, procedures, 
	as well as age, comorbidity indexes, etc. Large scale regularized regression is used to 
	fit the propensity and outcome models. Functions are included for trimming, stratifying, 
	(variable and fixed ratio) matching and weighting by propensity scores, as well as 
	diagnostic functions, such as propensity score distribution plots and plots showing 
	covariate balance before and after matching and/or trimming. Supported outcome models 
	are (conditional) logistic regression, (conditional) Poisson regression, and 
	(stratified) Cox regression. Also included are Kaplan-Meier plots that can adjust for 
	the stratification or matching.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CohortMethod}
  \item \url{https://github.com/OHDSI/CohortMethod}
  \item Report bugs at \url{https://github.com/OHDSI/CohortMethod/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

Authors:
\itemize{
  \item Marc Suchard
  \item Patrick Ryan
}

}
\keyword{internal}
