% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{createCmAnalysis}
\alias{createCmAnalysis}
\title{Create a CohortMethod analysis specification}
\usage{
createCmAnalysis(
  analysisId = 1,
  description = "",
  getDbCohortMethodDataArgs,
  createStudyPopArgs,
  createPs = FALSE,
  createPsArgs = NULL,
  trimByPs = FALSE,
  trimByPsArgs = NULL,
  trimByPsToEquipoise = FALSE,
  trimByPsToEquipoiseArgs = NULL,
  trimByIptw = FALSE,
  trimByIptwArgs = NULL,
  matchOnPs = FALSE,
  matchOnPsArgs = NULL,
  matchOnPsAndCovariates = FALSE,
  matchOnPsAndCovariatesArgs = NULL,
  stratifyByPs = FALSE,
  stratifyByPsArgs = NULL,
  stratifyByPsAndCovariates = FALSE,
  stratifyByPsAndCovariatesArgs = NULL,
  computeSharedCovariateBalance = FALSE,
  computeSharedCovariateBalanceArgs = NULL,
  computeCovariateBalance = FALSE,
  computeCovariateBalanceArgs = NULL,
  fitOutcomeModel = FALSE,
  fitOutcomeModelArgs = NULL
)
}
\arguments{
\item{analysisId}{An integer that will be used later to refer to this specific
set of analysis choices.}

\item{description}{A short description of the analysis.}

\item{getDbCohortMethodDataArgs}{An object representing the arguments to be used when calling
the \code{\link[=getDbCohortMethodData]{getDbCohortMethodData()}} function.}

\item{createStudyPopArgs}{An object representing the arguments to be used when calling
the \code{\link[=createStudyPopulation]{createStudyPopulation()}} function.}

\item{createPs}{Should the \code{\link[=createPs]{createPs()}} function be used in this
analysis?}

\item{createPsArgs}{An object representing the arguments to be used when calling
the \code{\link[=createPs]{createPs()}} function.}

\item{trimByPs}{Should the \code{\link[=trimByPs]{trimByPs()}} function be used in this
analysis?}

\item{trimByPsArgs}{An object representing the arguments to be used when calling
the \code{\link[=trimByPs]{trimByPs()}} function.}

\item{trimByPsToEquipoise}{Should the \code{\link[=trimByPsToEquipoise]{trimByPsToEquipoise()}} function be used
in this analysis?}

\item{trimByPsToEquipoiseArgs}{An object representing the arguments to be used when calling
the \code{\link[=trimByPsToEquipoise]{trimByPsToEquipoise()}} function.}

\item{trimByIptw}{Should the \code{\link[=trimByPsToEquipoise]{trimByPsToEquipoise()}} function be used
in this analysis?}

\item{trimByIptwArgs}{An object representing the arguments to be used when calling
the \code{\link[=trimByIptw]{trimByIptw()}} function.}

\item{matchOnPs}{Should the \code{\link[=matchOnPs]{matchOnPs()}} function be used in this
analysis?}

\item{matchOnPsArgs}{An object representing the arguments to be used when calling
the \code{\link[=matchOnPs]{matchOnPs()}} function.}

\item{matchOnPsAndCovariates}{Should the \code{\link[=matchOnPsAndCovariates]{matchOnPsAndCovariates()}} function be
used in this analysis?}

\item{matchOnPsAndCovariatesArgs}{An object representing the arguments to be used when calling
the \code{\link[=matchOnPsAndCovariates]{matchOnPsAndCovariates()}} function.}

\item{stratifyByPs}{Should the \code{\link[=stratifyByPs]{stratifyByPs()}} function be used in
this analysis?}

\item{stratifyByPsArgs}{An object representing the arguments to be used when calling
the \code{\link[=stratifyByPs]{stratifyByPs()}} function.}

\item{stratifyByPsAndCovariates}{Should the \code{\link[=stratifyByPsAndCovariates]{stratifyByPsAndCovariates()}} function
be used in this analysis?}

\item{stratifyByPsAndCovariatesArgs}{An object representing the arguments to be used when calling
the \code{\link[=stratifyByPsAndCovariates]{stratifyByPsAndCovariates()}} function.}

\item{computeSharedCovariateBalance}{Compute the covariate balance per target-comparator-analysis (using \code{\link[=computeCovariateBalance]{computeCovariateBalance()}})?
This will compute covariate balance across outcomes, for a study population that is not specific
to any outcome.}

\item{computeSharedCovariateBalanceArgs}{An object representing the arguments to be used when calling
the \code{\link[=computeCovariateBalance]{computeCovariateBalance()}} function per target-comparator-analysis.}

\item{computeCovariateBalance}{Compute the covariate balance per target-comparator-outcome-analysis (using \code{\link[=computeCovariateBalance]{computeCovariateBalance()}})?}

\item{computeCovariateBalanceArgs}{An object representing the arguments to be used when calling
the \code{\link[=computeCovariateBalance]{computeCovariateBalance()}} function per target-comparator-outcome-analysis.}

\item{fitOutcomeModel}{Should the \code{\link[=fitOutcomeModel]{fitOutcomeModel()}} function be used in
this analysis?}

\item{fitOutcomeModelArgs}{An object representing the arguments to be used when calling
the \code{\link[=fitOutcomeModel]{fitOutcomeModel()}} function.}
}
\description{
Create a CohortMethod analysis specification
}
\details{
Create a set of analysis choices, to be used with the \code{\link[=runCmAnalyses]{runCmAnalyses()}} function.
}
