% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{createCmAnalysis}
\alias{createCmAnalysis}
\title{Create a CohortMethod analysis specification}
\usage{
createCmAnalysis(analysisId = 1, description = "", targetType = NULL,
  comparatorType = NULL, getDbCohortMethodDataArgs, createStudyPopArgs,
  createPs = FALSE, createPsArgs = NULL, trimByPs = FALSE,
  trimByPsArgs = NULL, trimByPsToEquipoise = FALSE,
  trimByPsToEquipoiseArgs = NULL, matchOnPs = FALSE,
  matchOnPsArgs = NULL, matchOnPsAndCovariates = FALSE,
  matchOnPsAndCovariatesArgs = NULL, stratifyByPs = FALSE,
  stratifyByPsArgs = NULL, stratifyByPsAndCovariates = FALSE,
  stratifyByPsAndCovariatesArgs = NULL, fitOutcomeModel = FALSE,
  fitOutcomeModelArgs = NULL)
}
\arguments{
\item{analysisId}{An integer that will be used later to refer to this specific
set of analysis choices.}

\item{description}{A short description of the analysis.}

\item{targetType}{If more than one target is provided for each
drugComparatorOutcome, this field should be used to select
the specific target to use in this analysis.}

\item{comparatorType}{If more than one comparator is provided for each
drugComparatorOutcome, this field should be used to select
the specific comparator to use in this analysis.}

\item{getDbCohortMethodDataArgs}{An object representing the arguments to be used when calling
the \code{\link{getDbCohortMethodData}} function.}

\item{createStudyPopArgs}{An object representing the arguments to be used when calling
the \code{\link{createStudyPopulation}} function.}

\item{createPs}{Should the \code{\link{createPs}} function be used in this
analysis?}

\item{createPsArgs}{An object representing the arguments to be used when calling
the \code{\link{createPs}} function.}

\item{trimByPs}{Should the \code{\link{trimByPs}} function be used in this
analysis?}

\item{trimByPsArgs}{An object representing the arguments to be used when calling
the \code{\link{trimByPs}} function.}

\item{trimByPsToEquipoise}{Should the \code{\link{trimByPsToEquipoise}} function be used
in this analysis?}

\item{trimByPsToEquipoiseArgs}{An object representing the arguments to be used when calling
the \code{\link{trimByPsToEquipoise}} function.}

\item{matchOnPs}{Should the \code{\link{matchOnPs}} function be used in this
analysis?}

\item{matchOnPsArgs}{An object representing the arguments to be used when calling
the \code{\link{matchOnPs}} function.}

\item{matchOnPsAndCovariates}{Should the \code{\link{matchOnPsAndCovariates}} function be
used in this analysis?}

\item{matchOnPsAndCovariatesArgs}{An object representing the arguments to be used when calling
the \code{\link{matchOnPsAndCovariates}} function.}

\item{stratifyByPs}{Should the \code{\link{stratifyByPs}} function be used in
this analysis?}

\item{stratifyByPsArgs}{An object representing the arguments to be used when calling
the \code{\link{stratifyByPs}} function.}

\item{stratifyByPsAndCovariates}{Should the \code{\link{stratifyByPsAndCovariates}} function
be used in this analysis?}

\item{stratifyByPsAndCovariatesArgs}{An object representing the arguments to be used when calling
the \code{\link{stratifyByPsAndCovariates}} function.}

\item{fitOutcomeModel}{Should the \code{\link{fitOutcomeModel}} function be used in
this analysis?}

\item{fitOutcomeModelArgs}{An object representing the arguments to be used when calling
the \code{\link{fitOutcomeModel}} function.}
}
\description{
Create a CohortMethod analysis specification
}
\details{
Create a set of analysis choices, to be used with the \code{\link{runCmAnalyses}} function.
}
