% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettingsObjects.R
\name{createCmDiagnosticThresholds}
\alias{createCmDiagnosticThresholds}
\title{Create CohortMethod diagnostics thresholds}
\usage{
createCmDiagnosticThresholds(
  mdrrThreshold = 10,
  easeThreshold = 0.25,
  sdmThreshold = 0.1,
  sdmAlpha = NULL,
  equipoiseThreshold = 0.2,
  generalizabilitySdmThreshold = 999
)
}
\arguments{
\item{mdrrThreshold}{What is the maximum allowed minimum detectable relative risk
(MDRR)?}

\item{easeThreshold}{What is the maximum allowed expected absolute systematic error
(EASE).}

\item{sdmThreshold}{What is the maximum allowed standardized difference of mean (SDM)? If
any covariate has an SDM exceeding this threshold, the diagnostic will
fail.}

\item{sdmAlpha}{What is the alpha for testing whether the absolute SDM exceeds
\code{sdmThreshold}? If not provided, no significance testing will be
performed and any absolute SDM greater than the threshold will be
considered imbalance. Note that a Bonferroni adjustment will
automatically be applied to adjust for the number of tests performed.}

\item{equipoiseThreshold}{What is the minimum required equipoise?}

\item{generalizabilitySdmThreshold}{What is the maximum allowed standardized difference of mean
(SDM)when comparing the population before and after PS
adjustments? If the SDM is greater than this value, the diagnostic
will fail.}
}
\value{
An object of type \code{CmDiagnosticThresholds}.
}
\description{
Threshold used when calling \code{\link[=exportToCsv]{exportToCsv()}} to determine if we pass or fail diagnostics.
}
\details{
The \code{sdmThreshold} and \code{sdmAlpha} arguments are independent of the \code{threshold} and \code{alpha}
threshold provided to the \code{createComputeCovariateBalanceArgs()} function. The latter have no
impact on blinding and diagnostics reported in the export.
}
