% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{createCohortMethodDataSimulationProfile}
\alias{createCohortMethodDataSimulationProfile}
\title{Create simulation profile}
\usage{
createCohortMethodDataSimulationProfile(cohortMethodData, minCellCount = 5)
}
\arguments{
\item{cohortMethodData}{An object of type \link{CohortMethodData} as generated using
\code{\link[=getDbCohortMethodData]{getDbCohortMethodData()}}.}

\item{minCellCount}{If > 0, will apply \code{.truncateSimulationProfile()}}
}
\value{
An object of type \code{CohortDataSimulationProfile}.
}
\description{
Creates a profile based on the provided
\link{CohortMethodData} object, which can be used to generate simulated data that has similar
characteristics.
}
\details{
The output of this function is an object that can be used by the \code{\link[=simulateCohortMethodData]{simulateCohortMethodData()}}
function to generate a cohortMethodData object.
}
