% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{createCohortMethodDataSimulationProfile}
\alias{createCohortMethodDataSimulationProfile}
\title{Create simulation profile}
\usage{
createCohortMethodDataSimulationProfile(cohortMethodData)
}
\arguments{
\item{cohortMethodData}{An object of type \code{cohortMethodData} as generated using
\code{getDbCohortMethodData}.}
}
\value{
An object of type \code{cohortDataSimulationProfile}.
}
\description{
\code{createCohortMethodDataSimulationProfile} creates a profile based on the provided
cohortMethodData object, which can be used to generate simulated data that has similar
characteristics.
}
\details{
The output of this function is an object that can be used by the \code{simulateCohortMethodData}
function to generate a cohortMethodData object.
}
