% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettingsObjects.R
\name{createComputeCovariateBalanceArgs}
\alias{createComputeCovariateBalanceArgs}
\title{Create a parameter object for the function \code{\link[=computeCovariateBalance]{computeCovariateBalance()}}}
\usage{
createComputeCovariateBalanceArgs(
  subgroupCovariateId = NULL,
  maxCohortSize = 250000,
  covariateFilter = NULL,
  threshold = 0.1,
  alpha = 0.05
)
}
\arguments{
\item{subgroupCovariateId}{Optional: a covariate ID of a binary covariate that indicates a subgroup of
interest. Both the before and after populations will be restricted to this
subgroup before computing covariate balance.}

\item{maxCohortSize}{If the target or comparator cohort are larger than this number, they
will be downsampled before computing covariate balance to save time.
Setting this number to 0 means no downsampling will be applied.}

\item{covariateFilter}{Determines the covariates for which to compute covariate balance. Either a vector
of covariate IDs, or a table 1 specifications object as generated for example using
\code{\link[FeatureExtraction:getDefaultTable1Specifications]{FeatureExtraction::getDefaultTable1Specifications()}}. If \code{covariateFilter = NULL},
balance will be computed for all variables found in the data.}

\item{threshold}{Threshold value for the absolute value of the standardized difference of means (ASDM).
If the ASDM exceeds this threshold it will be marked as unbalanced. (Hripcsak et al. 2025)}

\item{alpha}{The family-wise alpha for testing whether the absolute value of the standardized
difference of means is greater than the threshold. If not provided, any value greater
than the threshold will be marked as unbalanced.}
}
\value{
An object of type \code{ComputeCovariateBalanceArgs}.
}
\description{
Create a parameter object for the function \code{\link[=computeCovariateBalance]{computeCovariateBalance()}}
}
\details{
Create an object defining the parameter values.
}
\references{
Hripcsak G, Zhang L, Chen Y, Li K, Suchard MA, Ryan PB, Schuemie MJ (2025)
Assessing Covariate Balance with Small Sample Sizes. medRxiv. Feb 21:2024.04.23.24306230.
}
