% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createCreatePsArgs}
\alias{createCreatePsArgs}
\title{Create parameter object for the function createPs}
\usage{
createCreatePsArgs(excludeCovariateIds = NULL,
  prior = createPrior("laplace", exclude = c(0), useCrossValidation = TRUE),
  control = createControl(noiseLevel = "silent", cvType = "auto",
  startingVariance = 0.1))
}
\arguments{
\item{excludeCovariateIds}{Exclude these covariates from the propensity model.}

\item{prior}{The prior used to fit the model. See createPriorfor details.}

\item{control}{The control object used to control the cross-validation used
                             todetermine the hyperparameters of the prior (if applicable).
                             SeecreateControl for details.}
}
\description{
Create parameter object for the function createPs
}
\details{
Create an object defining the parameter values. To be used in the
\code{\link{createCohortMethodAnalysis}} function.
}

