% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createCreateStudyPopulationArgs}
\alias{createCreateStudyPopulationArgs}
\title{Create a parameter object for the function createStudyPopulation}
\usage{
createCreateStudyPopulationArgs(firstExposureOnly = FALSE,
  washoutPeriod = 0, removeDuplicateSubjects = FALSE,
  removeSubjectsWithPriorOutcome = TRUE, priorOutcomeLookback = 99999,
  minDaysAtRisk = 1, riskWindowStart = 0, addExposureDaysToStart = FALSE,
  riskWindowEnd = 0, addExposureDaysToEnd = TRUE)
}
\arguments{
\item{firstExposureOnly}{Should only the first exposure per subject be included? Note
that this is typically done in the createStudyPopulation
function,}

\item{washoutPeriod}{The mininum required continuous observation time prior to
index date for a person to be included in the cohort.}

\item{removeDuplicateSubjects}{Remove subjects that are in both the treated and comparator
cohort?}

\item{removeSubjectsWithPriorOutcome}{Remove subjects that have the outcome prior to the risk
window start?}

\item{priorOutcomeLookback}{How many days should we look back when identifying prior
outcomes?}

\item{minDaysAtRisk}{The minimum required number of days at risk.}

\item{riskWindowStart}{The start of the risk window (in days) relative to the index
date (+days of exposure if the addExposureDaysToStart
parameter isspecified).}

\item{addExposureDaysToStart}{Add the length of exposure the start of the risk window?}

\item{riskWindowEnd}{The end of the risk window (in days) relative to the index
data (+days of exposure if the addExposureDaysToEnd
parameter isspecified).}

\item{addExposureDaysToEnd}{Add the length of exposure the risk window?}
}
\description{
Create a parameter object for the function createStudyPopulation
}
\details{
Create an object defining the parameter values.
}

