% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createCreateStudyPopulationArgs}
\alias{createCreateStudyPopulationArgs}
\title{Create a parameter object for the function createStudyPopulation}
\usage{
createCreateStudyPopulationArgs(firstExposureOnly = FALSE,
  restrictToCommonPeriod = FALSE, washoutPeriod = 0,
  removeDuplicateSubjects = FALSE,
  removeSubjectsWithPriorOutcome = TRUE, priorOutcomeLookback = 99999,
  minDaysAtRisk = 1, riskWindowStart = 0,
  addExposureDaysToStart = NULL, startAnchor = "cohort start",
  riskWindowEnd = 0, addExposureDaysToEnd = NULL,
  endAnchor = "cohort end", censorAtNewRiskWindow = FALSE)
}
\arguments{
\item{firstExposureOnly}{Should only the first exposure per subject be included?
Notethat this is typically done in thecreateStudyPopulation
function,}

\item{restrictToCommonPeriod}{Restrict the analysis to the period when both exposures are
observed?}

\item{washoutPeriod}{The mininum required continuous observation time prior
toindex date for a person to be included in the cohort.}

\item{removeDuplicateSubjects}{Remove subjects that are in both the target and
comparatorcohort? See details for allowed values.}

\item{removeSubjectsWithPriorOutcome}{Remove subjects that have the outcome prior to the
riskwindow start?}

\item{priorOutcomeLookback}{How many days should we look back when identifying
prioroutcomes?}

\item{minDaysAtRisk}{The minimum required number of days at risk.}

\item{riskWindowStart}{The start of the risk window (in days) relative to the
startAnchor.}

\item{addExposureDaysToStart}{DEPRECATED: Add the length of exposure the start of the risk
window?Use startAnchor instead.}

\item{startAnchor}{The anchor point for the start of the risk window. Can be
"cohort start"or "cohort end".}

\item{riskWindowEnd}{The end of the risk window (in days) relative to the
endAnchor.}

\item{addExposureDaysToEnd}{DEPRECATED: Add the length of exposure the risk window?Use
endAnchor instead.}

\item{endAnchor}{The anchor point for the end of the risk window. Can be
"cohort start"or "cohort end".}

\item{censorAtNewRiskWindow}{If a subject is in multiple cohorts, should time-at-risk be
censoredwhen the new time-at-risk starts to prevent overlap?}
}
\description{
Create a parameter object for the function createStudyPopulation
}
\details{
Create an object defining the parameter values.
}
