% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createCyclopsData.ffdf}
\alias{createCyclopsData.ffdf}
\title{Convert data from two ffdf objects into a CyclopsData object}
\usage{
createCyclopsData.ffdf(outcomes, covariates, modelType = "lr",
  addIntercept = TRUE, offsetAlreadyOnLogScale = FALSE,
  makeCovariatesDense = NULL, checkSorting = TRUE)
}
\arguments{
\item{outcomes}{A ffdf object containing the outcomes with predefined columns (see below).}

\item{covariates}{A ffdf object containing the covariates with predefined columns (see below).}

\item{modelType}{Cyclops model type. Current supported types are "pr", "cpr", lr", "clr", or "cox"}

\item{addIntercept}{Add an intercept to the model?}

\item{offsetAlreadyOnLogScale}{Is the time variable already on a log scale?}

\item{checkSorting}{Check if the data is sorted appropriately, and if not, sort.}

\item{useOffsetCovariate}{Use the time variable in the model as an offset?}
}
\value{
An object of type CyclopsModel
}
\description{
\code{createCyclopsData.ffdf} loads data from two ffdf objects, and inserts it into a Cyclops data object.
}
\details{
These columns are expected in the outcome object:
\tabular{lll}{
  \verb{stratum_id}    \tab(integer) \tab (optional) Stratum ID for conditional regression models \cr
  \verb{row_id}  	\tab(integer) \tab Row ID is used to link multiple covariates (x) to a single outcome (y) \cr
  \verb{y}    \tab(real) \tab The outcome variable \cr
  \verb{time}    \tab(real) \tab For models that use time (e.g. Poisson or Cox regression) this contains time (e.g. number of days) \cr
}
The outcome table should be sorted by stratum_id (if present) and then row_id except for Cox regression when
the table should be sorted by stratum_id (if present), -time, y, and row_id.

These columns are expected in the covariates object:
\tabular{lll}{
  \verb{stratum_id}    \tab(integer) \tab (optional) Stratum ID for conditional regression models \cr
  \verb{row_id}  	\tab(integer) \tab Row ID is used to link multiple covariates (x) to a single outcome (y) \cr
  \verb{covariate_id}    \tab(integer) \tab A numeric identifier of a covariate  \cr
  \verb{covariate_value}    \tab(real) \tab The value of the specified covariate \cr
}
The covariate table should be sorted by stratum_id (if present), row_id and covariate_id except for Cox regression when
the table should be sorted by stratum_id (if present), -time, y, and row_id.
}
\examples{
#Convert infert dataset to Cyclops format:
covariates <- data.frame(stratum_id = rep(infert$stratum,2),
                         row_id = rep(1:nrow(infert),2),
                         covariate_id = rep(1:2,each=nrow(infert)),
                         covariate_value = c(infert$spontaneous,infert$induced))
outcomes <- data.frame(stratum_id = infert$stratum,
                       row_id = 1:nrow(infert),
                       y = infert$case)
#Make sparse:
covariates <- covariates[covariates$covariate_value != 0,]

#Sort:
covariates <- covariates[order(covariates$stratum_id,covariates$row_id,covariates$covariate_id),]
outcomes <- outcomes[order(outcomes$stratum_id,outcomes$row_id),]

#Convert to ffdf:
covariates <- as.ffdf(covariates)
outcomes <- as.ffdf(outcomes)

#Create Cyclops data object:
cyclopsData <- createCyclopsData.ffdf(outcomes,covariates,modelType = "clr",addIntercept = FALSE)

#Fit model:
fit <- fitCyclopsModel(cyclopsData,prior = prior("none"))
}

