% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createFitOutcomeModelArgs}
\alias{createFitOutcomeModelArgs}
\title{Create a parameter object for the function fitOutcomeModel}
\usage{
createFitOutcomeModelArgs(stratifiedCox = TRUE, riskWindowStart = 0,
  riskWindowEnd = 9999, addExposureDaysToEnd = FALSE,
  useCovariates = TRUE, fitModel = TRUE, modelType = "cox",
  prior = createPrior("laplace", useCrossValidation = TRUE),
  control = createControl(cvType = "auto", startingVariance = 0.1,
  selectorType = "byPid", noiseLevel = "quiet"))
}
\arguments{
\item{stratifiedCox}{Specifically for Cox regressions: specify whether to use the
stratadefined in subPopulation in the analysis. For Poissonregression
and logistic regression, this is implied in 'clr' and'cpr'.}

\item{riskWindowStart}{The start of the risk window (in days) relative to the index data.}

\item{riskWindowEnd}{The end of the risk window (in days) relative to the index data (+days
of exposure if the addExposureDaysToEnd parameter isspecified).}

\item{addExposureDaysToEnd}{Add the length of exposure the risk window?}

\item{useCovariates}{Whether to use the covariate matrix in the cohortMethodData in
theoutcome model.}

\item{fitModel}{If false, the model will not be fit, and only summary statistics
areavailable.}

\item{modelType}{The type of model to be fitted. See details for options.}

\item{prior}{The prior used to fit the model. SeecreatePrior for details.}

\item{control}{The control object used to control the cross-validation used
todetermine the hyperparameters of the prior (if applicable).
SeecreateControl for details.}
}
\description{
Create a parameter object for the function fitOutcomeModel
}
\details{
Create an object defining the parameter values.
}

