% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createFitOutcomeModelArgs}
\alias{createFitOutcomeModelArgs}
\title{Create a parameter object for the function fitOutcomeModel}
\usage{
createFitOutcomeModelArgs(modelType = "logistic", stratified = TRUE,
  useCovariates = TRUE, excludeCovariateIds = c(),
  includeCovariateIds = c(), prior = createPrior("laplace",
  useCrossValidation = TRUE), control = createControl(cvType = "auto",
  startingVariance = 0.01, tolerance = 2e-07, cvRepetitions = 10, noiseLevel =
  "quiet"))
}
\arguments{
\item{modelType}{The type of outcome model that will be used. Possible values
are"logistic", "poisson", or "cox".}

\item{stratified}{Should the regression be conditioned on the strata defined in the
populationobject (e.g. by matching or stratifying on propensity
scores)?}

\item{useCovariates}{Whether to use the covariate matrix in the cohortMethodData object in
theoutcome model.}

\item{excludeCovariateIds}{Exclude these covariates from the outcome model.}

\item{includeCovariateIds}{Include only these covariates in the outcome model.}

\item{prior}{The prior used to fit the model. SeecreatePrior for details.}

\item{control}{The control object used to control the cross-validation used
todetermine the hyperparameters of the prior (if applicable).
SeecreateControl for details.}
}
\description{
Create a parameter object for the function fitOutcomeModel
}
\details{
Create an object defining the parameter values.
}
