% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createFitOutcomeModelArgs}
\alias{createFitOutcomeModelArgs}
\title{Create a parameter object for the function fitOutcomeModel}
\usage{
createFitOutcomeModelArgs(modelType = "logistic", stratified = FALSE,
  useCovariates = FALSE, inversePtWeighting = FALSE,
  interactionCovariateIds = c(), excludeCovariateIds = c(),
  includeCovariateIds = c(), prior = createPrior("laplace",
  useCrossValidation = TRUE), control = createControl(cvType = "auto",
  seed = 1, startingVariance = 0.01, tolerance = 2e-07, cvRepetitions = 10,
  noiseLevel = "quiet"))
}
\arguments{
\item{modelType}{The type of outcome model that will be used. Possible values
are"logistic", "poisson", or "cox".}

\item{stratified}{Should the regression be conditioned on the strata defined in
thepopulation object (e.g. by matching or stratifying on
propensityscores)?}

\item{useCovariates}{Whether to use the covariate matrix in the cohortMethodDataobject
in the outcome model.}

\item{inversePtWeighting}{Use inverse probability of treatment weigting?}

\item{interactionCovariateIds}{An optional vector of covariate IDs to use to estimate
interactionswith the main treatment effect.}

\item{excludeCovariateIds}{Exclude these covariates from the outcome model.}

\item{includeCovariateIds}{Include only these covariates in the outcome model.}

\item{prior}{The prior used to fit the model. See createPriorfor details.}

\item{control}{The control object used to control the cross-validation used
todetermine the hyperparameters of the prior (if applicable).
SeecreateControl for details.}
}
\description{
Create a parameter object for the function fitOutcomeModel
}
\details{
Create an object defining the parameter values.
}
