% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createGetDbCohortMethodDataArgs}
\alias{createGetDbCohortMethodDataArgs}
\title{Create parameter object for the function getDbCohortMethodData}
\usage{
createGetDbCohortMethodDataArgs(washoutWindow = 183,
  indicationLookbackWindow = 183, studyStartDate = "", studyEndDate = "",
  outcomeConditionTypeConceptIds = c(), excludeDrugsFromCovariates = TRUE,
  useCovariateDemographics = TRUE, useCovariateDemographicsGender = TRUE,
  useCovariateDemographicsRace = TRUE,
  useCovariateDemographicsEthnicity = TRUE,
  useCovariateDemographicsAge = TRUE, useCovariateDemographicsYear = TRUE,
  useCovariateDemographicsMonth = TRUE,
  useCovariateConditionOccurrence = TRUE,
  useCovariateConditionOccurrence365d = TRUE,
  useCovariateConditionOccurrence30d = TRUE,
  useCovariateConditionOccurrenceInpt180d = TRUE,
  useCovariateConditionEra = FALSE, useCovariateConditionEraEver = FALSE,
  useCovariateConditionEraOverlap = FALSE,
  useCovariateConditionGroup = FALSE,
  useCovariateConditionGroupMeddra = FALSE,
  useCovariateConditionGroupSnomed = FALSE,
  useCovariateDrugExposure = FALSE, useCovariateDrugExposure365d = FALSE,
  useCovariateDrugExposure30d = FALSE, useCovariateDrugEra = FALSE,
  useCovariateDrugEra365d = FALSE, useCovariateDrugEra30d = FALSE,
  useCovariateDrugEraOverlap = FALSE, useCovariateDrugEraEver = FALSE,
  useCovariateDrugGroup = FALSE, useCovariateProcedureOccurrence = FALSE,
  useCovariateProcedureOccurrence365d = FALSE,
  useCovariateProcedureOccurrence30d = FALSE,
  useCovariateProcedureGroup = FALSE, useCovariateObservation = FALSE,
  useCovariateObservation365d = FALSE, useCovariateObservation30d = FALSE,
  useCovariateObservationBelow = FALSE,
  useCovariateObservationAbove = FALSE,
  useCovariateObservationCount365d = FALSE,
  useCovariateConceptCounts = FALSE, useCovariateRiskScores = FALSE,
  useCovariateRiskScoresCharlson = FALSE,
  useCovariateRiskScoresDCSI = FALSE, useCovariateRiskScoresCHADS2 = FALSE,
  useCovariateInteractionYear = FALSE, useCovariateInteractionMonth = FALSE,
  excludedCovariateConceptIds = c(), includedCovariateConceptIds = c(),
  deleteCovariatesSmallCount = 100)
}
\arguments{
\item{washoutWindow}{The mininum required continuous observation time
prior to indexdate for a person to be included in
the cohort.}

\item{indicationLookbackWindow}{The window to look back prior to cohort index date
to identifyrecords of a indication condition.  Only
applicable ifindicationConceptIds != ”.}

\item{studyStartDate}{A calendar date specifying the minimum date that a
cohort indexdate can appear. Date format is
'yyyymmdd'.}

\item{studyEndDate}{A calendar date specifying the maximum date that a
cohort indexdate can appear. Date format is
'yyyymmdd'.}

\item{outcomeConditionTypeConceptIds}{A list of TYPE_CONCEPT_ID values that will
restrictcondition occurrences.  Only applicable if
outcomeTable =CONDITION_OCCURRENCE.}

\item{excludeDrugsFromCovariates}{Should the target and comparator drugs (and their
descendantconcepts) be excluded from the
covariates? Note that this willwork if the drugs
are actualy drug concept IDs (and not cohortIDs).}

\item{useCovariateDemographics}{A boolean value (TRUE/FALSE) to determine
ifdemographic covariates (age in 5-yr
increments,gender, race, ethnicity, year of index
date, monthof index date) will be created and
included infuture models.}

\item{useCovariateDemographicsGender}{A boolean value (TRUE/FALSE) to determine if
gendershould be included in the model.}

\item{useCovariateDemographicsRace}{A boolean value (TRUE/FALSE) to determine if
raceshould be included in the model.}

\item{useCovariateDemographicsEthnicity}{A boolean value (TRUE/FALSE) to determine
ifethnicity should be included in the model.}

\item{useCovariateDemographicsAge}{A boolean value (TRUE/FALSE) to determine if age(in
5 year increments) should be included in themodel.}

\item{useCovariateDemographicsYear}{A boolean value (TRUE/FALSE) to determine
ifcalendar year should be included in the model.}

\item{useCovariateDemographicsMonth}{A boolean value (TRUE/FALSE) to determine
ifcalendar month should be included in the model.}

\item{useCovariateConditionOccurrence}{A boolean value (TRUE/FALSE) to determine
ifcovariates derived from CONDITION_OCCURRENCE
tablewill be created and included in future models.}

\item{useCovariateConditionOccurrence365d}{A boolean value (TRUE/FALSE) to determine
ifcovariates will be created and used in models
thatlook for presence/absence of condition in
365dwindow prior to or on cohort index date.
Onlyapplicable if useCovariateConditionOccurrence
=TRUE.}

\item{useCovariateConditionOccurrence30d}{A boolean value (TRUE/FALSE) to determine
ifcovariates will be created and used in models
thatlook for presence/absence of condition in
30dwindow prior to or on cohort index date.
Onlyapplicable if useCovariateConditionOccurrence
=TRUE.}

\item{useCovariateConditionOccurrenceInpt180d}{A boolean value (TRUE/FALSE) to determine
ifcovariates will be created and used in models
thatlook for presence/absence of condition
withininpatient type in 180d window prior to or on
cohortindex date.  Only applicable
ifuseCovariateConditionOccurrence = TRUE.}

\item{useCovariateConditionEra}{A boolean value (TRUE/FALSE) to determine
ifcovariates derived from CONDITION_ERA table will
becreated and included in future models.}

\item{useCovariateConditionEraEver}{A boolean value (TRUE/FALSE) to determine
ifcovariates will be created and used in models
thatlook for presence/absence of condition era
anytimeprior to or on cohort index date.  Only
applicableif useCovariateConditionEra = TRUE.}

\item{useCovariateConditionEraOverlap}{A boolean value (TRUE/FALSE) to determine
ifcovariates will be created and used in models
thatlook for presence/absence of condition era
thatoverlaps the cohort index date.  Only
applicable ifuseCovariateConditionEra = TRUE.}

\item{useCovariateConditionGroup}{A boolean value (TRUE/FALSE) to determine if
allCONDITION_OCCURRENCE and CONDITION_ERA
covariatesshould be aggregated or rolled-up to
higher-levelconcepts based on vocabluary
classification.}

\item{useCovariateConditionGroupMeddra}{A boolean value (TRUE/FALSE) to determine if
allCONDITION_OCCURRENCE and CONDITION_ERA
covariatesshould be aggregated or rolled-up to
higher-levelconcepts based on the MEDDRA
classification.}

\item{useCovariateConditionGroupSnomed}{A boolean value (TRUE/FALSE) to determine if
allCONDITION_OCCURRENCE and CONDITION_ERA
covariatesshould be aggregated or rolled-up to
higher-levelconcepts based on the SNOMED
classification.}

\item{useCovariateDrugExposure}{A boolean value (TRUE/FALSE) to determine
ifcovariates derived from DRUG_EXPOSURE table will
becreated and included in future models.}

\item{useCovariateDrugExposure365d}{A boolean value (TRUE/FALSE) to determine
ifcovariates will be created and used in models
thatlook for presence/absence of drug in 365d
windowprior to or on cohort index date.  Only
applicableif useCovariateDrugExposure = TRUE.}

\item{useCovariateDrugExposure30d}{A boolean value (TRUE/FALSE) to determine
ifcovariates will be created and used in models
thatlook for presence/absence of drug in 30d
windowprior to or on cohort index date.  Only
applicableif useCovariateDrugExposure = TRUE.}

\item{useCovariateDrugEra}{A boolean value (TRUE/FALSE) to determine
ifcovariates derived from DRUG_ERA table will
becreated and included in future models.}

\item{useCovariateDrugEra365d}{A boolean value (TRUE/FALSE) to determine
ifcovariates will be created and used in models
thatlook for presence/absence of drug era in
365dwindow prior to or on cohort index date.
Onlyapplicable if useCovariateDrugEra = TRUE.}

\item{useCovariateDrugEra30d}{A boolean value (TRUE/FALSE) to determine
ifcovariates will be created and used in models
thatlook for presence/absence of drug era in 30d
windowprior to or on cohort index date.  Only
applicableif useCovariateDrugEra = TRUE.}

\item{useCovariateDrugEraOverlap}{A boolean value (TRUE/FALSE) to determine
ifcovariates will be created and used in models
thatlook for presence/absence of drug era that
overlapsthe cohort index date.  Only applicable
ifuseCovariateDrugEra = TRUE.}

\item{useCovariateDrugEraEver}{A boolean value (TRUE/FALSE) to determine
ifcovariates will be created and used in models
thatlook for presence/absence of drug era anytime
priorto or on cohort index date.  Only applicable
ifuseCovariateDrugEra = TRUE.}

\item{useCovariateDrugGroup}{A boolean value (TRUE/FALSE) to determine if
allDRUG_EXPOSURE and DRUG_ERA covariates should
beaggregated or rolled-up to higher-level concepts
ofdrug classes based on vocabluary classification.}

\item{useCovariateProcedureOccurrence}{A boolean value (TRUE/FALSE) to determine
ifcovariates derived from PROCEDURE_OCCURRENCE
tablewill be created and included in future models.}

\item{useCovariateProcedureOccurrence365d}{A boolean value (TRUE/FALSE) to determine
ifcovariates will be created and used in models
thatlook for presence/absence of procedure in
365dwindow prior to or on cohort index date.
Onlyapplicable if useCovariateProcedureOccurrence
=TRUE.}

\item{useCovariateProcedureOccurrence30d}{A boolean value (TRUE/FALSE) to determine
ifcovariates will be created and used in models
thatlook for presence/absence of procedure in
30dwindow prior to or on cohort index date.
Onlyapplicable if useCovariateProcedureOccurrence
=TRUE.}

\item{useCovariateProcedureGroup}{A boolean value (TRUE/FALSE) to determine if
allPROCEDURE_OCCURRENCE covariates should
beaggregated or rolled-up to higher-level
conceptsbased on vocabluary classification.}

\item{useCovariateObservation}{A boolean value (TRUE/FALSE) to determine
ifcovariates derived from OBSERVATION table will
becreated and included in future models.}

\item{useCovariateObservation365d}{A boolean value (TRUE/FALSE) to determine
ifcovariates will be created and used in models
thatlook for presence/absence of observation in
365dwindow prior to or on cohort index date.
Onlyapplicable if useCovariateObservation = TRUE.}

\item{useCovariateObservation30d}{A boolean value (TRUE/FALSE) to determine
ifcovariates will be created and used in models
thatlook for presence/absence of observation in
30dwindow prior to or on cohort index date.
Onlyapplicable if useCovariateObservation = TRUE.}

\item{useCovariateObservationBelow}{A boolean value (TRUE/FALSE) to determine
ifcovariates will be created and used in models
thatlook for presence/absence of observation with
anumeric value below normal range for latest
valuewithin 180d of cohort index.  Only applicable
ifuseCovariateObservation = TRUE.}

\item{useCovariateObservationAbove}{A boolean value (TRUE/FALSE) to determine
ifcovariates will be created and used in models
thatlook for presence/absence of observation with
anumeric value above normal range for latest
valuewithin 180d of cohort index.  Only applicable
ifuseCovariateObservation = TRUE.}

\item{useCovariateObservationCount365d}{A boolean value (TRUE/FALSE) to determine
ifcovariates will be created and used in models
thatlook for the count of each observation concept
in365d window prior to or on cohort index date.
Onlyapplicable if useCovariateObservation = TRUE.}

\item{useCovariateConceptCounts}{A boolean value (TRUE/FALSE) to determine
ifcovariates will be created and used in models
thatcount the number of concepts that a person
haswithin each domain (CONDITION, DRUG,
PROCEDURE,OBSERVATION)}

\item{useCovariateRiskScores}{A boolean value (TRUE/FALSE) to determine
ifcovariates will be created and used in models
thatcalculate various Risk Scores, including
Charlson,DCSI.}

\item{useCovariateRiskScoresCharlson}{A boolean value (TRUE/FALSE) to determine if
theCharlson comorbidity index should be included
inthe model.}

\item{useCovariateRiskScoresDCSI}{A boolean value (TRUE/FALSE) to determine if
theDCSI score should be included in the model.}

\item{useCovariateRiskScoresCHADS2}{A boolean value (TRUE/FALSE) to determine if
theCHADS2 score should be included in the model.}

\item{useCovariateInteractionYear}{A boolean value (TRUE/FALSE) to determine
ifcovariates will be created and used in models
thatrepresent interaction terms between all
othercovariates and the year of the cohort index
date.}

\item{useCovariateInteractionMonth}{A boolean value (TRUE/FALSE) to determine
ifcovariates will be created and used in models
thatrepresent interaction terms between all
othercovariates and the month of the cohort index
date.}

\item{excludedCovariateConceptIds}{A list of concept IDs that should NOT be used
toconstruct covariates.}

\item{includedCovariateConceptIds}{A list of concept IDs that should be used
toconstruct covariates.}

\item{deleteCovariatesSmallCount}{A numeric value used to remove covariates thatoccur
                                                 in both cohorts fewer
                                                 thandeleteCovariateSmallCounts time.}
}
\description{
Create parameter object for the function getDbCohortMethodData
}
\details{
Create an object defining the parameter values. To be used in the
\code{\link{createCohortMethodAnalysis}} function.
}

