% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createGetDbCohortMethodDataArgs}
\alias{createGetDbCohortMethodDataArgs}
\title{Create a parameter object for the function getDbCohortMethodData}
\usage{
createGetDbCohortMethodDataArgs(
  studyStartDate = "",
  studyEndDate = "",
  firstExposureOnly = FALSE,
  removeDuplicateSubjects = FALSE,
  restrictToCommonPeriod = FALSE,
  washoutPeriod = 0,
  maxCohortSize = 0,
  covariateSettings
)
}
\arguments{
\item{studyStartDate}{A calendar date specifying the minimum date that a cohort index date can appear. Date format is 'yyyymmdd'.}

\item{studyEndDate}{A calendar date specifying the maximum date that a cohort index date can appear. Date format is 'yyyymmdd'. Important: the study end data is also used to truncate risk windows, meaning no outcomes beyond the study end date will be considered.}

\item{firstExposureOnly}{Should only the first exposure per subject be included? Note that this is typically done in the createStudyPopulation() function, but can already be done here for efficiency reasons.}

\item{removeDuplicateSubjects}{Remove subjects that are in both the target and comparator cohort? See details for allowed values.Note that this is typically done in the createStudyPopulation function, but can already be done here for efficiency reasons.}

\item{restrictToCommonPeriod}{Restrict the analysis to the period when both treatments are observed?}

\item{washoutPeriod}{The minimum required continuous observation time prior to index date for a person to be included in the cohort. Note that this is typically done in the createStudyPopulation function, but can already be done here for efficiency reasons.}

\item{maxCohortSize}{If either the target or the comparator cohort is larger than this number it will be sampled to this size. maxCohortSize = 0 indicates no maximum size.}

\item{covariateSettings}{An object of type covariateSettings as created using the FeatureExtraction::createCovariateSettings() function.}
}
\description{
Create a parameter object for the function getDbCohortMethodData
}
\details{
Create an object defining the parameter values.
}
