% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettingsObjects.R
\name{createGetDbCohortMethodDataArgs}
\alias{createGetDbCohortMethodDataArgs}
\title{Create a parameter object for the function \code{\link[=getDbCohortMethodData]{getDbCohortMethodData()}}}
\usage{
createGetDbCohortMethodDataArgs(
  removeDuplicateSubjects = "keep first, truncate to second",
  firstExposureOnly = TRUE,
  washoutPeriod = 365,
  nestingCohortId = NULL,
  restrictToCommonPeriod = TRUE,
  minAge = NULL,
  maxAge = NULL,
  genderConceptIds = NULL,
  studyStartDate = "",
  studyEndDate = "",
  maxCohortSize = 0,
  covariateSettings
)
}
\arguments{
\item{removeDuplicateSubjects}{Remove subjects that are in both the target and comparator cohort? See details for allowed values.Note that this is typically done in the createStudyPopulation function, but can already be done here for efficiency reasons.}

\item{firstExposureOnly}{Should only the first exposure per subject be included? Note that this is typically done in the createStudyPopulation() function, but can already be done here for efficiency reasons.}

\item{washoutPeriod}{The minimum required continuous observation time prior to index date for a person to be included in the cohort. Note that this is typically done in the createStudyPopulation function, but can already be done here for efficiency reasons.}

\item{nestingCohortId}{A cohort definition ID identifying the records in the \code{nestingCohortTable} to use as nesting cohort.}

\item{restrictToCommonPeriod}{Restrict the analysis to the period when both treatments are observed?}

\item{minAge}{Minimum age at index date at which patient time will be included in the analysis. If not specified, no minimum age restriction will be applied.}

\item{maxAge}{Maximum age at index date at which patient time will be included in the analysis. If not specified, no maximum age restriction will be applied.}

\item{genderConceptIds}{Set of gender concept IDs to restrict the population to. If not specified, no restriction on gender will be applied.}

\item{studyStartDate}{A calendar date specifying the minimum date that a cohort index date can appear. Date format is 'yyyymmdd'.}

\item{studyEndDate}{A calendar date specifying the maximum date that a cohort index date can appear. Date format is 'yyyymmdd'. Important: the study end data is also used to truncate risk windows, meaning no outcomes beyond the study end date will be considered.}

\item{maxCohortSize}{If either the target or the comparator cohort is larger than this number it will be sampled to this size. maxCohortSize = 0 indicates no maximum size.}

\item{covariateSettings}{An object of type covariateSettings as created using the FeatureExtraction::createCovariateSettings() function, or a list of covariate settings objects.}
}
\value{
An object of type \code{GetDbCohortMethodDataArgs}.
}
\description{
Create a parameter object for the function \code{\link[=getDbCohortMethodData]{getDbCohortMethodData()}}
}
\details{
Create an object defining the parameter values.

The \code{removeduplicateSubjects} argument can have one of the following values:
\itemize{
\item \code{"keep first, truncate to second"}: When a subjects appear in both target and comparator cohort, only keep whichever cohort is first in time. If the other cohort starts before the first has ended, the first cohort will be truncated to stop the day before the second starts. If both cohorts start simultaneous, the person is removed from the analysis.
\item \code{"keep first"}: When a subjects appear in both target and comparator cohort, only keep whichever cohort is first in time. If both cohorts start simultaneous, the person is removed from the analysis.
\item \code{"remove all"}: Remove subjects that appear in both target and comparator cohort completely from the analysis."
\item \code{"keep all"}: Do not remove subjects that appear in both target and comparator cohort
}
}
