% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createGetDbCohortMethodDataArgs}
\alias{createGetDbCohortMethodDataArgs}
\title{Create a parameter object for the function getDbCohortMethodData}
\usage{
createGetDbCohortMethodDataArgs(studyStartDate = "", studyEndDate = "",
  excludeDrugsFromCovariates = TRUE, firstExposureOnly = FALSE,
  removeDuplicateSubjects = FALSE, washoutPeriod = 0, covariateSettings)
}
\arguments{
\item{studyStartDate}{A calendar date specifying the minimum date that a cohort
indexdate can appear. Date format is 'yyyymmdd'.}

\item{studyEndDate}{A calendar date specifying the maximum date that a cohort
indexdate can appear. Date format is 'yyyymmdd'. Important: the
studyend data is also used to truncate risk windows, meaning no
outcomesbeyond the study end date will be considered.}

\item{excludeDrugsFromCovariates}{Should the target and comparator drugs (and their
descendantconcepts) be excluded from the covariates? Note that
this willwork if the drugs are actualy drug concept IDs (and not
cohortIDs).}

\item{firstExposureOnly}{Should only the first exposure per subject be included? Note
thatthis is typically done in the createStudyPopulation
function,but can already be done here for efficiency reasons.}

\item{removeDuplicateSubjects}{Remove subjects that are in both the treated and comparator
cohort? Note thatthis is typically done in the
createStudyPopulation function,but can already be done here for
efficiency reasons.}

\item{washoutPeriod}{The mininum required continuous observation time prior to
indexdate for a person to be included in the cohort. Note
thatthis is typically done in the createStudyPopulation
function,but can already be done here for efficiency reasons.}

\item{covariateSettings}{An object of type covariateSettings as created using
thecreateCovariateSettings function in theFeatureExtraction
package.}
}
\description{
Create a parameter object for the function getDbCohortMethodData
}
\details{
Create an object defining the parameter values.
}
