% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createMatchOnPsAndCovariatesArgs}
\alias{createMatchOnPsAndCovariatesArgs}
\title{Create a parameter object for the function matchOnPsAndCovariates}
\usage{
createMatchOnPsAndCovariatesArgs(caliper = 0.25,
  caliperScale = "standardized", maxRatio = 1, covariateIds)
}
\arguments{
\item{caliper}{The caliper for matching. A caliper is the distance which is acceptablefor any
match. Observations which are outside of the caliper are dropped.A caliper of
0 means no caliper is used.}

\item{caliperScale}{The scale on which the caliper is defined. Two scales
aresupported:caliperScale = 'propensity score' orcaliperScale =
'standardized'. On the standardized scale, thecaliper is interpreted in
standard deviations of the propensity scoredistribution.}

\item{maxRatio}{The maximum number of persons int the comparator arm to be matched to
eachperson in the treatment arm. A maxRatio of 0 means no maximum:
allcomparators will be assigned to a treated person.}

\item{covariateIds}{One or more covariate IDs in the cohortMethodData object on whichsubjects
should be also matched.}
}
\description{
Create a parameter object for the function matchOnPsAndCovariates
}
\details{
Create an object defining the parameter values.
}

