% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createMatchOnPsAndCovariatesArgs}
\alias{createMatchOnPsAndCovariatesArgs}
\title{Create a parameter object for the function matchOnPsAndCovariates}
\usage{
createMatchOnPsAndCovariatesArgs(caliper = 0.2,
  caliperScale = "standardized logit", maxRatio = 1, covariateIds)
}
\arguments{
\item{caliper}{The caliper for matching. A caliper is the distance which is acceptablefor any
match. Observations which are outside of the caliper are dropped.A caliper of
0 means no caliper is used.}

\item{caliperScale}{The scale on which the caliper is defined. Three scales are
supported:caliperScale = 'propensity score', caliperScale ='standardized', or
caliperScale = 'standardized logit'.On the standardized scale, the caliper is
interpreted in standarddeviations of the propensity score distribution.
'standardized logit'is similar, except that the propensity score is
transformed to the logitscale because the PS is more likely to be normally
distributed on that scale(Austin, 2011).}

\item{maxRatio}{The maximum number of persons int the comparator arm to be matched to
eachperson in the treatment arm. A maxRatio of 0 means no maximum:
allcomparators will be assigned to a target person.}

\item{covariateIds}{One or more covariate IDs in the cohortMethodData object on whichsubjects
should be also matched.}
}
\description{
Create a parameter object for the function matchOnPsAndCovariates
}
\details{
Create an object defining the parameter values.
}
