% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{createOutcome}
\alias{createOutcome}
\title{Create outcome definition}
\usage{
createOutcome(
  outcomeId,
  outcomeOfInterest = TRUE,
  trueEffectSize = NA,
  priorOutcomeLookback = NULL,
  riskWindowStart = NULL,
  startAnchor = NULL,
  riskWindowEnd = NULL,
  endAnchor = NULL
)
}
\arguments{
\item{outcomeId}{An integer used to identify the outcome in the outcome cohort table.}

\item{outcomeOfInterest}{Is this an outcome of interest? If not, creation of non-essential
files will be skipped, including outcome=specific covariate balance
files. This could be helpful to speed up analyses with many controls,
for which we're only interested in the effect size estimate.}

\item{trueEffectSize}{For negative and positive controls: the known true effect size. To be used
for empirical calibration. Negative controls have \code{trueEffectSize = 1}. If
the true effect size is unknown, use \code{trueEffectSize = NA}}

\item{priorOutcomeLookback}{How many days should we look back when identifying prior.
outcomes?}

\item{riskWindowStart}{The start of the risk window (in days) relative to the \code{startAnchor}.}

\item{startAnchor}{The anchor point for the start of the risk window. Can be \code{"cohort start"}
or \code{"cohort end"}.}

\item{riskWindowEnd}{The end of the risk window (in days) relative to the \code{endAnchor}.}

\item{endAnchor}{The anchor point for the end of the risk window. Can be \code{"cohort start"}
or \code{"cohort end"}.}
}
\value{
An object of type \code{outcome}, to be used in \code{\link[=createTargetComparatorOutcomes]{createTargetComparatorOutcomes()}}.
}
\description{
Create outcome definition
}
\details{
Any settings here that are not \code{NULL} will override any values set in \code{\link[=createCreateStudyPopulationArgs]{createCreateStudyPopulationArgs()}}.
}
