% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsFunctions.R
\name{createPs}
\alias{createPs}
\title{Create propensity scores}
\usage{
createPs(cohortMethodData, excludePriorOutcome = TRUE, outcomeId = NULL,
  excludeCovariateIds = NULL, stopOnHighCorrelation = TRUE,
  prior = createPrior("laplace", exclude = c(0), useCrossValidation = TRUE),
  control = createControl(noiseLevel = "silent", cvType = "auto",
  startingVariance = 0.1))
}
\arguments{
\item{cohortMethodData}{An object of type \code{cohortMethodData} as generated using
\code{getDbCohortMethodData}.}

\item{excludePriorOutcome}{Remove people that have the outcome prior to the index date?}

\item{outcomeId}{The concept ID of the outcome. Persons marked for removal for the
outcome will be removed prior to creating the propensity score model.}

\item{excludeCovariateIds}{Exclude these covariates from the propensity model.}

\item{stopOnHighCorrelation}{If true, the function will test each covariate for correlation with the
treatment assignment. If any covariate has an unusually high correlation
(either positive or negative), this will be reported and the function
will stop.}

\item{prior}{The prior used to fit the model. See \code{\link[Cyclops]{createPrior}}
for details.}

\item{control}{The control object used to control the cross-validation used to
determine the hyperparameters of the prior (if applicable). See
\code{\link[Cyclops]{createControl}} for details.}
}
\description{
\code{createPs} creates propensity scores using a regularized logistic regression.
}
\details{
\code{createPs} creates propensity scores using a regularized logistic regression.
}
\examples{
data(cohortMethodDataSimulationProfile)
cohortMethodData <- simulateCohortMethodData(cohortMethodDataSimulationProfile, n = 1000)
ps <- createPs(cohortMethodData)

}

