% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createPs}
\alias{createPs}
\title{Create propensity scores}
\usage{
createPs(cohortData, outcomeConceptId = NULL, prior = createPrior("laplace",
  useCrossValidation = TRUE), control = createControl(lowerLimit = 0.01,
  upperLimit = 10, fold = 5, noiseLevel = "quiet"))
}
\arguments{
\item{cohortData}{An object of type \code{cohortData} as generated using \code{getDbCohortData}.}

\item{outcomeConceptId}{The concept ID of the outcome. Persons marked for removal for the outcome will be removed prior to
creating the propensity score model.}

\item{prior}{The prior used to fit the model. See \code{?createPrior} in the Cyclops package for details.}

\item{control}{The control object used to control the cross-validation used to determine the
hyperparameters of the prior (if applicable). See \code{?createControl} in the Cyclops package for details.}
}
\description{
\code{createPs} creates propensity scores using a regularized logistic regression.
}
\details{
\code{createPs} creates propensity scores using a regularized logistic regression.
}
\examples{
#todo
}

