% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsFunctions.R
\name{createPs}
\alias{createPs}
\title{Create propensity scores}
\usage{
createPs(cohortMethodData, population, excludeCovariateIds = c(),
  includeCovariateIds = c(), errorOnHighCorrelation = TRUE,
  stopOnError = TRUE, prior = createPrior("laplace", exclude = c(0),
  useCrossValidation = TRUE), control = createControl(noiseLevel = "silent",
  cvType = "auto", tolerance = 2e-07, cvRepetitions = 10, startingVariance =
  0.01))
}
\arguments{
\item{cohortMethodData}{An object of type \code{cohortMethodData} as generated using
\code{getDbCohortMethodData}.}

\item{population}{A data frame describing the population. This should at least have a 'rowId' column
corresponding to the rowId column in the \code{cohortMethodData} covariates object and a 'treatment' column.
If population is not specified, the full population in the \code{cohortMethodData} will be used.}

\item{excludeCovariateIds}{Exclude these covariates from the propensity model.}

\item{includeCovariateIds}{Include only these covariates in the propensity model.}

\item{errorOnHighCorrelation}{If true, the function will test each covariate for correlation with the
treatment assignment. If any covariate has an unusually high correlation
(either positive or negative), this will throw and error.}

\item{stopOnError}{If an error occurrs, should the function stop? Else, the two cohorts
will be assumed to be perfectly separable.}

\item{prior}{The prior used to fit the model. See \code{\link[Cyclops]{createPrior}}
for details.}

\item{control}{The control object used to control the cross-validation used to
determine the hyperparameters of the prior (if applicable). See
\code{\link[Cyclops]{createControl}} for details.}
}
\description{
\code{createPs} creates propensity scores using a regularized logistic regression.
}
\details{
\code{createPs} creates propensity scores using a regularized logistic regression.
}
\examples{
data(cohortMethodDataSimulationProfile)
cohortMethodData <- simulateCohortMethodData(cohortMethodDataSimulationProfile, n = 1000)
ps <- createPs(cohortMethodData)

}

