% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createStratifyByPsAndCovariatesArgs}
\alias{createStratifyByPsAndCovariatesArgs}
\title{Create a parameter object for the function stratifyByPsAndCovariates}
\usage{
createStratifyByPsAndCovariatesArgs(numberOfStrata = 5, covariateIds)
}
\arguments{
\item{numberOfStrata}{Into how many strata should the propensity score be divided? Theboundaries
of the strata are automatically defined to contain equalnumbers of treated
persons.}

\item{covariateIds}{One or more covariate IDs in the cohortMethodData object on whichsubjects
should also be stratified.}
}
\description{
Create a parameter object for the function stratifyByPsAndCovariates
}
\details{
Create an object defining the parameter values.
}

