% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createStratifyByPsAndCovariatesArgs}
\alias{createStratifyByPsAndCovariatesArgs}
\title{Create a parameter object for the function stratifyByPsAndCovariates}
\usage{
createStratifyByPsAndCovariatesArgs(
  numberOfStrata = 5,
  baseSelection = "all",
  covariateIds
)
}
\arguments{
\item{numberOfStrata}{Into how many strata should the propensity score be divided? The boundaries of the strata are automatically defined to contain equal numbers of target persons.}

\item{baseSelection}{What is the base selection of subjects where the strata bounds are to be determined? Strata are defined as equally-sized strata inside this selection. Possible values are "all", "target", and "comparator".}

\item{covariateIds}{One or more covariate IDs in the cohortMethodData object on which subjects should also be stratified.}
}
\description{
Create a parameter object for the function stratifyByPsAndCovariates
}
\details{
Create an object defining the parameter values.
}
