% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettingsObjects.R
\name{createStratifyByPsArgs}
\alias{createStratifyByPsArgs}
\title{Create a parameter object for the function \code{\link[=stratifyByPs]{stratifyByPs()}}}
\usage{
createStratifyByPsArgs(
  numberOfStrata = 10,
  baseSelection = "all",
  stratificationColumns = c(),
  stratificationCovariateIds = c()
)
}
\arguments{
\item{numberOfStrata}{How many strata? The boundaries of the strata are automatically
defined to contain equal numbers of target persons.}

\item{baseSelection}{What is the base selection of subjects where the strata bounds are
to be determined? Strata are defined as equally-sized strata inside
this selection. Possible values are "all", "target", and "comparator".}

\item{stratificationColumns}{Names or numbers of one or more columns in the \code{data} data.frame
on which subjects should be stratified prior to matching. No persons
will be matched with persons outside of the strata identified by the
values in these columns.}

\item{stratificationCovariateIds}{One or more covariate IDs in the \code{cohortMethodData} object on which
subjects should also be stratified.}
}
\value{
An object of type \code{StratifyByPsArgs}.
}
\description{
Create a parameter object for the function \code{\link[=stratifyByPs]{stratifyByPs()}}
}
\details{
Create an object defining the parameter values.
}
