% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createStratifyByPsArgs}
\alias{createStratifyByPsArgs}
\title{Create a parameter object for the function stratifyByPs}
\usage{
createStratifyByPsArgs(numberOfStrata = 5, stratificationColumns = c())
}
\arguments{
\item{numberOfStrata}{How many strata? The boundaries of the strata are
automaticallydefined to contain equal numbers of treated persons.}

\item{stratificationColumns}{Names of one or more columns in the data data.frame on whichsubjects
should also be stratified in addition to stratification onpropensity
score.}
}
\description{
Create a parameter object for the function stratifyByPs
}
\details{
Create an object defining the parameter values.
}

