% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyPopulation.R
\name{createStudyPopulation}
\alias{createStudyPopulation}
\title{Create a study population}
\usage{
createStudyPopulation(
  cohortMethodData,
  population = NULL,
  outcomeId,
  riskId = NULL,
  firstExposureOnly = FALSE,
  restrictToCommonPeriod = FALSE,
  washoutPeriod = 0,
  removeDuplicateSubjects = FALSE,
  removeSubjectsWithPriorOutcome = TRUE,
  priorOutcomeLookback = 99999,
  minDaysAtRisk = 1,
  riskWindowStart = 0,
  addExposureDaysToStart = NULL,
  startAnchor = "cohort start",
  riskWindowEnd = 0,
  addExposureDaysToEnd = NULL,
  endAnchor = "cohort end",
  censorAtNewRiskWindow = FALSE
)
}
\arguments{
\item{cohortMethodData}{An object of type \link{CohortMethodData} as generated using
\code{\link[=getDbCohortMethodData]{getDbCohortMethodData()}}.}

\item{population}{If specified, this population will be used as the starting
point instead of the cohorts in the \code{cohortMethodData} object.}

\item{outcomeId}{The ID of the outcome. If not specified, no outcome-specific
transformations will be performed.}

\item{riskId}{The ID of the competing risk outcome. If not specified, no competing-risk-specific
transformations will be performed.}

\item{firstExposureOnly}{Should only the first exposure per subject be included?}

\item{restrictToCommonPeriod}{Restrict the analysis to the period when both exposures are observed?}

\item{washoutPeriod}{The minimum required continuous observation time prior to
index date for a person to be included in the cohort.}

\item{removeDuplicateSubjects}{Remove subjects that are in both the target and comparator
cohort? See details for allowed values.}

\item{removeSubjectsWithPriorOutcome}{Remove subjects that have the outcome prior to the risk
window start?}

\item{priorOutcomeLookback}{How many days should we look back when identifying prior
outcomes?}

\item{minDaysAtRisk}{The minimum required number of days at risk.}

\item{riskWindowStart}{The start of the risk window (in days) relative to the \code{startAnchor}.}

\item{addExposureDaysToStart}{DEPRECATED: Add the length of exposure the start of the risk window?
Use \code{startAnchor} instead.}

\item{startAnchor}{The anchor point for the start of the risk window. Can be \code{"cohort start"}
or \code{"cohort end"}.}

\item{riskWindowEnd}{The end of the risk window (in days) relative to the \code{endAnchor}.}

\item{addExposureDaysToEnd}{DEPRECATED: Add the length of exposure the risk window?
Use \code{endAnchor} instead.}

\item{endAnchor}{The anchor point for the end of the risk window. Can be \code{"cohort start"}
or \code{"cohort end"}.}

\item{censorAtNewRiskWindow}{If a subject is in multiple cohorts, should time-at-risk be censored
when the new time-at-risk starts to prevent overlap?}
}
\value{
A \code{tibble} specifying the study population. This \code{tibble} will have the following columns:
\itemize{
\item \code{rowId}: A unique identifier for an exposure.
\item \code{subjectId}: The person ID of the subject.
\item \code{cohortStartdate}: The index date.
\item \code{outcomeCount} The number of outcomes observed during the risk window.
\item \code{timeAtRisk}: The number of days in the risk window.
\item \code{survivalTime}: The number of days until either the outcome or the end of the risk window.
}
}
\description{
Create a study population
}
\details{
Create a study population by enforcing certain inclusion and exclusion criteria, defining a risk
window, and determining which outcomes fall inside the risk window.

The \code{removeduplicateSubjects} argument can have one of the following values:
\itemize{
\item \code{"keep all"}: Do not remove subjects that appear in both target and comparator cohort
\item \code{"keep first"}: When a subjects appear in both target and comparator cohort, only keep whichever cohort is first in time. If both cohorts start simultaneous, the person is removed from the analysis.
\item \code{"remove all"}: Remove subjects that appear in both target and comparator cohort completely from the analysis."
}
}
