% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{createTargetComparatorOutcomes}
\alias{createTargetComparatorOutcomes}
\title{Create target-comparator-outcomes combinations.}
\usage{
createTargetComparatorOutcomes(
  targetId,
  comparatorId,
  outcomes,
  excludedCovariateConceptIds = c(),
  includedCovariateConceptIds = c()
)
}
\arguments{
\item{targetId}{A cohort ID identifying the target exposure in the exposure
table.}

\item{comparatorId}{A cohort ID identifying the comparator exposure in the exposure
table.}

\item{outcomes}{A list of object of type \code{outcome} as created by
\code{\link[=createOutcome]{createOutcome()}}.}

\item{excludedCovariateConceptIds}{A list of concept IDs that cannot be used to construct
covariates. This argument is to be used only for exclusion
concepts that are specific to the target-comparator combination.}

\item{includedCovariateConceptIds}{A list of concept IDs that must be used to construct
covariates. This argument is to be used only for inclusion
concepts that are specific to the target-comparator combination.}
}
\value{
An object of type \code{targetComparatorOutcomes}.
}
\description{
Create target-comparator-outcomes combinations.
}
\details{
Create a set of hypotheses of interest, to be used with the \code{\link[=runCmAnalyses]{runCmAnalyses()}} function.
}
