% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{createTargetComparatorOutcomes}
\alias{createTargetComparatorOutcomes}
\title{Create target-comparator-outcomes combinations.}
\usage{
createTargetComparatorOutcomes(targetId, comparatorId, outcomeIds,
  excludedCovariateConceptIds = c(), includedCovariateConceptIds = c())
}
\arguments{
\item{targetId}{A concept ID indentifying the target drug in the exposure
table. If multiple strategies for picking the target will be
tested in the analysis, a named list of numbers can be provided
instead. In the analysis, the name of the number to be used can
be specified using the #' \code{targetType} parameter in the
\code{\link{createCmAnalysis}} function.}

\item{comparatorId}{A concept ID indentifying the comparator drug in the exposure
table. If multiple strategies for picking the comparator will
be tested in the analysis, a named list of numbers can be
provided instead. In the analysis, the name of the number to be
used can be specified using the #' \code{comparatorType}
parameter in the \code{\link{createCmAnalysis}} function.}

\item{outcomeIds}{A vector of concept IDs indentifying the outcome(s) in the
outcome table.}

\item{excludedCovariateConceptIds}{A list of concept IDs that cannot be used to construct
covariates. This argument is to be used only for exclusion
concepts that are specific to the drug-comparator combination.}

\item{includedCovariateConceptIds}{A list of concept IDs that must be used to construct
covariates. This argument is to be used only for inclusion
concepts that are specific to the drug-comparator combination.}
}
\description{
Create target-comparator-outcomes combinations.
}
\details{
Create a set of hypotheses of interest, to be used with the \code{\link{runCmAnalyses}} function.
}
