% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createTrimByPsArgs}
\alias{createTrimByPsArgs}
\title{Create a parameter object for the function trimByPs}
\usage{
createTrimByPsArgs(trimFraction = 0.05)
}
\arguments{
\item{trimFraction}{This fraction will be removed from each treatment group. In the targetgroup,
persons with the highest propensity scores will be removed, in thecomparator
group person with the lowest scores will be removed.}
}
\description{
Create a parameter object for the function trimByPs
}
\details{
Create an object defining the parameter values.
}
