% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createTruncateIptwArgs}
\alias{createTruncateIptwArgs}
\title{Create a parameter object for the function truncateIptw}
\usage{
createTruncateIptwArgs(maxWeight = 10)
}
\arguments{
\item{maxWeight}{The maximum allowed IPTW.}
}
\description{
Create a parameter object for the function truncateIptw
}
\details{
Create an object defining the parameter values.
}
