% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutcomeModels.R
\name{fitOutcomeModel}
\alias{fitOutcomeModel}
\title{Create an outcome model, and compute the relative risk}
\usage{
fitOutcomeModel(
  population,
  cohortMethodData = NULL,
  fitOutcomeModelArgs = createFitOutcomeModelArgs()
)
}
\arguments{
\item{population}{A population object generated by \code{\link[=createStudyPopulation]{createStudyPopulation()}},
potentially filtered by other functions.}

\item{cohortMethodData}{An object of type \link{CohortMethodData} as generated using
\code{\link[=getDbCohortMethodData]{getDbCohortMethodData()}}. Can be omitted if not using covariates and
not using interaction terms.}

\item{fitOutcomeModelArgs}{An object of type \code{FitOutcomeModelArgs} as generated using the
\code{\link[=createFitOutcomeModelArgs]{createFitOutcomeModelArgs()}} function.}
}
\value{
An object of class \code{OutcomeModel}. Generic function \code{print}, \code{coef}, and
\code{confint} are available.
}
\description{
Create an outcome model, and computes the relative risk
}
\details{
For likelihood profiling, either specify the \code{profileGrid} for a completely user- defined grid, or
\code{profileBounds} for an adaptive grid. Both should be defined on the log effect size scale. When both
\code{profileGrid} and \code{profileGrid} are \code{NULL} likelihood profiling is disabled.
}
