% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fitOutcomeModel}
\alias{fitOutcomeModel}
\title{Create an outcome model, and compute the relative risk}
\usage{
fitOutcomeModel(outcomeConceptId, cohortData, strata = NULL,
  riskWindowStart = 0, riskWindowEnd = 9999, addExposureDaysToEnd = FALSE,
  useCovariates = TRUE, fitModel = TRUE, modelType = "cox",
  prior = createPrior("laplace", useCrossValidation = TRUE),
  control = createControl(lowerLimit = 0.01, upperLimit = 10, fold = 5,
  noiseLevel = "quiet"))
}
\arguments{
\item{cohortData}{An object of type \code{cohortData} as generated using \code{getDbCohortData}.}

\item{strata}{A data frame specifying the (matched and/or trimmed) subpopulation to be
used in the study, as well as their strata (for conditional models). This data frame should have at
least a \code{RowId}, and a \code{StratumId} when including stratification.}

\item{riskWindowEnd}{The maximum length (in days) of the risk window.}

\item{useCovariates}{Whether to use the covariate matrix in the cohortData in the outcome model.}

\item{fitModel}{If false, the model will not be fit, and only summary statistics are available.}

\item{modelType}{The type of model to be fitted. See details for options.}

\item{prior}{The prior used to fit the model. See \code{?prior} for details.}
}
\value{
An object of class \code{outcomeModel}. Generic function \code{summary}, \code{coef}, and \code{confint}
are available.
}
\description{
\code{fitOutcomeModel} creates an outcome model, and computes the relative risk
}
\details{
The model type can be one of these:
\tabular{ll}{
  \verb{lr}      \tab Logistic regression  \cr
  \verb{clr}     \tab Conditional logistic regression \cr
  \verb{cox}     \tab Cox regression (stratified or not, depending on whether \code{stata} is specified) \cr
  \verb{pr}      \tab Poisson regression  \cr
  \verb{cpr}     \tab Conditional Poisson regression \cr
}
}
\examples{
#todo
}

