% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutcomeModels.R
\name{fitOutcomeModel}
\alias{fitOutcomeModel}
\title{Create an outcome model, and compute the relative risk}
\usage{
fitOutcomeModel(
  population,
  cohortMethodData = NULL,
  modelType = "logistic",
  stratified = FALSE,
  useCovariates = FALSE,
  inversePtWeighting = FALSE,
  estimator = "ate",
  interactionCovariateIds = c(),
  excludeCovariateIds = c(),
  includeCovariateIds = c(),
  profileGrid = seq(log(0.1), log(10), length.out = 1000),
  prior = createPrior("laplace", useCrossValidation = TRUE),
  control = createControl(cvType = "auto", seed = 1, startingVariance = 0.01, tolerance
    = 2e-07, cvRepetitions = 10, noiseLevel = "quiet")
)
}
\arguments{
\item{population}{A population object generated by \code{\link[=createStudyPopulation]{createStudyPopulation()}},
potentially filtered by other functions.}

\item{cohortMethodData}{An object of type \link{CohortMethodData} as generated using
\code{\link[=getDbCohortMethodData]{getDbCohortMethodData()}}. Can be omitted if not using covariates and
not using interaction terms.}

\item{modelType}{The type of outcome model that will be used. Possible values are
"logistic", "poisson", or "cox".}

\item{stratified}{Should the regression be conditioned on the strata defined in the
population object (e.g. by matching or stratifying on propensity
scores)?}

\item{useCovariates}{Whether to use the covariates in the \code{cohortMethodData}
object in the outcome model.}

\item{inversePtWeighting}{Use inverse probability of treatment weighting (IPTW)? See details.}

\item{estimator}{for IPTW: the type of estimator. Options are \code{estimator = "ate"} for the
average treatment effect, and \code{estimator = "att"}for the average treatment
effect in the treated.}

\item{interactionCovariateIds}{An optional vector of covariate IDs to use to estimate interactions
with the main treatment effect.}

\item{excludeCovariateIds}{Exclude these covariates from the outcome model.}

\item{includeCovariateIds}{Include only these covariates in the outcome model.}

\item{profileGrid}{A one-dimensional grid of points on the log(relative risk) scale where
the likelihood for the treatment variable coefficient is sampled. Set to
NULL to skip profiling.}

\item{prior}{The prior used to fit the model. See \code{\link[Cyclops:createPrior]{Cyclops::createPrior()}}
for details.}

\item{control}{The control object used to control the cross-validation used to
determine the hyperparameters of the prior (if applicable). See
\code{\link[Cyclops:createControl]{Cyclops::createControl()}} for details.}
}
\value{
An object of class \code{OutcomeModel}. Generic function \code{print}, \code{coef}, and
\code{confint} are available.
}
\description{
Create an outcome model, and computes the relative risk
}
\details{
IPTW estimates either the average treatment effect (ate) or average treatment effect in the treated
(att) using stabilized inverse propensity scores (Xu et al. 2010).
}
\references{
Xu S, Ross C, Raebel MA, Shetterly S, Blanchette C, Smith D. Use of stabilized inverse propensity scores
as weights to directly estimate relative risk and its confidence intervals. Value Health.
2010;13(2):273-277. doi:10.1111/j.1524-4733.2009.00671.x
}
