% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutcomeModels.R
\name{fitOutcomeModel}
\alias{fitOutcomeModel}
\title{Create an outcome model, and compute the relative risk}
\usage{
fitOutcomeModel(
  population,
  cohortMethodData = NULL,
  modelType = "logistic",
  stratified = FALSE,
  useCovariates = FALSE,
  inversePtWeighting = FALSE,
  bootstrapCi = FALSE,
  interactionCovariateIds = c(),
  excludeCovariateIds = c(),
  includeCovariateIds = c(),
  profileGrid = NULL,
  profileBounds = c(log(0.1), log(10)),
  prior = createPrior("laplace", useCrossValidation = TRUE),
  control = createControl(cvType = "auto", seed = 1, resetCoefficients = TRUE,
    startingVariance = 0.01, tolerance = 2e-07, cvRepetitions = 10, noiseLevel = "quiet")
)
}
\arguments{
\item{population}{A population object generated by \code{\link[=createStudyPopulation]{createStudyPopulation()}},
potentially filtered by other functions.}

\item{cohortMethodData}{An object of type \link{CohortMethodData} as generated using
\code{\link[=getDbCohortMethodData]{getDbCohortMethodData()}}. Can be omitted if not using covariates and
not using interaction terms.}

\item{modelType}{The type of outcome model that will be used. Possible values are
"logistic", "poisson", or "cox".}

\item{stratified}{Should the regression be conditioned on the strata defined in the
population object (e.g. by matching or stratifying on propensity
scores)?}

\item{useCovariates}{Whether to use the covariates in the \code{cohortMethodData}
object in the outcome model.}

\item{inversePtWeighting}{Use inverse probability of treatment weighting (IPTW)}

\item{bootstrapCi}{Compute confidence interval using bootstrapping instead of likelihood profiling?}

\item{interactionCovariateIds}{An optional vector of covariate IDs to use to estimate interactions
with the main treatment effect.}

\item{excludeCovariateIds}{Exclude these covariates from the outcome model.}

\item{includeCovariateIds}{Include only these covariates in the outcome model.}

\item{profileGrid}{A one-dimensional grid of points on the log(relative risk) scale where
the likelihood for coefficient of variables is sampled. See details.}

\item{profileBounds}{The bounds (on the log relative risk scale) for the adaptive sampling
of the likelihood function. See details.}

\item{prior}{The prior used to fit the model. See \code{\link[Cyclops:createPrior]{Cyclops::createPrior()}}
for details.}

\item{control}{The control object used to control the cross-validation used to
determine the hyperparameters of the prior (if applicable). See
\code{\link[Cyclops:createControl]{Cyclops::createControl()}} for details.}
}
\value{
An object of class \code{OutcomeModel}. Generic function \code{print}, \code{coef}, and
\code{confint} are available.
}
\description{
Create an outcome model, and computes the relative risk
}
\details{
For likelihood profiling, either specify the \code{profileGrid} for a completely user- defined grid, or
\code{profileBounds} for an adaptive grid. Both should be defined on the log effect size scale. When both
\code{profileGrid} and \code{profileGrid} are \code{NULL} likelihood profiling is disabled.
}
