% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutcomeModels.R
\name{fitOutcomeModel}
\alias{fitOutcomeModel}
\title{Create an outcome model, and compute the relative risk}
\usage{
fitOutcomeModel(population, cohortMethodData = NULL,
  modelType = "logistic", stratified = FALSE, useCovariates = FALSE,
  inversePtWeighting = FALSE, interactionCovariateIds = c(),
  excludeCovariateIds = c(), includeCovariateIds = c(),
  prior = createPrior("laplace", useCrossValidation = TRUE),
  control = createControl(cvType = "auto", seed = 1, startingVariance =
  0.01, tolerance = 2e-07, cvRepetitions = 10, noiseLevel = "quiet"))
}
\arguments{
\item{population}{A population object generated by \code{createStudyPopulation},
potentially filtered by other functions.}

\item{cohortMethodData}{An object of type \code{cohortMethodData} as generated using
\code{getDbCohortMethodData}.}

\item{modelType}{The type of outcome model that will be used. Possible values are
"logistic", "poisson", or "cox".}

\item{stratified}{Should the regression be conditioned on the strata defined in the
population object (e.g. by matching or stratifying on propensity
scores)?}

\item{useCovariates}{Whether to use the covariate matrix in the \code{cohortMethodData}
object in the outcome model.}

\item{inversePtWeighting}{Use inverse probability of treatment weigting?}

\item{interactionCovariateIds}{An optional vector of covariate IDs to use to estimate interactions
with the main treatment effect.}

\item{excludeCovariateIds}{Exclude these covariates from the outcome model.}

\item{includeCovariateIds}{Include only these covariates in the outcome model.}

\item{prior}{The prior used to fit the model. See \code{\link[Cyclops]{createPrior}}
for details.}

\item{control}{The control object used to control the cross-validation used to
determine the hyperparameters of the prior (if applicable). See
\code{\link[Cyclops]{createControl}} for details.}
}
\value{
An object of class \code{outcomeModel}. Generic function \code{summary}, \code{coef}, and
\code{confint} are available.
}
\description{
\code{fitOutcomeModel} creates an outcome model, and computes the relative risk
}
