% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyPopulation.R
\name{getAttritionTable}
\alias{getAttritionTable}
\title{Get the attrition table for a population}
\usage{
getAttritionTable(population)
}
\arguments{
\item{population}{Either an object of type \code{cohortMethodData} or a population object generated by functions
like \code{createStudyPopulation}.}
}
\value{
A data frame specifiyng the number of people and exposures in the population after specific sets of filtering.
}
\description{
Get the attrition table for a population
}

