% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoadingSaving.R
\name{getDbCohortMethodData}
\alias{getDbCohortMethodData}
\title{Get the cohort data from the server}
\usage{
getDbCohortMethodData(
  connectionDetails,
  cdmDatabaseSchema,
  oracleTempSchema = cdmDatabaseSchema,
  targetId,
  comparatorId,
  outcomeIds,
  studyStartDate = "",
  studyEndDate = "",
  exposureDatabaseSchema = cdmDatabaseSchema,
  exposureTable = "drug_era",
  outcomeDatabaseSchema = cdmDatabaseSchema,
  outcomeTable = "condition_occurrence",
  cdmVersion = "5",
  excludeDrugsFromCovariates,
  firstExposureOnly = FALSE,
  removeDuplicateSubjects = FALSE,
  restrictToCommonPeriod = FALSE,
  washoutPeriod = 0,
  maxCohortSize = 0,
  covariateSettings
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the
\code{\link[DatabaseConnector:createConnectionDetails]{DatabaseConnector::createConnectionDetails()}} function.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance. Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{oracleTempSchema}{For Oracle only: the name of the database schema where you want
all temporary tables to be managed. Requires create/insert
permissions to this database.}

\item{targetId}{A unique identifier to define the target cohort.  If
exposureTable = DRUG_ERA, targetId is a concept ID and all
descendant concepts within that concept ID will be used to
define the cohort.  If exposureTable <> DRUG_ERA, targetId is
used to select the COHORT_DEFINITION_ID in the cohort-like table.}

\item{comparatorId}{A unique identifier to define the comparator cohort.  If
exposureTable = DRUG_ERA, comparatorId is a concpet ID and all
descendant concepts within that concept ID will be used to
define the cohort.  If exposureTable <> DRUG_ERA, comparatorId
is used to select the COHORT_DEFINITION_ID in the cohort-like
table.}

\item{outcomeIds}{A list of cohort IDs used to define outcomes.}

\item{studyStartDate}{A calendar date specifying the minimum date that a cohort index
date can appear. Date format is 'yyyymmdd'.}

\item{studyEndDate}{A calendar date specifying the maximum date that a cohort index
date can appear. Date format is 'yyyymmdd'. Important: the study
end data is also used to truncate risk windows, meaning no
outcomes beyond the study end date will be considered.}

\item{exposureDatabaseSchema}{The name of the database schema that is the location where the
exposure data used to define the exposure cohorts is available.}

\item{exposureTable}{The tablename that contains the exposure cohorts. If
exposureTable <> DRUG_ERA, then expectation is \code{exposureTable} has
format of COHORT table: COHORT_DEFINITION_ID, SUBJECT_ID,
COHORT_START_DATE, COHORT_END_DATE.}

\item{outcomeDatabaseSchema}{The name of the database schema that is the location where the
data used to define the outcome cohorts is available.}

\item{outcomeTable}{The tablename that contains the outcome cohorts.  If
outcomeTable <> CONDITION_OCCURRENCE, then expectation is
outcomeTable has format of COHORT table: COHORT_DEFINITION_ID,
SUBJECT_ID, COHORT_START_DATE, COHORT_END_DATE.}

\item{cdmVersion}{Define the OMOP CDM version used: currently supports "5".}

\item{excludeDrugsFromCovariates}{DEPRECATED: Should the target and comparator drugs (and their descendant
concepts) be excluded from the covariates? Note that this will
work if the drugs are actualy drug concept IDs (and not cohort
IDs).}

\item{firstExposureOnly}{Should only the first exposure per subject be included? Note
that this is typically done in the \code{\link[=createStudyPopulation]{createStudyPopulation()}}
function, but can already be done here for efficiency reasons.}

\item{removeDuplicateSubjects}{Remove subjects that are in both the target and comparator
cohort? See details for allowed values.N ote that this is typically done in the
\code{createStudyPopulation} function, but can already be done
here for efficiency reasons.}

\item{restrictToCommonPeriod}{Restrict the analysis to the period when both treatments are observed?}

\item{washoutPeriod}{The mininum required continuous observation time prior to index
date for a person to be included in the cohort. Note that this
is typically done in the \code{createStudyPopulation} function,
but can already be done here for efficiency reasons.}

\item{maxCohortSize}{If either the target or the comparator cohort is larger than
this number it will be sampled to this size. \code{maxCohortSize = 0}
indicates no maximum size.}

\item{covariateSettings}{An object of type \code{covariateSettings} as created using the
\code{\link[FeatureExtraction:createCovariateSettings]{FeatureExtraction::createCovariateSettings()}} function.}
}
\value{
A \link{CohortMethodData} object.
}
\description{
This function executes a large set of SQL statements against the database in OMOP CDM format to
extract the data needed to perform the analysis.
}
\details{
Based on the arguments, the treatment and comparator cohorts are retrieved, as well as outcomes
occurring in exposed subjects. The treatment and comparator cohorts can be identified using the
DRUG_ERA table, or through user-defined cohorts in a cohort table either inside the CDM schema or
in a separate schema. Similarly, outcomes are identified using the CONDITION_ERA table or through
user-defined cohorts in a cohort table either inside the CDM schema or in a separate schema.
Covariates are automatically extracted from the appropriate tables within the CDM.

\strong{Important}: The target and comparator drug must not be included in the covariates, including any descendant
concepts. You will need to manually add the drugs and descendants to the \code{excludedCovariateConceptIds}
of the \code{covariateSettings} argument.

The \code{removeduplicateSubjects} argument can have one of the following values:
\itemize{
\item \code{"keep all"}: Do not remove subjects that appear in both target and comparator cohort
\item \code{"keep first"}: When a subjects appear in both target and comparator cohort, only keep whichever cohort is first in time.
\item \code{"remove all"}: Remove subjects that appear in both target and comparator cohort completely from the analysis."
}
}
