% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{getDimensionTable}
\alias{getDimensionTable}
\title{Return data frame containing analysisId for dimensions}
\usage{
getDimensionTable(dimensions)
}
\arguments{
\item{dimensions}{names of data dimensions}
}
\value{
list of data frames containing the data dimension names and analysisId
}
\description{
Return data frame containing analysisId for dimensions
}

