% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalyses.R
\name{getFileReference}
\alias{getFileReference}
\title{Get a summary report of the analyses results}
\usage{
getFileReference(outputFolder)
}
\arguments{
\item{outputFolder}{Name of the folder where all the outputs have been written to.}
}
\value{
A tibble containing summary statistics for each target-comparator-outcome-analysis combination.
}
\description{
Get a summary report of the analyses results
}
