% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutcomeModels.R
\name{getOutcomeModel}
\alias{getOutcomeModel}
\title{Get the outcome model}
\usage{
getOutcomeModel(outcomeModel, cohortMethodData)
}
\arguments{
\item{outcomeModel}{An object of type \code{OutcomeModel} as generated using he
\code{\link[=fitOutcomeModel]{fitOutcomeModel()}} function.}

\item{cohortMethodData}{An object of type \link{CohortMethodData} as generated using
\code{\link[=getDbCohortMethodData]{getDbCohortMethodData()}}.}
}
\value{
A tibble.
}
\description{
Get the full outcome model, so showing the betas of all variables included
in the outcome model, not just the treatment variable.
}
