% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/OutcomeModels.R
\name{getOutcomeModel}
\alias{getOutcomeModel}
\title{Get the outcome model}
\usage{
getOutcomeModel(outcomeModel, cohortData)
}
\arguments{
\item{outcomeModel}{An object of type \code{outcomeModel} as generated using he \code{createOutcomeMode} function.}

\item{cohortData}{An object of type \code{cohortData} as generated using \code{getDbCohortData}.}
}
\description{
\code{getOutcomeModel} shows the full outcome model, so showing the betas of all variables
included in the outcome model, not just the treatment variable.
}
\details{
Shows the coefficients and names of the covariates with non-zero coefficients.
}
\examples{
#todo
}

