% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutcomeModels.R
\name{getOutcomeModel}
\alias{getOutcomeModel}
\title{Get the outcome model}
\usage{
getOutcomeModel(outcomeModel, cohortMethodData)
}
\arguments{
\item{outcomeModel}{An object of type \code{outcomeModel} as generated using he
\code{createOutcomeMode} function.}

\item{cohortMethodData}{An object of type \code{cohortMethodData} as generated using
\code{getDbCohortMethodData}.}
}
\description{
\code{getOutcomeModel} shows the full outcome model, so showing the betas of all variables included
in the outcome model, not just the treatment variable.
}
\details{
Shows the coefficients and names of the covariates with non-zero coefficients.
}
\examples{
# todo

}
