% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{getPredefinedCovariateRef}
\alias{getPredefinedCovariateRef}
\title{Get covariateRef for predefined codes}
\usage{
getPredefinedCovariateRef(cohortData, conceptIds, icd9, icd9AnalysisIds)
}
\arguments{
\item{cohortData}{CohortMethodData object}

\item{conceptIds}{List of conceptIds to include}

\item{icd9}{List of icd9 condition codes to include, with E replaced by 10 and V replaced by 11}

\item{icd9AnalysisIds}{List of analysisIds that include icd9 condition codes}
}
\value{
Returns entries of covariateRef that correspond to predefined codes to include
}
\description{
Get covariateRef entries for predefined codes to include
}

