% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{getPredefinedCovariates}
\alias{getPredefinedCovariates}
\title{Get predefined covariates}
\usage{
getPredefinedCovariates(cohortData, predefinedCovariateRef)
}
\arguments{
\item{cohortData}{CohortMethodData object}

\item{predefinedCovariateRef}{ffdf returned by \code{getPredefinedCovariateRef}}
}
\value{
Returns entries of covariates that correspond to predefined codes to include
}
\description{
Get predefined covariates entries for predefined codes to include
}

