% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsFunctions.R
\name{getPsModel}
\alias{getPsModel}
\title{Get the propensity model}
\usage{
getPsModel(propensityScore, cohortMethodData)
}
\arguments{
\item{propensityScore}{The propensity scores as generated using the \code{createPs} function.}

\item{cohortMethodData}{An object of type \code{cohortMethodData} as generated using
\code{getDbCohortMethodData}.}
}
\description{
\code{getPsModel} shows the propensity score model
}
\details{
Shows the coefficients and names of the covariates with non-zero coefficients.
}
\examples{
# todo

}
