% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoadingSaving.R
\name{loadCohortMethodData}
\alias{loadCohortMethodData}
\title{Load the cohort data from a folder}
\usage{
loadCohortMethodData(file, readOnly = TRUE, skipCovariates = FALSE)
}
\arguments{
\item{file}{The name of the folder containing the data.}

\item{readOnly}{If true, the data is opened read only.}

\item{skipCovariates}{Do not load the covariates. Can save a lot of time.}
}
\value{
An object of class cohortMethodData.
}
\description{
\code{loadCohortMethodData} loads an object of type cohortMethodData from a folder in the file
system.
}
\details{
The data will be written to a set of files in the folder specified by the user.
}
\examples{
# todo

}
