% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{loadTargetComparatorOutcomesList}
\alias{loadTargetComparatorOutcomesList}
\title{Load a list of targetComparatorOutcomes from file}
\usage{
loadTargetComparatorOutcomesList(file)
}
\arguments{
\item{file}{The name of the file}
}
\value{
A list of objects of type \code{targetComparatorOutcomes}.
}
\description{
Load a list of objects of type \code{targetComparatorOutcomes} from file. The file is in JSON format.
}
