% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Balance.R
\name{plotCovariateBalanceScatterPlot}
\alias{plotCovariateBalanceScatterPlot}
\title{Create a scatterplot of the covariate balance}
\usage{
plotCovariateBalanceScatterPlot(balance, absolute = TRUE,
  threshold = 0, title = "Standardized difference of mean",
  fileName = NULL, beforeLabel = "Before matching",
  afterLabel = "After matching", showCovariateCountLabel = FALSE,
  showMaxLabel = FALSE)
}
\arguments{
\item{balance}{A data frame created by the \code{computeCovariateBalance} funcion.}

\item{absolute}{Should the absolute value of the difference be used?}

\item{threshold}{Show a threshold value for after matching standardized difference.}

\item{title}{The main title for the plot.}

\item{fileName}{Name of the file where the plot should be saved, for example 'plot.png'. See the
function \code{ggsave} in the ggplot2 package for supported file formats.}

\item{beforeLabel}{Label for the x-axis.}

\item{afterLabel}{Label for the y-axis.}

\item{showCovariateCountLabel}{Show a label with the number of covariates included in the plot?}

\item{showMaxLabel}{Show a label with the maximum absolute standardized difference after matching/stratification?}
}
\value{
A ggplot object. Use the \code{\link[ggplot2]{ggsave}} function to save to file in a different
format.
}
\description{
Create a scatterplot of the covariate balance, showing all variables with balance before and after
matching on the x and y axis respectively. Requires running \code{computeCovariateBalance} first.
}
