% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotKaplanMeier}
\alias{plotKaplanMeier}
\title{Plot the Kaplan-Meier curve}
\usage{
plotKaplanMeier(outcomeModel, censorMarks = FALSE, legend = FALSE,
  labelsInGraph = TRUE, fileName = NULL)
}
\arguments{
\item{outcomeModel}{An object of type \code{outcomeModel} as generated using he \code{fitOutcomeModel} function.}

\item{censorMarks}{Whether or not to include censor marks in the plot.}

\item{legend}{Whether or not to include a legend in the plot.}

\item{labelsInGraph}{If true, the labels identifying the two curves will be added to the graph.}

\item{fileName}{Name of the file where the plot should be saved, for example 'plot.png'. See
the function \code{ggsave} in the ggplot2 package for supported file formats.}
}
\description{
\code{plotKaplanMeier} creates the Kaplain-Meier survival plot
}
\examples{
#todo
}

