% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyPopulation.R
\name{plotTimeToEvent}
\alias{plotTimeToEvent}
\title{Plot time-to-event}
\usage{
plotTimeToEvent(cohortMethodData, population = NULL, outcomeId,
  firstExposureOnly = FALSE, restrictToCommonPeriod = FALSE,
  washoutPeriod = 0, removeDuplicateSubjects = FALSE,
  minDaysAtRisk = 1, riskWindowStart = 0,
  addExposureDaysToStart = NULL, startAnchor = "cohort start",
  riskWindowEnd = 0, addExposureDaysToEnd = NULL,
  endAnchor = "cohort end", censorAtNewRiskWindow = FALSE,
  periodLength = 7, numberOfPeriods = 52, showFittedLines = TRUE,
  targetLabel = "Target", comparatorLabel = "Comparator",
  title = NULL, fileName = NULL)
}
\arguments{
\item{cohortMethodData}{An object of type \code{cohortMethodData} as generated using
\code{getDbCohortMethodData}.}

\item{population}{If specified, this population will be used as the starting
point instead of the cohorts in the \code{cohortMethodData}
object.}

\item{outcomeId}{The ID of the outcome. If not specified, no outcome-specific
transformations will be performed.}

\item{firstExposureOnly}{Should only the first exposure per subject be included? Note
that this is typically done in the
\code{createStudyPopulation} function,}

\item{restrictToCommonPeriod}{Restrict the analysis to the period when both exposures are observed?}

\item{washoutPeriod}{The mininum required continuous observation time prior to
index date for a person to be included in the cohort.}

\item{removeDuplicateSubjects}{Remove subjects that are in both the target and comparator
cohort? See details for allowed values.}

\item{minDaysAtRisk}{The minimum required number of days at risk.}

\item{riskWindowStart}{The start of the risk window (in days) relative to the \code{startAnchor}.}

\item{addExposureDaysToStart}{DEPRECATED: Add the length of exposure the start of the risk window?
Use \code{startAnchor} instead.}

\item{startAnchor}{The anchor point for the start of the risk window. Can be "cohort start"
or "cohort end".}

\item{riskWindowEnd}{The end of the risk window (in days) relative to the \code{endAnchor}.}

\item{addExposureDaysToEnd}{DEPRECATED: Add the length of exposure the risk window?
Use \code{endAnchor} instead.}

\item{endAnchor}{The anchor point for the end of the risk window. Can be "cohort start"
or "cohort end".}

\item{censorAtNewRiskWindow}{If a subject is in multiple cohorts, should time-at-risk be censored
when the new time-at-risk starts to prevent overlap?}

\item{periodLength}{The length in days of each period shown in the plot.}

\item{numberOfPeriods}{Number of periods to show in the plot. The periods are
equaly divided before and after the index date.}

\item{showFittedLines}{Fit lines to the proportions and show them in the plot?}

\item{targetLabel}{A label to us for the target cohort.}

\item{comparatorLabel}{A label to us for the comparator cohort.}

\item{title}{Optional: the main title for the plot.}

\item{fileName}{Name of the file where the plot should be saved, for example
'plot.png'. See the function \code{ggsave} in the ggplot2 package for
supported file formats.}
}
\value{
A ggplot object. Use the \code{\link[ggplot2]{ggsave}} function to save to file in a different
format.
}
\description{
Plot time-to-event
}
\details{
Creates a plot showing the number of events over time in the target and comparator cohorts, both before and after
index date. The plot also distinguishes between events inside and outside the time-at-risk period. This requires
the user to (re)specify the time-at-risk using the same arguments as the \code{\link{createStudyPopulation}} function.
Note that it is not possible to specify that people with the outcome prior should be removed, since the plot will
show these prior events.
}
