% Generated by roxygen2 (4.0.1): do not edit by hand
\name{psCreate}
\alias{psCreate}
\title{Create propensity scores}
\usage{
psCreate(cohortData, prior = prior("laplace", useCrossValidation = TRUE))
}
\arguments{
\item{cohortData}{An object of type \code{cohortData} as generated using \code{dbGetCohortData}}

\item{prior}{The prior used to fit the model}
}
\description{
\code{psCreate} creates propensity scores
}
\details{
todo
}
\examples{
#todo
}

