% Generated by roxygen2 (4.0.2): do not edit by hand
\name{psCreate}
\alias{psCreate}
\title{Create propensity scores}
\usage{
psCreate(cohortData, outcomeConceptId = NULL, prior = prior("laplace",
  useCrossValidation = TRUE))
}
\arguments{
\item{cohortData}{An object of type \code{cohortData} as generated using \code{dbGetCohortData}.}

\item{outcomeConceptId}{The concept ID of the outcome. Persons marked for removal for the outcome will be removed prior to
creating the propensity score model.}

\item{prior}{The prior used to fit the model.}
}
\description{
\code{psCreate} creates propensity scores using a regularized logistic regression.
}
\details{
\code{psCreate} creates propensity scores using a regularized logistic regression.
}
\examples{
#todo
}

