% Generated by roxygen2 (4.0.2): do not edit by hand
\name{psGetModel}
\alias{psGetModel}
\title{Get the propensity model}
\usage{
psGetModel(propensityScore, cohortData)
}
\arguments{
\item{propensityScore}{The propensity scores as generated using the \code{psCreate} function.}

\item{connectionDetails}{An R object of type ConnectionDetail (details for the function that contains server info, database type, optionally username/password, port).}

\item{cdmSchema}{Database schema that contains the vocabulary.}
}
\description{
\code{psGetModel} shows the propensity score model
}
\details{
Shows the coefficients and names of the covariates with non-zero coefficients.
}
\examples{
#todo
}

