% Generated by roxygen2 (4.0.1): do not edit by hand
\name{psPlot}
\alias{psPlot}
\title{Plot the propensity score distribution}
\usage{
psPlot(data, scale = "preference", type = "density", binWidth = 0.05)
}
\arguments{
\item{data}{A data frame with at least the two columns described below}

\item{scale}{The scale of the graph. Two scales are supported: \code{
scale = "propensity"} or  \code{#' scale = "preference"}. The preference score scale is defined by Walker
et al (2013).}

\item{type}{Type of plot. Two possible values: \code{type = "density"} or \code{type = "histogram"}}

\item{binWidth}{For histograms, the width of the bins}
}
\value{
Returns a GGPlot2 object. Use the ggsave function to save to file.
}
\description{
\code{psPlot} shows the propensity (or preference) score distribution
}
\details{
The data frame should have a least the following two columns:
\tabular{lll}{
  \verb{treatment}          \tab(integer) \tab Column indicating whether the person is in the treated (1) or comparator (0) group  \cr
  \verb{propensity_score}   \tab(real)    \tab Propensity score \cr
}
}
\examples{
treatment = rep(0:1, each = 100)
propensity_score = c(rnorm(100,mean=0.4, sd=0.25),rnorm(100,mean=0.6, sd=0.25))
data <- data.frame(treatment = treatment, propensity_score = propensity_score)
data <- data[data$propensity_score > 0 & data$propensity_score < 1,]
psPlot(data)
}
\references{
Walker AM, Patrick AR, Lauer MS, Hornbrook MC, Marin MG, Platt R, Roger VL, Stang P, and Schneeweiss S.
(2013) A tool for assessing the feasibility of comparative effectiveness research, Comparative Effective
Research, 3, 11-20
}

