% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdpsFunctions.R
\name{removeLowRank}
\alias{removeLowRank}
\title{Keeps only covariates with highest bias.}
\usage{
removeLowRank(data, rankings, rankCutoff, useExpRank)
}
\arguments{
\item{data}{List covariate information for data dimensions, as ffdf objects with columns \code{rowId}, \code{covariateId},
\code{treatment}, \code{outcome}}

\item{rankings}{List of bias and exposure rankings for each data dimension's covariates, as ffdf objects}

\item{rankCutoff}{Number of total covariates to keep}

\item{useExpRank}{Boolean to use exposure rank instead of bias rank}
}
\value{
Returns \code{data} with only the highest bias covariates.
}
\description{
This function eliminates data associated with covariates that have low bias. Only the top \code{rankCutoff} codes are kept.
}

