% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runHdps.R
\name{runHdps}
\alias{runHdps}
\title{Runs hdps implementation}
\usage{
runHdps(cohortData, connectionDetails, useConditionICD9 = TRUE,
  useDrug = TRUE, useProcedureCPT4 = TRUE, demographicsAnalysisIds = c(2,
  3, 5, 6), predefinedIncludeConceptIds = c(),
  predefinedExcludeConceptIds = c(), lowPopCutoff = 100,
  dimensionCutoff = 200, biasCutoff = 500, dropRareAfterExpand = FALSE,
  includeRR0 = TRUE, RR0Constant = 99999)
}
\arguments{
\item{cohortData}{\code{cohortData} object constructed by \code{getDbCohortData}}

\item{connectionDetails}{An R object of type connectionDetails created using the function createConnectionDetails in the DatabaseConnector package.}

\item{useConditionICD9}{Boolean to include ICD9 diagnosis codes}

\item{useDrug}{Boolean to include drugs}

\item{useProcedureCPT4}{Boolean to include CPT4 procedure codes}

\item{demographicsAnalysisIds}{Vector of analysis identifiers that correspond to demographics covariates}

\item{predefinedIncludeConceptIds}{Vector of predefined concept IDs to include as covariates}

\item{predefinedExcludeConceptIds}{Vector of predefined concept IDs to exclude as covariates}

\item{lowPopCutoff}{Cutoff for discarding rare covariates}

\item{dimensionCutoff}{Number of most prevalent covariates to include from each data dimension}

\item{biasCutoff}{Number of covariates to include in propensity score model (minus the demographics covariates)}

\item{dropRareAfterExpand}{Boolean for dropping codes below lowPopCutoff after expanding covariates}

\item{includeRR0}{Boolean for including covariates with large bias due to zero RR}

\item{RR0Constant}{Large constant to use for handling zero RR covariates}
}
\value{
Returns the input \code{cohortData} with new entries for \code{covariates} and \code{covariateRef}
}
\description{
This function runs the hdps implementation on \code{cohortData}.
}
\details{
Currently implemented for ICD9 diagnosis codes, CPT4 procedure codes, and drug concept ID codes. The mapping from CDM concept Id to
source codes is obtained, and used to relabel the covariates and covariateRef components of cohortData. Granularity for ICD9 codes
is implemented by "truncating" the source codes to 3 digits. Codes with fewer than \code{lowPopCutoff} number of patients are dropped.
For each covariate, three covariates are constructed: "once", "sporadic", "frequent" for any occurrences, occurrences greater than
the median, and occurrences greater than 3rd quartile, respectively. For each covariate, the "bias" as specified in hdps model is
calculated, and the highest \code{biasCutoff} codes across all data dimensions are selected for inclusion. Demographic information of
age, sex, race, and year are also included. \code{cohortData$covariates} and \code{cohortData$covariateRef} are altered accordingly.
}

