% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runHdps.R
\name{runHdps}
\alias{runHdps}
\title{Runs hdps implementation}
\usage{
runHdps(cohortData, useInpatientDiagnosis = TRUE,
  useAmbulatoryDiagnosis = TRUE, useDrugIngredient = TRUE,
  useInpatientProcedure = TRUE, useAmbulatoryProcedure = TRUE,
  demographicsAnalysisIds = c(2, 3, 5, 6), predefinedIncludeICD9Dx = c(),
  predefinedIncludeConceptIds = c(), predefinedExcludeICD9Dx = c(),
  predefinedExcludeConceptIds = c(), icd9AnalysisIds = c(104, 105, 106, 107,
  108, 109), dimensionCutoff = 200, rankCutoff = 500, useExpRank = FALSE,
  fudge = 0)
}
\arguments{
\item{cohortData}{\code{cohortData} object constructed by \code{getDbCohortMethodData}}

\item{useInpatientDiagnosis}{boolean to use inpatient icd9 condition codes}

\item{useAmbulatoryDiagnosis}{boolean to use ambulatory icd9 condition codes}

\item{useDrugIngredient}{boolean to use drug ingredient codes}

\item{useInpatientProcedure}{boolean to use inpatient procedure codes}

\item{useAmbulatoryProcedure}{boolean to use ambulatory procedure codes}

\item{demographicsAnalysisIds}{Vector of analysis identifiers that correspond to demographics covariates}

\item{predefinedIncludeICD9Dx}{Vector of predefined icd9 condition codes to include as covariates}

\item{predefinedIncludeConceptIds}{Vector of predefined concept IDs to include as covariates}

\item{predefinedExcludeICD9Dx}{Vector of predefined icd9 condition codes to exclude as covariates}

\item{predefinedExcludeConceptIds}{Vector of predefined concept IDs to exclude as covariates}

\item{icd9AnalysisIds}{Analysis IDs that contain icd9 condition codes}

\item{dimensionCutoff}{Number of most prevalent covariates to include from each data dimension}

\item{rankCutoff}{Number of covariates to include in propensity score model (minus the demographics covariates)}

\item{useExpRank}{Use exposure rank instead of bias rank}

\item{fudge}{Constant to avoid 0/Inf in calculating RR. Setting equal to 0 will cause those covariates to be discarded}
}
\value{
Returns the input \code{cohortData} with new entries for \code{covariates} and \code{covariateRef}
}
\description{
This function runs the hdps implementation on \code{cohortData}.
}
\details{
Currently implemented for ICD9 diagnosis codes, CPT4 procedure codes, and drug concept ID codes. The mapping from CDM concept Id to
source codes is obtained, and used to relabel the covariates and covariateRef components of cohortData. Granularity for ICD9 codes
is implemented by "truncating" the source codes to 3 digits. Codes with fewer than \code{lowPopCutoff} number of patients are dropped.
For each covariate, three covariates are constructed: "once", "sporadic", "frequent" for any occurrences, occurrences greater than
the median, and occurrences greater than 3rd quartile, respectively. For each covariate, the "bias" as specified in hdps model is
calculated, and the highest \code{biasCutoff} codes across all data dimensions are selected for inclusion. Demographic information of
age, sex, race, and year are also included. \code{cohortData$covariates} and \code{cohortData$covariateRef} are altered accordingly.
}

