% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{saveCmAnalysisList}
\alias{saveCmAnalysisList}
\title{Save a list of cmAnalysis to file}
\usage{
saveCmAnalysisList(cmAnalysisList, file)
}
\arguments{
\item{cmAnalysisList}{The cmAnalysis list to be written to file}

\item{file}{The name of the file where the results will be written}
}
\description{
Write a list of objects of type \code{cmAnalysis} to file. The file is in JSON format.
}
