% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataLoadingSaving.R
\name{saveCohortMethodData}
\alias{saveCohortMethodData}
\title{Save the cohort data to folder}
\usage{
saveCohortMethodData(cohortMethodData, file)
}
\arguments{
\item{cohortMethodData}{An object of type \code{cohortMethodData} as generated using
\code{getDbCohortMethodData}.}

\item{file}{The name of the folder where the data will be written. The folder should
                          not yet exist.}
}
\description{
\code{saveCohortMethodData} saves an object of type cohortMethodData to folder.
}
\details{
The data will be written to a set of files in the folder specified by the user.
}
\examples{
# todo
}

