% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettingsObjects.R
\name{saveTargetComparatorOutcomesList}
\alias{saveTargetComparatorOutcomesList}
\title{Save a list of \code{TargetComparatorOutcomes} to file}
\usage{
saveTargetComparatorOutcomesList(targetComparatorOutcomesList, file)
}
\arguments{
\item{file}{The name of the file where the results will be written}

\item{exposuresOutcomeList}{A list of objects of type \code{TargetComparatorOutcomes} as created
using the \code{\link[=createTargetComparatorOutcomes]{createTargetComparatorOutcomes()}} function.}
}
\description{
Write a list of objects of type \code{TargetComparatorOutcomes} to file. The file is in JSON format.
}
