% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{saveTargetComparatorOutcomesList}
\alias{saveTargetComparatorOutcomesList}
\title{Save a list of targetComparatorOutcomes to file}
\usage{
saveTargetComparatorOutcomesList(targetComparatorOutcomesList, file)
}
\arguments{
\item{targetComparatorOutcomesList}{The targetComparatorOutcomes list to be written to file}

\item{file}{The name of the file where the results will be written}
}
\description{
Write a list of objects of type \code{targetComparatorOutcomes} to file. The file is in JSON format.
}
