% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RunAnalyses.R
\name{summarizeAnalyses}
\alias{summarizeAnalyses}
\title{Create a summary report of the analyses}
\usage{
summarizeAnalyses(outcomeReference)
}
\arguments{
\item{outcomeReference}{A data.frame as created by the \code{\link{runAnalyses}} function.}
}
\description{
Create a summary report of the analyses
}

