% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalyses.R
\name{summarizeAnalyses}
\alias{summarizeAnalyses}
\title{Create a summary report of the analyses}
\usage{
summarizeAnalyses(referenceTable, outputFolder)
}
\arguments{
\item{referenceTable}{A \link[dplyr:reexports]{dplyr::tibble} as created by the \link{runCmAnalyses} function.}

\item{outputFolder}{Name of the folder where all the outputs have been written to.}
}
\value{
A tibble containing summary statistics for each target-comparator-outcome-analysis combination.
}
\description{
Create a summary report of the analyses
}
