% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsFunctions.R
\name{trimByIptw}
\alias{trimByIptw}
\title{Remove subjects with a high IPTW}
\usage{
trimByIptw(population, maxWeight = 10)
}
\arguments{
\item{population}{A data frame with at least the two columns described in the details}

\item{maxWeight}{The maximum allowed IPTW.}
}
\value{
Returns a tibble with the same  columns as the input.
}
\description{
Remove subjects having a weight higher than the user-specified threshold.
}
\details{
The data frame should have the following two columns:
\itemize{
\item treatment (integer): Column indicating whether the person is in the target (1) or comparator (0) group.
\item iptw (numeric): Propensity score.
}
}
\examples{
rowId <- 1:2000
treatment <- rep(0:1, each = 1000)
iptw <- 1 / c(runif(1000, min = 0, max = 1), runif(1000, min = 0, max = 1))
data <- data.frame(rowId = rowId, treatment = treatment, iptw = iptw)
result <- trimByIptw(data)

}
