% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsFunctions.R
\name{trimByPs}
\alias{trimByPs}
\title{Trim persons by propensity score}
\usage{
trimByPs(population, trimFraction = 0.05)
}
\arguments{
\item{population}{A data frame with the three columns described below}

\item{trimFraction}{This fraction will be removed from each treatment group. In the target
group, persons with the highest propensity scores will be removed, in the
comparator group person with the lowest scores will be removed.}
}
\value{
Returns a date frame with the same three columns as the input.
}
\description{
\code{trimByPs} uses the provided propensity scores to trim subjects with extreme scores.
}
\details{
The data frame should have the following three columns: \tabular{lll}{ \verb{rowId} \tab(numeric)
\tab A unique identifier for each row (e.g. the person ID) \cr \verb{treatment} \tab(integer) \tab
Column indicating whether the person is in the target (1) or comparator\cr \tab \tab (0) group \cr
\verb{propensityScore} \tab(numeric) \tab Propensity score \cr }
}
\examples{
rowId <- 1:2000
treatment <- rep(0:1, each = 1000)
propensityScore <- c(runif(1000, min = 0, max = 1), runif(1000, min = 0, max = 1))
data <- data.frame(rowId = rowId, treatment = treatment, propensityScore = propensityScore)
result <- trimByPs(data, 0.05)

}
