% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CohortMethod.R
\docType{data}
\name{vignetteOutcomeModel2}
\alias{vignetteOutcomeModel2}
\title{Outcome data for the vignette}
\usage{
data(vignetteOutcomeModel2)
}
\description{
Outcome data for the vignette
}
\keyword{datasets}

